Fetches Simulation Parameter Meta Information

This provides meta information about simulatino options. This includes option names, text descriptions, ui_names used, etc.

Usage

CTS_fetch_sc_meta(
  MOD_yaml_file = system.file(package = "ruminate", "templates", "CTS.yaml")
)

Arguments

MOD_yaml_file

Module configuration file with MC as main section.

Value

List with the following elements:

config List from the YAML->MC->sim_config.

summary: Dataframe with elements of config in tabular format.

ui_config Vector of all the ui_ids for configuration options.

Examples

CTS_fetch_sc_meta()
#> $config
#> $config$atol
#> $config$atol$group
#> [1] "ODE Solver"
#> 
#> $config$atol$label
#> [1] "Absolute tolerance"
#> 
#> $config$atol$tooltip
#> [1] "Absolute tolerance used by the ODE solver to determine if a good solution has been achieved; This is also used in the solved linear model to check if prior doses do not add anything to the solution."
#> 
#> $config$atol$uitype
#> [1] "text"
#> 
#> $config$atol$type
#> [1] "numeric"
#> 
#> $config$atol$value
#> [1] "1e-8"
#> 
#> $config$atol$options
#> NULL
#> 
#> $config$atol$use
#> [1] "rxSolve"
#> 
#> $config$atol$placeholder
#> [1] "1e-8"
#> 
#> $config$atol$ui
#> [1] "cts_config_atol"
#> 
#> 
#> $config$rtol
#> $config$rtol$group
#> [1] "ODE Solver"
#> 
#> $config$rtol$label
#> [1] "Relative tolerance"
#> 
#> $config$rtol$tooltip
#> [1] "Relative tolerance used by the ODE solver to determine if a good solution has been achieved. This is also used in the solved linear model to check if prior doses do not add anything to the solution."
#> 
#> $config$rtol$uitype
#> [1] "text"
#> 
#> $config$rtol$type
#> [1] "numeric"
#> 
#> $config$rtol$value
#> [1] "1e-6"
#> 
#> $config$rtol$options
#> NULL
#> 
#> $config$rtol$use
#> [1] "rxSolve"
#> 
#> $config$rtol$placeholder
#> [1] "1e-6"
#> 
#> $config$rtol$ui
#> [1] "cts_config_rtol"
#> 
#> 
#> $config$sigdig
#> $config$sigdig$group
#> [1] "ODE Solver"
#> 
#> $config$sigdig$label
#> [1] "Significant digits"
#> 
#> $config$sigdig$tooltip
#> [1] "Specifies the \"significant digits\" that the ode solving requests. When specified this controls the relative and absolute tolerances of the ODE solvers. By default the tolerance is 0.5*10^(-sigdig-2) for regular ODEs. For the sensitivity equations the default is ⁠0.5*10\\^(-sigdig-1.5)⁠ (sensitivity changes only applicable for liblsoda). This also controls the atol/rtol of the steady state solutions. The ssAtol/ssRtol is ⁠0.5*10\\^(-sigdig)⁠ and for the sensitivities ⁠0.5*10\\^(-sigdig+0.625)⁠. By default this is unspecified (NULL) and uses the standard atol/rtol."
#> 
#> $config$sigdig$uitype
#> [1] "text"
#> 
#> $config$sigdig$type
#> [1] "numeric"
#> 
#> $config$sigdig$value
#> [1] "NULL"
#> 
#> $config$sigdig$options
#> NULL
#> 
#> $config$sigdig$use
#> [1] "rxSolve"
#> 
#> $config$sigdig$placeholder
#> [1] "NULL"
#> 
#> $config$sigdig$ui
#> [1] "cts_config_sigdig"
#> 
#> 
#> $config$seed
#> $config$seed$group
#> [1] "Misc"
#> 
#> $config$seed$label
#> [1] "Random Seed"
#> 
#> $config$seed$tooltip
#> [1] "Integer value to set the seed for the random number generator"
#> 
#> $config$seed$uitype
#> [1] "text"
#> 
#> $config$seed$type
#> [1] "numeric"
#> 
#> $config$seed$value
#> [1] "86753090"
#> 
#> $config$seed$options
#> NULL
#> 
#> $config$seed$use
#> [1] "Misc"
#> 
#> $config$seed$placeholder
#> [1] "8675309"
#> 
#> $config$seed$ui
#> [1] "cts_config_seed"
#> 
#> 
#> $config$nsteps
#> $config$nsteps$group
#> [1] "Misc"
#> 
#> $config$nsteps$label
#> [1] "Time Steps"
#> 
#> $config$nsteps$tooltip
#> [1] "Integer value representing the number of time steps to add between zero and the end of trial. A larger value will give smoother curves but it will also take more time to simulate."
#> 
#> $config$nsteps$uitype
#> [1] "text"
#> 
#> $config$nsteps$type
#> [1] "numeric"
#> 
#> $config$nsteps$value
#> [1] "200"
#> 
#> $config$nsteps$options
#> NULL
#> 
#> $config$nsteps$use
#> [1] "Misc"
#> 
#> $config$nsteps$placeholder
#> [1] "200"
#> 
#> $config$nsteps$ui
#> [1] "cts_config_nsteps"
#> 
#> 
#> $config$preamble
#> $config$preamble$group
#> [1] "Rule Environment Preamble"
#> 
#> $config$preamble$label
#> NULL
#> 
#> $config$preamble$tooltip
#> [1] "Optional user-defined code to be created in the rule-evaluation environment."
#> 
#> $config$preamble$uitype
#> [1] "textarea"
#> 
#> $config$preamble$type
#> [1] "character"
#> 
#> $config$preamble$value
#> NULL
#> 
#> $config$preamble$options
#> NULL
#> 
#> $config$preamble$use
#> [1] "rules"
#> 
#> $config$preamble$placeholder
#> [1] "my_func = function(id=NULL){\n\n}"
#> 
#> $config$preamble$ui
#> [1] "cts_config_preamble"
#> 
#> 
#> 
#> $cfg_summary
#>       name                  ui     use      type                     group
#> 1     atol     cts_config_atol rxSolve   numeric                ODE Solver
#> 2     rtol     cts_config_rtol rxSolve   numeric                ODE Solver
#> 3   sigdig   cts_config_sigdig rxSolve   numeric                ODE Solver
#> 4     seed     cts_config_seed    Misc   numeric                      Misc
#> 5   nsteps   cts_config_nsteps    Misc   numeric                      Misc
#> 6 preamble cts_config_preamble   rules character Rule Environment Preamble
#> 
#> $ui_config
#> [1] "cts_config_atol"     "cts_config_rtol"     "cts_config_sigdig"  
#> [4] "cts_config_seed"     "cts_config_nsteps"   "cts_config_preamble"
#>