Initialize CTS Module State

Creates a list of the initialized module state

Usage

CTS_init_state(FM_yaml_file, MOD_yaml_file, id, id_MB, session)

Arguments

FM_yaml_file

App configuration file with FM as main section

MOD_yaml_file

Module configuration file with MC as main section

id

ID string for the module

id_MB

An ID string that corresponds with the ID used to call the MB modules

session

Shiny session variable

Value

list containing an empty CTS state

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = CTS_test_mksession(session=list())
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#> → MB: module checksum updated:81b00278da5d2205e016183a001bfd19
#> → MB: State initialized
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 4c1cee671a16d68ac228a0860a4346de
#> → MB: module checksum updated:e2ee66142df9cd7724aa9ac416a0399f
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 5e6fc90200958cc3f4f1d979475b1e86
#> → MB: module checksum updated:bba5b8caae381920a547eb80c192bac6
#> bottom
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → CTS:   dest:   file.path("config","report.docx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → CTS:   dest:   file.path("config","report.pptx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → CTS:   dest:   file.path("config","report.yaml")
#> → CTS: module checksum updated:4dc8f03bdf3ae334d0f4ebef877e80c8
#> → CTS: State initialized
#> → CTS: cohort checksum updated: df5e2c9cd978aee8e0d58ba388646f86
#> → CTS: module checksum updated:3cbb9da2caeed7d9a4ce2f24bdbb18a3
session = sess_res$session
input   = sess_res$input

state = CTS_init_state(
   FM_yaml_file  = system.file(package = "formods",
                               "templates",
                               "formods.yaml"),
   MOD_yaml_file = system.file(package = "ruminate",
                               "templates",
                               "CTS.yaml"),
   id              = "CTS",
   id_MB           = "MB",
   session         = session)
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → CTS:   dest:   file.path("config","report.docx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → CTS:   dest:   file.path("config","report.pptx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → CTS:   dest:   file.path("config","report.yaml")
#> → CTS: module checksum updated:4dc8f03bdf3ae334d0f4ebef877e80c8
#> → CTS: State initialized

state
#> $yaml
#> $yaml$FM
#> $yaml$FM$include
#> $yaml$FM$include$files
#> $yaml$FM$include$files[[1]]
#> $yaml$FM$include$files[[1]]$file
#> $yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[2]]
#> $yaml$FM$include$files[[2]]$file
#> $yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[3]]
#> $yaml$FM$include$files[[3]]$file
#> $yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $yaml$FM$deployed
#> [1] FALSE
#> 
#> $yaml$FM$code
#> $yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $yaml$FM$notifications
#> $yaml$FM$notifications$config
#> $yaml$FM$notifications$config$success
#> $yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $yaml$FM$notifications$config$failure
#> $yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $yaml$FM$notifications$config$info
#> $yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $yaml$FM$notifications$config$warning
#> $yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $yaml$FM$reporting
#> $yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $yaml$FM$reporting$content_init
#> $yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> 
#> $yaml$FM$ui
#> $yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $yaml$FM$ui$select_size
#> [1] 10
#> 
#> $yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $yaml$FM$data_meta
#> $yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $yaml$FM$data_meta$many_sep
#> [1] ",⋅⋅⋅,"
#> 
#> $yaml$FM$data_meta$data_types
#> $yaml$FM$data_meta$data_types$character
#> $yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $yaml$FM$data_meta$data_types$double
#> $yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$integer
#> $yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$other
#> $yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $yaml$FM$labels
#> $yaml$FM$labels$default_ds
#> [1] "Original data set"
#> 
#> 
#> $yaml$FM$user_files
#> $yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $yaml$FM$logging
#> $yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $MC
#> $MC$element_object_name
#> [1] "CTS_obj"
#> 
#> $MC$code
#> $MC$code$packages
#> [1] "XXX"
#> 
#> $MC$code$readOnly
#> [1] TRUE
#> 
#> $MC$code$mode
#> [1] "r"
#> 
#> 
#> $MC$compact
#> $MC$compact$code
#> [1] TRUE
#> 
#> $MC$compact$clip
#> [1] TRUE
#> 
#> 
#> $MC$reporting
#> $MC$reporting$enabled
#> [1] FALSE
#> 
#> $MC$reporting$priority
#> [1] 1
#> 
#> 
#> $MC$formatting
#> $MC$formatting$code
#> $MC$formatting$code$width
#> [1] 800
#> 
#> $MC$formatting$code$height
#> [1] 300
#> 
#> 
#> $MC$formatting$current_element
#> $MC$formatting$current_element$width
#> [1] "200px"
#> 
#> 
#> $MC$formatting$source_model
#> $MC$formatting$source_model$width
#> [1] "200px"
#> 
#> 
#> $MC$formatting$element_name
#> $MC$formatting$element_name$width
#> [1] "200px"
#> 
#> 
#> $MC$formatting$preview
#> $MC$formatting$preview$width
#> [1] "800px"
#> 
#> $MC$formatting$preview$height
#> [1] "500px"
#> 
#> 
#> $MC$formatting$input
#> NULL
#> 
#> $MC$formatting$button_clk_run
#> $MC$formatting$button_clk_run$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_run$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_clk_del
#> $MC$formatting$button_clk_del$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_del$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_del$tooltip
#> [1] "Delete the current cohort."
#> 
#> $MC$formatting$button_clk_del$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_save
#> $MC$formatting$button_clk_save$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_save$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_save$tooltip
#> [1] "Save the name for the current cohort."
#> 
#> $MC$formatting$button_clk_save$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_clip
#> $MC$formatting$button_clk_clip$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_clip$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_clip$tooltip
#> [1] "Copy the code to generate the current cohort to the clipboard."
#> 
#> $MC$formatting$button_clk_clip$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_copy
#> $MC$formatting$button_clk_copy$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_copy$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_copy$tooltip
#> [1] "Make a copy of the current cohort."
#> 
#> $MC$formatting$button_clk_copy$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_new
#> $MC$formatting$button_clk_new$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_new$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_new$tooltip
#> [1] "Create a new cohort."
#> 
#> $MC$formatting$button_clk_new$tooltip_position
#> [1] "right"
#> 
#> $MC$formatting$button_clk_new$width
#> [1] 100
#> 
#> 
#> $MC$formatting$rule_condition
#> $MC$formatting$rule_condition$width
#> [1] 300
#> 
#> $MC$formatting$rule_condition$height
#> [1] 100
#> 
#> $MC$formatting$rule_condition$tooltip
#> [1] "This is a condition that should evaluate to either TRUE (if the action should be taken) or FALSE (if it should be skipped)."
#> 
#> $MC$formatting$rule_condition$tooltip_position
#> [1] "bottom"
#> 
#> $MC$formatting$rule_condition$placeholder
#> [1] "examples: TRUE will run the action at at every visit or (time > 0) & (Ac < 10) will be more selective."
#> 
#> 
#> $MC$formatting$action_dosing_state
#> $MC$formatting$action_dosing_state$width
#> [1] 200
#> 
#> 
#> $MC$formatting$action_dosing_values
#> $MC$formatting$action_dosing_values$width
#> [1] 200
#> 
#> $MC$formatting$action_dosing_values$placeholder
#> [1] "c(1, 1, 1)"
#> 
#> $MC$formatting$action_dosing_values$tooltip
#> [1] "Expression defining dosing amounts that will evaluate as a numeric vector of numbers."
#> 
#> 
#> $MC$formatting$action_dosing_times
#> $MC$formatting$action_dosing_times$width
#> [1] 200
#> 
#> $MC$formatting$action_dosing_times$placeholder
#> [1] "c(0, 7, 14)"
#> 
#> $MC$formatting$action_dosing_times$tooltip
#> [1] "Expression defining dosing times that will evaluate as a numeric vector of numbers."
#> 
#> $MC$formatting$action_dosing_times$tooltip_position
#> [1] "bottom"
#> 
#> 
#> $MC$formatting$action_dosing_durations
#> $MC$formatting$action_dosing_durations$width
#> [1] 200
#> 
#> $MC$formatting$action_dosing_durations$placeholder
#> [1] "c(0, 0, 0)"
#> 
#> $MC$formatting$action_dosing_durations$tooltip
#> [1] "Expression containing dosing duration that will evaluate as a numeric vector of numbers (zeros for bolus dosing)."
#> 
#> $MC$formatting$action_dosing_durations$tooltip_position
#> [1] "bottom"
#> 
#> 
#> $MC$formatting$action_set_state_state
#> $MC$formatting$action_set_state_state$width
#> [1] 200
#> 
#> 
#> $MC$formatting$action_set_state_value
#> $MC$formatting$action_set_state_value$width
#> [1] 200
#> 
#> $MC$formatting$action_set_state_value$placeholder
#> [1] "A value like 20 or an expression in terms of a system element such as Ac/2"
#> 
#> $MC$formatting$action_set_state_value$tooltip
#> [1] "An expression that evaluates to a new number."
#> 
#> 
#> $MC$formatting$action_manual_code
#> $MC$formatting$action_manual_code$width
#> [1] 200
#> 
#> $MC$formatting$action_manual_code$height
#> [1] 75
#> 
#> $MC$formatting$action_manual_code$placeholder
#> [1] "SI_interval_ev = etRbind(SI_interval_ev,\n  et(cmt  = 'Ac',\n     id   = id,\n     amt  = 10,\n     evid = 4,\n     time = time))"
#> 
#> $MC$formatting$action_manual_code$tooltip
#> [1] "Expression with code to manually update the event table."
#> 
#> 
#> $MC$formatting$rule_type
#> $MC$formatting$rule_type$width
#> [1] 100
#> 
#> $MC$formatting$rule_type$options
#> $MC$formatting$rule_type$options$dose
#> $MC$formatting$rule_type$options$dose$choice
#> [1] "Dosing"
#> 
#> $MC$formatting$rule_type$options$dose$subtext
#> [1] "Add bolus or infusion dosing"
#> 
#> $MC$formatting$rule_type$options$dose$value
#> [1] "dose"
#> 
#> 
#> $MC$formatting$rule_type$options$state_reset
#> $MC$formatting$rule_type$options$state_reset$choice
#> [1] "Set State"
#> 
#> $MC$formatting$rule_type$options$state_reset$subtext
#> [1] "Reset the named state or compartment to a specified value"
#> 
#> $MC$formatting$rule_type$options$state_reset$value
#> [1] "set state"
#> 
#> 
#> $MC$formatting$rule_type$options$manual
#> $MC$formatting$rule_type$options$manual$choice
#> [1] "Manual"
#> 
#> $MC$formatting$rule_type$options$manual$subtext
#> [1] "Alter the event table directly."
#> 
#> $MC$formatting$rule_type$options$manual$value
#> [1] "manual"
#> 
#> 
#> 
#> 
#> 
#> $MC$labels
#> $MC$labels$current_element
#> NULL
#> 
#> $MC$labels$element_name
#> NULL
#> 
#> $MC$labels$save_btn
#> [1] "Save"
#> 
#> $MC$labels$clip_btn
#> [1] "Code"
#> 
#> $MC$labels$copy_btn
#> [1] "Copy"
#> 
#> $MC$labels$del_btn
#> [1] "Delete"
#> 
#> $MC$labels$new_btn
#> [1] "New"
#> 
#> $MC$labels$source_model
#> [1] "Select Model"
#> 
#> $MC$labels$rule_type
#> [1] "Type of rule"
#> 
#> $MC$labels$rule_condition
#> [1] "Condition to trigger rule"
#> 
#> $MC$labels$action_dosing_state
#> [1] "Dosing into"
#> 
#> $MC$labels$action_dosing_values
#> [1] "Values"
#> 
#> $MC$labels$action_dosing_times
#> [1] "Times"
#> 
#> $MC$labels$action_manual_code
#> [1] "Code to evaluate"
#> 
#> $MC$labels$action_dosing_durations
#> [1] "Durations"
#> 
#> $MC$labels$action_set_state_state
#> [1] "State/compartment to reset"
#> 
#> $MC$labels$action_set_state_value
#> [1] "New value for state/compartment"
#> 
#> 
#> $MC$errors
#> $MC$errors$bad_input
#> [1] "Input error message"
#> 
#> 
#> $MC$tooltips
#> $MC$tooltips$include
#> [1] TRUE
#> 
#> $MC$tooltips$elements
#> [1] "Show cohort components"
#> 
#> $MC$tooltips$show_code
#> [1] "Show cohort code"
#> 
#> $MC$tooltips$url_manual
#> [1] "https://ruminate.ubiquity.tools/articles/clinical_trial_simulation.html#example-manual-code-evaluation"
#> 
#> $MC$tooltips$url_rule_env
#> [1] "https://ruminate.ubiquity.tools/articles/clinical_trial_simulation.html#rule-evaluation-environment"
#> 
#> 
#> $MC$rule_env_content
#> [1] "uiele = tagList(tags$b(\"Simulation environment\"), tags$br(),\n                \"The condition and actions (e.g., dose times, amounts,\n                durations) that are triggered are all\n                evaluated in an environment with different objects that\n                you can use. For details on this see the \", \n                tags$a(href=state[[\"MC\"]][[\"tooltips\"]][[\"url_rule_env\"]],\n                 \"Clinical Trial Simulation\"), \" vignette.\")\nuiele = tagList(uiele, tags$br())\nuiele = tagList(uiele, \"For the current system the following are available\", tags$br())\nuiele = tagList(uiele, tags$br())\nuiele = tagList(uiele, rx_details[[\"ht_info\"]])"
#> 
#> $MC$sim_config
#> $MC$sim_config$atol
#> $MC$sim_config$atol$group
#> [1] "ODE Solver"
#> 
#> $MC$sim_config$atol$label
#> [1] "Absolute tolerance"
#> 
#> $MC$sim_config$atol$tooltip
#> [1] "Absolute tolerance used by the ODE solver to determine if a good solution has been achieved; This is also used in the solved linear model to check if prior doses do not add anything to the solution."
#> 
#> $MC$sim_config$atol$type
#> [1] "numeric"
#> 
#> $MC$sim_config$atol$value
#> [1] "1e-8"
#> 
#> $MC$sim_config$atol$options
#> NULL
#> 
#> $MC$sim_config$atol$use
#> [1] "rxSolve"
#> 
#> 
#> $MC$sim_config$rtol
#> $MC$sim_config$rtol$group
#> [1] "ODE Solver"
#> 
#> $MC$sim_config$rtol$label
#> [1] "Absolute tolerance"
#> 
#> $MC$sim_config$rtol$tooltip
#> [1] "Relative tolerance used by the ODE solver to determine if a good solution has been achieved. This is also used in the solved linear model to check if prior doses do not add anything to the solution."
#> 
#> $MC$sim_config$rtol$type
#> [1] "numeric"
#> 
#> $MC$sim_config$rtol$value
#> [1] "1e-6"
#> 
#> $MC$sim_config$rtol$options
#> NULL
#> 
#> $MC$sim_config$rtol$use
#> [1] "rxSolve"
#> 
#> 
#> $MC$sim_config$sigdig
#> $MC$sim_config$sigdig$group
#> [1] "ODE Solver"
#> 
#> $MC$sim_config$sigdig$label
#> [1] "Absolute tolerance"
#> 
#> $MC$sim_config$sigdig$tooltip
#> [1] "Specifies the \"significant digits\" that the ode solving requests. When specified this controls the relative and absolute tolerances of the ODE solvers. By default the tolerance is 0.5*10^(-sigdig-2) for regular ODEs. For the sensitivity equations the default is ⁠0.5*10\\^(-sigdig-1.5)⁠ (sensitivity changes only applicable for liblsoda). This also controls the atol/rtol of the steady state solutions. The ssAtol/ssRtol is ⁠0.5*10\\^(-sigdig)⁠ and for the sensitivities ⁠0.5*10\\^(-sigdig+0.625)⁠. By default this is unspecified (NULL) and uses the standard atol/rtol."
#> 
#> $MC$sim_config$sigdig$type
#> [1] "numeric"
#> 
#> $MC$sim_config$sigdig$value
#> NULL
#> 
#> $MC$sim_config$sigdig$options
#> NULL
#> 
#> $MC$sim_config$sigdig$use
#> [1] "rxSolve"
#> 
#> 
#> 
#> 
#> $CTS
#> $CTS$button_counters
#> $CTS$button_counters$button_clk_save
#> [1] 0
#> 
#> $CTS$button_counters$button_clk_clip
#> [1] 0
#> 
#> $CTS$button_counters$button_clk_del
#> [1] 0
#> 
#> $CTS$button_counters$button_clk_copy
#> [1] 0
#> 
#> $CTS$button_counters$button_clk_new
#> [1] 0
#> 
#> 
#> $CTS$ui_hold
#> $CTS$ui_hold$button_clk_save
#> [1] FALSE
#> 
#> $CTS$ui_hold$button_clk_clip
#> [1] FALSE
#> 
#> $CTS$ui_hold$button_clk_del
#> [1] FALSE
#> 
#> $CTS$ui_hold$button_clk_copy
#> [1] FALSE
#> 
#> $CTS$ui_hold$button_clk_new
#> [1] FALSE
#> 
#> $CTS$ui_hold$element_name
#> [1] FALSE
#> 
#> $CTS$ui_hold$rule_condition
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_dosing_state
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_dosing_values
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_dosing_times
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_dosing_durations
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_set_state_state
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_set_state_value
#> [1] FALSE
#> 
#> $CTS$ui_hold$action_manual_code
#> [1] FALSE
#> 
#> $CTS$ui_hold$cts_config_atol
#> [1] FALSE
#> 
#> $CTS$ui_hold$cts_config_rtol
#> [1] FALSE
#> 
#> $CTS$ui_hold$cts_config_sigdig
#> [1] FALSE
#> 
#> $CTS$ui_hold$rule_type
#> [1] FALSE
#> 
#> $CTS$ui_hold$source_model
#> [1] FALSE
#> 
#> $CTS$ui_hold$element_selection
#> [1] FALSE
#> 
#> 
#> $CTS$ui_ids
#>  [1] "button_clk_save"         "button_clk_clip"        
#>  [3] "button_clk_del"          "button_clk_copy"        
#>  [5] "button_clk_new"          "element_name"           
#>  [7] "rule_condition"          "action_dosing_state"    
#>  [9] "action_dosing_values"    "action_dosing_times"    
#> [11] "action_dosing_durations" "action_set_state_state" 
#> [13] "action_set_state_value"  "action_manual_code"     
#> [15] "cts_config_atol"         "cts_config_rtol"        
#> [17] "cts_config_sigdig"       "rule_type"              
#> [19] "source_model"            "element_selection"      
#> 
#> $CTS$isgood
#> [1] TRUE
#> 
#> $CTS$MDL
#> $CTS$MDL$isgood
#> [1] TRUE
#> 
#> $CTS$MDL$hasmdl
#> [1] TRUE
#> 
#> $CTS$MDL$catalog
#>                    label      object MOD_TYPE id
#> 1  One compartment model MB_obj_1_rx       MB MB
#> 2 mAb Model (Davda 2014) MB_obj_2_rx       MB MB
#>                           checksum                      MDLchecksum
#> 1 bba5b8caae381920a547eb80c192bac6 4c1cee671a16d68ac228a0860a4346de
#> 2 bba5b8caae381920a547eb80c192bac6 5e6fc90200958cc3f4f1d979475b1e86
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 code
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                MB_obj_1_fcn = function () \n{\n    description <- "One compartment PK model with linear clearance using differential equations"\n    ini({\n        lka <- 0.45\n        label("Absorption rate (Ka)")\n        lcl <- 1\n        label("Clearance (CL)")\n        lvc <- 3.45\n        label("Central volume of distribution (V)")\n        propSd <- c(0, 0.5)\n        label("Proportional residual error (fraction)")\n    })\n    model({\n        ka <- exp(lka)\n        cl <- exp(lcl)\n        vc <- exp(lvc)\n        kel <- cl/vc\n        d/dt(depot) <- -ka * depot\n        d/dt(central) <- ka * depot - kel * central\n        Cc <- central/vc\n        Cc ~ prop(propSd)\n    })\n}\nMB_obj_1_rx =  rxode2::rxode2(MB_obj_1_fcn)
#> 2 MB_obj_2_fcn = function () \n{\n    description <- "Two compartment PK model with linear clearance for average monoclonal antibodies (Davda 2014)"\n    reference <- "Davda JP, Dodds MG, Gibbs MA, Wisdom W, Gibbs JP. A model-based meta-analysis of monoclonal antibody pharmacokinetics to guide optimal first-in-human study design. MAbs. 2014;6(4):1094-1102. doi:10.4161/mabs.29095"\n    ini({\n        lfdepot <- -0.295714244149045\n        label("Subcutaneous bioavailability (fraction)")\n        lka <- -1.26584820804402\n        label("Absorption rate (Ka, 1/day)")\n        lcl <- -1.6094379124341\n        label("Clearance (CL, L/day)")\n        lv <- 1.28370777234479\n        label("Central volume of distribution (V, L)")\n        lvp <- 1.01160091167848\n        label("Peripheral volume of distribution (Vp, L)")\n        lq <- -0.29169009384932\n        label("Intercompartmental clearance (Q, L/day)")\n        allocl <- 0.865\n        label("Allometric exponent on clearance and intercompartmental clearance (unitless)")\n        allov <- 0.957\n        label("Allometric exponent on volumes of distribution (unitless)")\n        prop.err <- c(0, 0.144)\n        label("Proportional residual error (fraction)")\n        etafdepot ~ 0\n        etaka ~ 0.416\n        etacl + etav + etavp ~ c(0.0987, 0.0786, 0.116, 0.0377, \n            0.0619, 0.0789)\n        etaq ~ 0.699\n    })\n    model({\n        fdepot <- exp(lfdepot + etafdepot)\n        ka <- exp(lka + etaka)\n        wtnorm <- log(WT/70)\n        cl <- exp(lcl + allocl * wtnorm + etacl)\n        q <- exp(lq + allocl * wtnorm + etaq)\n        v <- exp(lv + allov * wtnorm + etav)\n        vp <- exp(lvp + allov * wtnorm + etavp)\n        Cc <- linCmt()\n        f(depot) <- fdepot\n        Cc ~ prop(prop.err)\n    })\n}\nMB_obj_2_rx =  rxode2::rxode2(MB_obj_2_fcn)
#> 
#> $CTS$MDL$modules
#> $CTS$MDL$modules$MB
#> $CTS$MDL$modules$MB$MB
#> [1] "bba5b8caae381920a547eb80c192bac6"
#> 
#> 
#> 
#> $CTS$MDL$mdl
#> $CTS$MDL$mdl$MB_obj_1_rx
#> $CTS$MDL$mdl$MB_obj_1_rx$label
#> [1] "One compartment model"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$MOD_TYPE
#> [1] "MB"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$id
#> [1] "MB"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$rx_obj
#>  ── rxode2-based free-form 2-cmt ODE model ────────────────────────────────────── 
#>  ── Initalization: ──  
#> Fixed Effects ($theta): 
#>    lka    lcl    lvc propSd 
#>   0.45   1.00   3.45   0.50 
#> 
#> States ($state or $stateDf): 
#>   Compartment Number Compartment Name
#> 1                  1            depot
#> 2                  2          central
#>  ── Model (Normalized Syntax): ── 
#> function() {
#>     description <- "One compartment PK model with linear clearance using differential equations"
#>     ini({
#>         lka <- 0.45
#>         label("Absorption rate (Ka)")
#>         lcl <- 1
#>         label("Clearance (CL)")
#>         lvc <- 3.45
#>         label("Central volume of distribution (V)")
#>         propSd <- c(0, 0.5)
#>         label("Proportional residual error (fraction)")
#>     })
#>     model({
#>         ka <- exp(lka)
#>         cl <- exp(lcl)
#>         vc <- exp(lvc)
#>         kel <- cl/vc
#>         d/dt(depot) <- -ka * depot
#>         d/dt(central) <- ka * depot - kel * central
#>         Cc <- central/vc
#>         Cc ~ prop(propSd)
#>     })
#> }
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$fcn_def
#> [1] "function () \n{\n    description <- \"One compartment PK model with linear clearance using differential equations\"\n    ini({\n        lka <- 0.45\n        label(\"Absorption rate (Ka)\")\n        lcl <- 1\n        label(\"Clearance (CL)\")\n        lvc <- 3.45\n        label(\"Central volume of distribution (V)\")\n        propSd <- c(0, 0.5)\n        label(\"Proportional residual error (fraction)\")\n    })\n    model({\n        ka <- exp(lka)\n        cl <- exp(lcl)\n        vc <- exp(lvc)\n        kel <- cl/vc\n        d/dt(depot) <- -ka * depot\n        d/dt(central) <- ka * depot - kel * central\n        Cc <- central/vc\n        Cc ~ prop(propSd)\n    })\n}"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$MDLMETA
#> [1] "base model"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$code
#> [1] "MB_obj_1_fcn = function () \n{\n    description <- \"One compartment PK model with linear clearance using differential equations\"\n    ini({\n        lka <- 0.45\n        label(\"Absorption rate (Ka)\")\n        lcl <- 1\n        label(\"Clearance (CL)\")\n        lvc <- 3.45\n        label(\"Central volume of distribution (V)\")\n        propSd <- c(0, 0.5)\n        label(\"Proportional residual error (fraction)\")\n    })\n    model({\n        ka <- exp(lka)\n        cl <- exp(lcl)\n        vc <- exp(lvc)\n        kel <- cl/vc\n        d/dt(depot) <- -ka * depot\n        d/dt(central) <- ka * depot - kel * central\n        Cc <- central/vc\n        Cc ~ prop(propSd)\n    })\n}\nMB_obj_1_rx =  rxode2::rxode2(MB_obj_1_fcn)"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$checksum
#> [1] "bba5b8caae381920a547eb80c192bac6"
#> 
#> $CTS$MDL$mdl$MB_obj_1_rx$MDLchecksum
#> [1] "4c1cee671a16d68ac228a0860a4346de"
#> 
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx
#> $CTS$MDL$mdl$MB_obj_2_rx$label
#> [1] "mAb Model (Davda 2014)"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$MOD_TYPE
#> [1] "MB"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$id
#> [1] "MB"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$rx_obj
#>  ── rxode2-based solved PK 2-compartment model with first-order absorption ────── 
#>  ── Initalization: ──  
#> Fixed Effects ($theta): 
#>    lfdepot        lka        lcl         lv        lvp         lq     allocl 
#> -0.2957142 -1.2658482 -1.6094379  1.2837078  1.0116009 -0.2916901  0.8650000 
#>      allov   prop.err 
#>  0.9570000  0.1440000 
#> 
#> Omega ($omega): 
#>           etafdepot etaka  etacl   etav  etavp  etaq
#> etafdepot         0 0.000 0.0000 0.0000 0.0000 0.000
#> etaka             0 0.416 0.0000 0.0000 0.0000 0.000
#> etacl             0 0.000 0.0987 0.0786 0.0377 0.000
#> etav              0 0.000 0.0786 0.1160 0.0619 0.000
#> etavp             0 0.000 0.0377 0.0619 0.0789 0.000
#> etaq              0 0.000 0.0000 0.0000 0.0000 0.699
#>  ── μ-referencing ($muRefTable): ──  
#>     theta       eta level covariates
#> 1 lfdepot etafdepot    id           
#> 2     lka     etaka    id           
#> 3     lcl     etacl    id           
#> 4      lq      etaq    id           
#> 5      lv      etav    id           
#> 6     lvp     etavp    id           
#> 
#>  ── Model (Normalized Syntax): ── 
#> function() {
#>     description <- "Two compartment PK model with linear clearance for average monoclonal antibodies (Davda 2014)"
#>     reference <- "Davda JP, Dodds MG, Gibbs MA, Wisdom W, Gibbs JP. A model-based meta-analysis of monoclonal antibody pharmacokinetics to guide optimal first-in-human study design. MAbs. 2014;6(4):1094-1102. doi:10.4161/mabs.29095"
#>     ini({
#>         lfdepot <- -0.295714244149045
#>         label("Subcutaneous bioavailability (fraction)")
#>         lka <- -1.26584820804402
#>         label("Absorption rate (Ka, 1/day)")
#>         lcl <- -1.6094379124341
#>         label("Clearance (CL, L/day)")
#>         lv <- 1.28370777234479
#>         label("Central volume of distribution (V, L)")
#>         lvp <- 1.01160091167848
#>         label("Peripheral volume of distribution (Vp, L)")
#>         lq <- -0.29169009384932
#>         label("Intercompartmental clearance (Q, L/day)")
#>         allocl <- 0.865
#>         label("Allometric exponent on clearance and intercompartmental clearance (unitless)")
#>         allov <- 0.957
#>         label("Allometric exponent on volumes of distribution (unitless)")
#>         prop.err <- c(0, 0.144)
#>         label("Proportional residual error (fraction)")
#>         etafdepot ~ 0
#>         etaka ~ 0.416
#>         etacl + etav + etavp ~ c(0.0987, 0.0786, 0.116, 0.0377, 
#>             0.0619, 0.0789)
#>         etaq ~ 0.699
#>     })
#>     model({
#>         fdepot <- exp(lfdepot + etafdepot)
#>         ka <- exp(lka + etaka)
#>         wtnorm <- log(WT/70)
#>         cl <- exp(lcl + allocl * wtnorm + etacl)
#>         q <- exp(lq + allocl * wtnorm + etaq)
#>         v <- exp(lv + allov * wtnorm + etav)
#>         vp <- exp(lvp + allov * wtnorm + etavp)
#>         Cc <- linCmt()
#>         f(depot) <- fdepot
#>         Cc ~ prop(prop.err)
#>     })
#> }
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$fcn_def
#> [1] "function () \n{\n    description <- \"Two compartment PK model with linear clearance for average monoclonal antibodies (Davda 2014)\"\n    reference <- \"Davda JP, Dodds MG, Gibbs MA, Wisdom W, Gibbs JP. A model-based meta-analysis of monoclonal antibody pharmacokinetics to guide optimal first-in-human study design. MAbs. 2014;6(4):1094-1102. doi:10.4161/mabs.29095\"\n    ini({\n        lfdepot <- -0.295714244149045\n        label(\"Subcutaneous bioavailability (fraction)\")\n        lka <- -1.26584820804402\n        label(\"Absorption rate (Ka, 1/day)\")\n        lcl <- -1.6094379124341\n        label(\"Clearance (CL, L/day)\")\n        lv <- 1.28370777234479\n        label(\"Central volume of distribution (V, L)\")\n        lvp <- 1.01160091167848\n        label(\"Peripheral volume of distribution (Vp, L)\")\n        lq <- -0.29169009384932\n        label(\"Intercompartmental clearance (Q, L/day)\")\n        allocl <- 0.865\n        label(\"Allometric exponent on clearance and intercompartmental clearance (unitless)\")\n        allov <- 0.957\n        label(\"Allometric exponent on volumes of distribution (unitless)\")\n        prop.err <- c(0, 0.144)\n        label(\"Proportional residual error (fraction)\")\n        etafdepot ~ 0\n        etaka ~ 0.416\n        etacl + etav + etavp ~ c(0.0987, 0.0786, 0.116, 0.0377, \n            0.0619, 0.0789)\n        etaq ~ 0.699\n    })\n    model({\n        fdepot <- exp(lfdepot + etafdepot)\n        ka <- exp(lka + etaka)\n        wtnorm <- log(WT/70)\n        cl <- exp(lcl + allocl * wtnorm + etacl)\n        q <- exp(lq + allocl * wtnorm + etaq)\n        v <- exp(lv + allov * wtnorm + etav)\n        vp <- exp(lvp + allov * wtnorm + etavp)\n        Cc <- linCmt()\n        f(depot) <- fdepot\n        Cc ~ prop(prop.err)\n    })\n}"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$MDLMETA
#> [1] "base model"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$code
#> [1] "MB_obj_2_fcn = function () \n{\n    description <- \"Two compartment PK model with linear clearance for average monoclonal antibodies (Davda 2014)\"\n    reference <- \"Davda JP, Dodds MG, Gibbs MA, Wisdom W, Gibbs JP. A model-based meta-analysis of monoclonal antibody pharmacokinetics to guide optimal first-in-human study design. MAbs. 2014;6(4):1094-1102. doi:10.4161/mabs.29095\"\n    ini({\n        lfdepot <- -0.295714244149045\n        label(\"Subcutaneous bioavailability (fraction)\")\n        lka <- -1.26584820804402\n        label(\"Absorption rate (Ka, 1/day)\")\n        lcl <- -1.6094379124341\n        label(\"Clearance (CL, L/day)\")\n        lv <- 1.28370777234479\n        label(\"Central volume of distribution (V, L)\")\n        lvp <- 1.01160091167848\n        label(\"Peripheral volume of distribution (Vp, L)\")\n        lq <- -0.29169009384932\n        label(\"Intercompartmental clearance (Q, L/day)\")\n        allocl <- 0.865\n        label(\"Allometric exponent on clearance and intercompartmental clearance (unitless)\")\n        allov <- 0.957\n        label(\"Allometric exponent on volumes of distribution (unitless)\")\n        prop.err <- c(0, 0.144)\n        label(\"Proportional residual error (fraction)\")\n        etafdepot ~ 0\n        etaka ~ 0.416\n        etacl + etav + etavp ~ c(0.0987, 0.0786, 0.116, 0.0377, \n            0.0619, 0.0789)\n        etaq ~ 0.699\n    })\n    model({\n        fdepot <- exp(lfdepot + etafdepot)\n        ka <- exp(lka + etaka)\n        wtnorm <- log(WT/70)\n        cl <- exp(lcl + allocl * wtnorm + etacl)\n        q <- exp(lq + allocl * wtnorm + etaq)\n        v <- exp(lv + allov * wtnorm + etav)\n        vp <- exp(lvp + allov * wtnorm + etavp)\n        Cc <- linCmt()\n        f(depot) <- fdepot\n        Cc ~ prop(prop.err)\n    })\n}\nMB_obj_2_rx =  rxode2::rxode2(MB_obj_2_fcn)"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$checksum
#> [1] "bba5b8caae381920a547eb80c192bac6"
#> 
#> $CTS$MDL$mdl$MB_obj_2_rx$MDLchecksum
#> [1] "5e6fc90200958cc3f4f1d979475b1e86"
#> 
#> 
#> 
#> 
#> $CTS$ui_ele
#>  [1] "element_name"            "rule_condition"         
#>  [3] "action_dosing_state"     "action_dosing_values"   
#>  [5] "action_dosing_times"     "action_dosing_durations"
#>  [7] "action_set_state_state"  "action_set_state_value" 
#>  [9] "action_manual_code"      "cts_config_atol"        
#> [11] "cts_config_rtol"         "cts_config_sigdig"      
#> 
#> $CTS$element_cntr
#> [1] 1
#> 
#> $CTS$sc_meta
#> $CTS$sc_meta$config
#> $CTS$sc_meta$config$atol
#> $CTS$sc_meta$config$atol$group
#> [1] "ODE Solver"
#> 
#> $CTS$sc_meta$config$atol$label
#> [1] "Absolute tolerance"
#> 
#> $CTS$sc_meta$config$atol$tooltip
#> [1] "Absolute tolerance used by the ODE solver to determine if a good solution has been achieved; This is also used in the solved linear model to check if prior doses do not add anything to the solution."
#> 
#> $CTS$sc_meta$config$atol$type
#> [1] "numeric"
#> 
#> $CTS$sc_meta$config$atol$value
#> [1] "1e-8"
#> 
#> $CTS$sc_meta$config$atol$options
#> NULL
#> 
#> $CTS$sc_meta$config$atol$use
#> [1] "rxSolve"
#> 
#> $CTS$sc_meta$config$atol$ui
#> [1] "cts_config_atol"
#> 
#> 
#> $CTS$sc_meta$config$rtol
#> $CTS$sc_meta$config$rtol$group
#> [1] "ODE Solver"
#> 
#> $CTS$sc_meta$config$rtol$label
#> [1] "Absolute tolerance"
#> 
#> $CTS$sc_meta$config$rtol$tooltip
#> [1] "Relative tolerance used by the ODE solver to determine if a good solution has been achieved. This is also used in the solved linear model to check if prior doses do not add anything to the solution."
#> 
#> $CTS$sc_meta$config$rtol$type
#> [1] "numeric"
#> 
#> $CTS$sc_meta$config$rtol$value
#> [1] "1e-6"
#> 
#> $CTS$sc_meta$config$rtol$options
#> NULL
#> 
#> $CTS$sc_meta$config$rtol$use
#> [1] "rxSolve"
#> 
#> $CTS$sc_meta$config$rtol$ui
#> [1] "cts_config_rtol"
#> 
#> 
#> $CTS$sc_meta$config$sigdig
#> $CTS$sc_meta$config$sigdig$group
#> [1] "ODE Solver"
#> 
#> $CTS$sc_meta$config$sigdig$label
#> [1] "Absolute tolerance"
#> 
#> $CTS$sc_meta$config$sigdig$tooltip
#> [1] "Specifies the \"significant digits\" that the ode solving requests. When specified this controls the relative and absolute tolerances of the ODE solvers. By default the tolerance is 0.5*10^(-sigdig-2) for regular ODEs. For the sensitivity equations the default is ⁠0.5*10\\^(-sigdig-1.5)⁠ (sensitivity changes only applicable for liblsoda). This also controls the atol/rtol of the steady state solutions. The ssAtol/ssRtol is ⁠0.5*10\\^(-sigdig)⁠ and for the sensitivities ⁠0.5*10\\^(-sigdig+0.625)⁠. By default this is unspecified (NULL) and uses the standard atol/rtol."
#> 
#> $CTS$sc_meta$config$sigdig$type
#> [1] "numeric"
#> 
#> $CTS$sc_meta$config$sigdig$value
#> NULL
#> 
#> $CTS$sc_meta$config$sigdig$options
#> NULL
#> 
#> $CTS$sc_meta$config$sigdig$use
#> [1] "rxSolve"
#> 
#> $CTS$sc_meta$config$sigdig$ui
#> [1] "cts_config_sigdig"
#> 
#> 
#> 
#> $CTS$sc_meta$ui_config
#> [1] "cts_config_atol"   "cts_config_rtol"   "cts_config_sigdig"
#> 
#> 
#> $CTS$elements
#> $CTS$elements$element_1
#> $CTS$elements$element_1$isgood
#> [1] TRUE
#> 
#> $CTS$elements$element_1$ui
#> $CTS$elements$element_1$ui$element_name
#> [1] "cohort 1"
#> 
#> $CTS$elements$element_1$ui$source_model
#> [1] "MB_obj_1_rx"
#> 
#> $CTS$elements$element_1$ui$cts_config_atol
#> [1] "1e-8"
#> 
#> $CTS$elements$element_1$ui$cts_config_rtol
#> [1] "1e-6"
#> 
#> 
#> $CTS$elements$element_1$id
#> [1] "element_1"
#> 
#> $CTS$elements$element_1$idx
#> [1] 1
#> 
#> $CTS$elements$element_1$element_object_name
#> [1] "CTS_obj_1"
#> 
#> $CTS$elements$element_1$code_previous
#> [1] ""
#> 
#> $CTS$elements$element_1$rx_details
#> $CTS$elements$element_1$rx_details$isgood
#> [1] TRUE
#> 
#> $CTS$elements$element_1$rx_details$msgs
#> NULL
#> 
#> $CTS$elements$element_1$rx_details$elements
#> $CTS$elements$element_1$rx_details$elements$covariates
#> character(0)
#> 
#> $CTS$elements$element_1$rx_details$elements$parameters
#> [1] "lka"    "lcl"    "lvc"    "propSd"
#> 
#> $CTS$elements$element_1$rx_details$elements$iiv
#> character(0)
#> 
#> $CTS$elements$element_1$rx_details$elements$states
#> [1] "depot"   "central"
#> 
#> 
#> $CTS$elements$element_1$rx_details$txt_info
#> [1] "States/Compartments\ndepot, central\n\nCovariates\nNone found\nParameters\nlka, lcl, lvc, propSd\n\nBetween-subject variability\nNone found\n"
#> 
#> $CTS$elements$element_1$rx_details$ht_info
#> <b>States/Compartments</b>
#> <br/>
#> depot, central
#> <br/>
#> <br/>
#> <b>Covariates</b>
#> <br/>
#> <em>None found</em>
#> <br/>
#> <br/>
#> <b>Parameters</b>
#> <br/>
#> lka, lcl, lvc, propSd
#> <br/>
#> <br/>
#> <b>Between-subject variability</b>
#> <br/>
#> <em>None found</em>
#> <br/>
#> <br/>
#> 
#> 
#> $CTS$elements$element_1$components_table
#> NULL
#> 
#> $CTS$elements$element_1$checksum
#> [1] "92b2bb9eac2df77bcdfcbe38eefa51f4"
#> 
#> $CTS$elements$element_1$code
#> NULL
#> 
#> $CTS$elements$element_1$code_ele_only
#> NULL
#> 
#> 
#> 
#> $CTS$current_element
#> [1] "element_1"
#> 
#> $CTS$checksum
#> [1] "4dc8f03bdf3ae334d0f4ebef877e80c8"
#> 
#> 
#> $MOD_TYPE
#> [1] "CTS"
#> 
#> $id
#> [1] "CTS"
#> 
#> $FM_yaml_file
#> [1] "/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/formods/templates/formods.yaml"
#> 
#> $MOD_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/RtmpZZjDuH/temp_libpathd8e060f606a0/ruminate/templates/CTS.yaml"
#> 
#> $shiny_token
#> [1] "non_shiny"
#>