New ===Module_Name=== cohort

Appends a new empty cohort to the CTS state object and makes this new cohort the active cohort.

Usage

CTS_new_element(state)

Arguments

state

CTS state from CTS_fetch_state()

Value

CTS state object containing a new cohort and that cohort is set as the current active cohort. See the help for CTS_fetch_state() for ===ELEMENT== format.

Examples

# For more information see the Clinical Trial Simulation vignette:
# https://ruminate.ubiquity.tools/articles/clinical_trial_simulation.html

# None of this will work if rxode2 isn't installed:
library(formods)
if(is_installed("rxode2")){

sess_res = CTS_test_mksession(session=list(), full_session=FALSE)
state = sess_res$state
session = sess_res$session
input   = sess_res$input

# Creates a new empty element
state = CTS_new_element(state)

# Delete the current element
state = CTS_del_current_element(state)

# Fetch a list of the current element
element = CTS_fetch_current_element(state)

# You can modify the element
element[["element_name"]] = "A more descriptive name"

# Simulating the current element
element = CTS_simulate_element(state, element)

# You can now place element back in the state
state = CTS_set_current_element(state, element)

# This will extract the code for the current module
code = CTS_fetch_code(state)
code

# This will update the checksum of the module state
state = CTS_update_checksum(state)

#'@example inst/test_apps/CTS_funcs.R
}
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#> → MB: module checksum updated:a8d5d330f91feb7be769e08518458dd8
#> → MB: State initialized
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: fd50f686258ced0cf49f094cee90f1aa
#> → MB: module checksum updated:3557cd4e3571b5414f8f3037ad7b29c8
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → CTS:   dest:   file.path("config","report.docx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → CTS:   dest:   file.path("config","report.pptx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → CTS:   dest:   file.path("config","report.yaml")
#> → CTS: module checksum updated: 4affbc8f5039c017fb6d555c8aa0f450
#> → CTS: State initialized
#> → CTS: cohort checksum updated: 7b7f6b1ffa36a534ff4d377fb5317f69
#> → CTS: module checksum updated: 15b4e3d941fb672f0346104f34aa2149
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: c4e664d7b349ffd79d757f6c40d0c3f3
#> → CTS: module checksum updated: 63cccedeb8c1f215b5f1bdb3251cf32f
#>  
#>  
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> → CTS: cohort checksum updated: c8c9f2770a80b6e684999ad3280cf89f
#> → CTS: module checksum updated: c5408dd76d32b14bec6868a873035dc1
#>  
#>  
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> → CTS: cohort checksum updated: 5ccbfa31ae9f099494aca239222cee8a
#> → CTS: module checksum updated: abb64ec2c533d6039bc04f28cb32c258