Fetches List of Available Models

Creates a catalog of the models available in the system file.

Usage

MB_fetch_catalog(state)

Arguments

state

MB state from MB_fetch_state()

Value

List with the following attributes:

summary: Dataframe with a summary of the models in the catlog

sources: Same information a that found in the summary table but in list form.

select_group: List with the models grouped by source.

select_plain: Flat list with the models (ungrouped).

select_subtext: Subtext for pulldown menus.

msgs: Messages to be passed back to the user.

hasmdl: Boolean value indicating if any models were found.

isgood: Boolean variable indicating success or failure.

Examples

library(ruminate)
# This will get the full session:
sess_res = MB_test_mksession(session=list(), full_session=TRUE)
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6
#> → MB: State initialized
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6
# This is just for CRAN
#sess_res = MB_test_mksession(session=list())
session = sess_res$session
input   = sess_res$input

# Configuration files
FM_yaml_file  = system.file(package = "formods", "templates", "formods.yaml")
MOD_yaml_file = system.file(package = "ruminate", "templates", "MB.yaml")

# Creating an empty state object
state = MB_fetch_state(id              = "MB",
                       input           = input,
                       session         = session,
                       FM_yaml_file    = FM_yaml_file,
                       MOD_yaml_file   = MOD_yaml_file,
                       react_state     = NULL)
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6
#> → MB: State initialized
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6

# This will provide a list of the available models
models = MB_fetch_catalog(state)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
# This is a summary of the tables in the model:
models$summary
#> NULL

# This will test the models in the catalog, set as_cran
# to FALSE to test all the models.
mtres = MB_test_catalog(state, as_cran=TRUE)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
mtres$isgood
#> [1] FALSE

# Creates a new empty element
state = MB_new_element(state)

# Delete the current element
state = MB_del_current_element(state)

# Fetch a list of the current element
element = MB_fetch_current_element(state)

# This will attach a model to it:
# Pulling the first model from the catalog
fcn_def  = models[["summary"]][1, ][["Model"]]
fcn_obj  = models[["summary"]][1, ][["Object"]]
mdl_type = models[["summary"]][1, ][["Type"]]
fcn_desc = models[["summary"]][1, ][["Description"]]

# This will build the rxode2 object from the model
mk_rx_res = mk_rx_obj(
  type  = mdl_type,
  model = list(fcn_def = fcn_def,
               fcn_obj = fcn_obj))

# This will attach the model to the current element
element = MB_update_model(
  state       = state,
  session     = session,
  current_ele = element,
  rx_obj      = mk_rx_res[["capture"]][["rx_obj"]],
  note        = fcn_desc,
  reset       = TRUE)

# You can now place element back in the state
state = MB_set_current_element(state, element)
#> → MB: module checksum updated:e5352fb5e1e48e8e151dcd9d80cfbfa1

# This will fetch the current component
component = MB_fetch_component(state, element)

# You can use the component to build the code to generate the model:
gen_code =
  MB_build_code(state        = state, session = session,
              fcn_def      = component[["fcn_def"]],
              fcn_obj_name = "my_fcn_obj",
              rx_obj_name  = "my_obj_name")

# Model code to be included in a larger script
cat(paste0(gen_code$model_code, collapse="\n"))
#> # rxode2 package was not found.

# Stand-alone model code
cat(paste0(gen_code$model_code_sa, collapse="\n"))
#> 
#> # rxode2 package was not found.

# JMH to add:
# MB_fetch_code