Fetch Model Builder Module Models

Fetches the models contained in the module.

Usage

MB_fetch_mdl(state)

Arguments

state

MB state from MB_fetch_state()

Value

list containing the following elements

isgood: Return status of the function.

hasmdl: Boolean indicator if the module has any models

msgs: Messages to be passed back to the user.

mdl: List with models. Each list element has the name of the R-object for that dataset. Each element has the following structure:

label: Text label for the model (e.g. one-compartment model).

MOD_TYPE: Type of module.

id: Module ID.

rx_obj: The rxode2 object.

rx_obj_name: The rxode2 object name that holds the model.

ts_obj List with elements system and details

ts_obj_name: The object name that holds the model time scale information.

fcn_def: Text to define the model

MDLMETA: Notes about the model.

code: Code to generate the model.

checksum: Module checksum.

MDLchecksum: Model checksum.

Examples

# We need a module state:
sess_res = MB_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> → MB: module checksum updated:f7de2055542d3ff1b085fafc2d30a1f8
#> → MB: State initialized
#> → MB: loading model idx: 1
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 2e54d844608917f8ae4068ee7ff1a6e3
#> → MB: module checksum updated:37572852285f9f23e95300a55eab68fb
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 0010f620387c815d825ed50cbc0df9e2
#> → MB: module checksum updated:d104251f1f2abde1da87637e43095824
#> → MB:   -> setting name: One compartment model
#> → MB:   -> setting time scale: hours
#> → MB:   -> setting base from: user
#> → MB:   -> setting catalog selection: 
#> → MB:   -> setting base model id: manual
#> → MB:   -> setting base model name: manual
#> → MB: model checksum updated: 037c7d41ff7567cf6c1c0f6111a46cff
#> → MB: module checksum updated:ffd42a1cd1e61b9f1f89f1b9fc383b4b
#> → MB: added element idx: 1
#> → MB: loading model idx: 2
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: d6ab490d1aaa5578c932378054fce466
#> → MB: module checksum updated:5f2d8029a721bdb31cea39455794754b
#> → MB:   -> setting name: PK Biomarker
#> → MB:   -> setting time scale: days
#> → MB:   -> setting base from: user
#> → MB:   -> setting catalog selection: 
#> → MB:   -> setting base model id: manual
#> → MB:   -> setting base model name: manual
#> → MB: model checksum updated: 6998cab7e8a506d52c4dd172e379c4eb
#> → MB: module checksum updated:d91fd187fd383d2351e254f4d542ba0f
#> → MB: added element idx: 2
#> → MB: module isgood: TRUE
state = sess_res$state

mdls = MB_fetch_mdl(state)

names(mdls)
#> [1] "hasmdl"     "isgood"     "ts_details" "msgs"       "mdl"