Initialize MB Module State

Creates a list of the initialized module state

Usage

MB_init_state(FM_yaml_file, MOD_yaml_file, id, session)

Arguments

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

id

ID string for the module.

session

Shiny session variable

Value

list containing an empty MB state

Examples

# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = MB_test_mksession(session=list(), full_session=FALSE)
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6
#> → MB: State initialized
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6
session = sess_res$session
input   = sess_res$input

state = MB_init_state(
   FM_yaml_file  = system.file(package = "formods",
                               "templates",
                               "formods.yaml"),
   MOD_yaml_file = system.file(package = "ruminate",
                               "templates",
                               "MB.yaml"),
   id              = "MB",
   session         = session)
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> → MB: module checksum updated:8ac755447dc73c185137d18702a2cdb6
#> → MB: State initialized

state
#> $yaml
#> $yaml$FM
#> $yaml$FM$include
#> $yaml$FM$include$files
#> $yaml$FM$include$files[[1]]
#> $yaml$FM$include$files[[1]]$file
#> $yaml$FM$include$files[[1]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.docx\")"
#> 
#> $yaml$FM$include$files[[1]]$file$dest
#> [1] "file.path(\"config\",\"report.docx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[2]]
#> $yaml$FM$include$files[[2]]$file
#> $yaml$FM$include$files[[2]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.pptx\")"
#> 
#> $yaml$FM$include$files[[2]]$file$dest
#> [1] "file.path(\"config\",\"report.pptx\")"
#> 
#> 
#> 
#> $yaml$FM$include$files[[3]]
#> $yaml$FM$include$files[[3]]$file
#> $yaml$FM$include$files[[3]]$file$source
#> [1] "file.path(system.file(package=\"onbrand\"), \"templates\", \"report.yaml\")"
#> 
#> $yaml$FM$include$files[[3]]$file$dest
#> [1] "file.path(\"config\",\"report.yaml\")"
#> 
#> 
#> 
#> 
#> 
#> $yaml$FM$deployed
#> [1] FALSE
#> 
#> $yaml$FM$code
#> $yaml$FM$code$theme
#> [1] "vibrant_ink"
#> 
#> $yaml$FM$code$showLineNumbers
#> [1] TRUE
#> 
#> $yaml$FM$code$gen_file
#> [1] "run_analysis.R"
#> 
#> $yaml$FM$code$gen_preamble
#> [1] "# formods automated output ------------------------------------------------\n# https://formods.ubiquity.tools/\nrm(list=ls())"
#> 
#> $yaml$FM$code$packages
#> [1] "onbrand" "writexl"
#> 
#> 
#> $yaml$FM$notifications
#> $yaml$FM$notifications$config
#> $yaml$FM$notifications$config$success
#> $yaml$FM$notifications$config$success$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$success$background
#> [1] "#5bb85b"
#> 
#> 
#> $yaml$FM$notifications$config$failure
#> $yaml$FM$notifications$config$failure$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$failure$background
#> [1] "#d9534f"
#> 
#> 
#> $yaml$FM$notifications$config$info
#> $yaml$FM$notifications$config$info$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$info$background
#> [1] "#5bc0de"
#> 
#> 
#> $yaml$FM$notifications$config$warning
#> $yaml$FM$notifications$config$warning$useFontAwesome
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$useIcon
#> [1] FALSE
#> 
#> $yaml$FM$notifications$config$warning$background
#> [1] "#f0ac4d"
#> 
#> 
#> 
#> 
#> $yaml$FM$reporting
#> $yaml$FM$reporting$enabled
#> [1] TRUE
#> 
#> $yaml$FM$reporting$content_init
#> $yaml$FM$reporting$content_init$xlsx
#> [1] "rpt = list(summary = NULL,\n           sheets  = list())"
#> 
#> $yaml$FM$reporting$content_init$docx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.docx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> $yaml$FM$reporting$content_init$pptx
#> [1] "rpt  = onbrand::read_template(\n  template = file.path(\"config\", \"report.pptx\"),\n  mapping  = file.path(\"config\", \"report.yaml\"))"
#> 
#> 
#> 
#> $yaml$FM$ui
#> $yaml$FM$ui$button_style
#> [1] "fill"
#> 
#> $yaml$FM$ui$select_size
#> [1] 10
#> 
#> $yaml$FM$ui$color_green
#> [1] "#00BB8A"
#> 
#> $yaml$FM$ui$color_red
#> [1] "#FF475E"
#> 
#> $yaml$FM$ui$color_blue
#> [1] "#0088FF"
#> 
#> $yaml$FM$ui$color_purple
#> [1] "#bd2cf4"
#> 
#> 
#> $yaml$FM$data_meta
#> $yaml$FM$data_meta$data_header
#> [1] "<span style='color:===COLOR==='><b>===NAME===</b><br/><font size='-3'>===LABEL===</font></span>"
#> 
#> $yaml$FM$data_meta$subtext
#> [1] "===LABEL===: ===RANGE==="
#> 
#> $yaml$FM$data_meta$many_sep
#> [1] ",⋅⋅⋅,"
#> 
#> $yaml$FM$data_meta$data_types
#> $yaml$FM$data_meta$data_types$character
#> $yaml$FM$data_meta$data_types$character$color
#> [1] "#DD4B39"
#> 
#> $yaml$FM$data_meta$data_types$character$label
#> [1] "text"
#> 
#> 
#> $yaml$FM$data_meta$data_types$double
#> $yaml$FM$data_meta$data_types$double$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$double$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$integer
#> $yaml$FM$data_meta$data_types$integer$color
#> [1] "#3C8DBC"
#> 
#> $yaml$FM$data_meta$data_types$integer$label
#> [1] "num"
#> 
#> 
#> $yaml$FM$data_meta$data_types$other
#> $yaml$FM$data_meta$data_types$other$color
#> [1] "black"
#> 
#> $yaml$FM$data_meta$data_types$other$label
#> [1] "other"
#> 
#> 
#> 
#> 
#> $yaml$FM$labels
#> $yaml$FM$labels$default_ds
#> [1] "Original data set"
#> 
#> 
#> $yaml$FM$user_files
#> $yaml$FM$user_files$use_tmpdir
#> [1] TRUE
#> 
#> 
#> $yaml$FM$logging
#> $yaml$FM$logging$enabled
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp
#> [1] TRUE
#> 
#> $yaml$FM$logging$timestamp_fmt
#> [1] "%Y-%m-%d %H:%M:%S"
#> 
#> $yaml$FM$logging$log_file
#> [1] "formods_log.txt"
#> 
#> $yaml$FM$logging$console
#> [1] TRUE
#> 
#> 
#> 
#> 
#> $MC
#> $MC$element_object_name
#> [1] "MB_obj"
#> 
#> $MC$sources
#> $MC$sources[[1]]
#> $MC$sources[[1]]$source
#> $MC$sources[[1]]$source$group
#> [1] "User-defined Models"
#> 
#> $MC$sources[[1]]$source$name
#> [1] "rxode2 User model"
#> 
#> $MC$sources[[1]]$source$description
#> [1] "User-defined rxode2 model"
#> 
#> $MC$sources[[1]]$source$type
#> [1] "rxode2"
#> 
#> $MC$sources[[1]]$source$obj
#> [1] "my_fcn"
#> 
#> $MC$sources[[1]]$source$file
#> [1] "file.path(getwd(),\"user_model.R\")"
#> 
#> 
#> 
#> $MC$sources[[2]]
#> $MC$sources[[2]]$source
#> $MC$sources[[2]]$source$group
#> [1] "User-defined Models"
#> 
#> $MC$sources[[2]]$source$name
#> [1] "NONMEM User model"
#> 
#> $MC$sources[[2]]$source$description
#> [1] "User-defined NONMEM model"
#> 
#> $MC$sources[[2]]$source$type
#> [1] "NONMEM"
#> 
#> $MC$sources[[2]]$source$file
#> [1] "file.path(getwd(),\"user_model.ctl\")"
#> 
#> 
#> 
#> $MC$sources[[3]]
#> $MC$sources[[3]]$source
#> $MC$sources[[3]]$source$group
#> [1] "Model Library"
#> 
#> $MC$sources[[3]]$source$type
#> [1] "nlmixr2lib"
#> 
#> $MC$sources[[3]]$source$name
#> [1] "nlmixr2 Model Library"
#> 
#> 
#> 
#> 
#> $MC$code
#> $MC$code$packages
#> [1] "rxode2"     "nonmem2rx"  "nlmixr2lib"
#> 
#> $MC$code$readOnly
#> [1] TRUE
#> 
#> $MC$code$mode
#> [1] "r"
#> 
#> 
#> $MC$compact
#> $MC$compact$code
#> [1] TRUE
#> 
#> $MC$compact$clip
#> [1] TRUE
#> 
#> 
#> $MC$reporting
#> $MC$reporting$enabled
#> [1] FALSE
#> 
#> $MC$reporting$priority
#> [1] 1
#> 
#> 
#> $MC$formatting
#> $MC$formatting$code
#> $MC$formatting$code$width
#> [1] 800
#> 
#> $MC$formatting$code$height
#> [1] 300
#> 
#> 
#> $MC$formatting$preview
#> $MC$formatting$preview$width
#> [1] "800px"
#> 
#> $MC$formatting$preview$height
#> [1] "500px"
#> 
#> 
#> $MC$formatting$input
#> NULL
#> 
#> $MC$formatting$current_element
#> $MC$formatting$current_element$width
#> [1] "200px"
#> 
#> $MC$formatting$current_element$tooltip
#> [1] "Change the current model."
#> 
#> $MC$formatting$current_element$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$element_name
#> $MC$formatting$element_name$width
#> [1] "200px"
#> 
#> 
#> $MC$formatting$button_clk_run
#> $MC$formatting$button_clk_run$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_run$block
#> [1] TRUE
#> 
#> 
#> $MC$formatting$button_clk_del
#> $MC$formatting$button_clk_del$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_del$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_del$tooltip
#> [1] "Delete the current model."
#> 
#> $MC$formatting$button_clk_del$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_save
#> $MC$formatting$button_clk_save$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_save$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_save$tooltip
#> [1] "Save the name for the current model."
#> 
#> $MC$formatting$button_clk_save$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_clip
#> $MC$formatting$button_clk_clip$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_clip$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_clip$tooltip
#> [1] "Copy the code to generate the current model to the clipboard."
#> 
#> $MC$formatting$button_clk_clip$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_copy
#> $MC$formatting$button_clk_copy$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_copy$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_copy$tooltip
#> [1] "Make a copy of the current model."
#> 
#> $MC$formatting$button_clk_copy$tooltip_position
#> [1] "right"
#> 
#> 
#> $MC$formatting$button_clk_new
#> $MC$formatting$button_clk_new$size
#> [1] "sm"
#> 
#> $MC$formatting$button_clk_new$block
#> [1] TRUE
#> 
#> $MC$formatting$button_clk_new$tooltip
#> [1] "Create a new model."
#> 
#> $MC$formatting$button_clk_new$tooltip_position
#> [1] "right"
#> 
#> $MC$formatting$button_clk_new$width
#> [1] 100
#> 
#> 
#> $MC$formatting$catalog_selection
#> $MC$formatting$catalog_selection$width
#> [1] "200px"
#> 
#> $MC$formatting$catalog_selection$truncate
#> [1] 65
#> 
#> $MC$formatting$catalog_selection$tooltip
#> [1] "Select base model from the catalog."
#> 
#> $MC$formatting$catalog_selection$tooltip_position
#> [1] "top"
#> 
#> 
#> $MC$formatting$model_type_selection
#> $MC$formatting$model_type_selection$width
#> [1] "200px"
#> 
#> $MC$formatting$model_type_selection$tooltip
#> [1] "Select the type of the model being uploaded."
#> 
#> $MC$formatting$model_type_selection$tooltip_position
#> [1] "right"
#> 
#> $MC$formatting$model_type_selection$choices
#> $MC$formatting$model_type_selection$choices$rxode2
#> [1] "rxode2 function"
#> 
#> $MC$formatting$model_type_selection$choices$NONMEM
#> [1] "NONMEM (XML, lst, or ctl)"
#> 
#> 
#> $MC$formatting$model_type_selection$default
#> [1] "rxode2"
#> 
#> 
#> $MC$formatting$upload_model_file
#> $MC$formatting$upload_model_file$width
#> [1] "400px"
#> 
#> 
#> $MC$formatting$base_from
#> $MC$formatting$base_from$size
#> [1] "normal"
#> 
#> $MC$formatting$base_from$status
#> [1] "primary"
#> 
#> $MC$formatting$base_from$tooltip
#> [1] "You can select the base model from a catalog or upload your own"
#> 
#> $MC$formatting$base_from$tooltip_position
#> [1] "bottom"
#> 
#> $MC$formatting$base_from$choices
#> $MC$formatting$base_from$choices$catalog
#> [1] "Model Catalog"
#> 
#> $MC$formatting$base_from$choices$user
#> [1] "User-defined model"
#> 
#> 
#> $MC$formatting$base_from$default
#> [1] "catalog"
#> 
#> 
#> $MC$formatting$time_scales
#> $MC$formatting$time_scales$width
#> [1] "200px"
#> 
#> $MC$formatting$time_scales$default
#> [1] "weeks"
#> 
#> $MC$formatting$time_scales$tooltip
#> [1] "Choose the timescale of the model."
#> 
#> $MC$formatting$time_scales$tooltip_position
#> [1] "top"
#> 
#> $MC$formatting$time_scales$choices
#> $MC$formatting$time_scales$choices$months
#> $MC$formatting$time_scales$choices$months$conv
#> [1] "60*60*24*4"
#> 
#> $MC$formatting$time_scales$choices$months$verb
#> [1] "Months"
#> 
#> 
#> $MC$formatting$time_scales$choices$weeks
#> $MC$formatting$time_scales$choices$weeks$conv
#> [1] "60*60*24*7"
#> 
#> $MC$formatting$time_scales$choices$weeks$verb
#> [1] "Weeks"
#> 
#> 
#> $MC$formatting$time_scales$choices$days
#> $MC$formatting$time_scales$choices$days$conv
#> [1] "60*60*24"
#> 
#> $MC$formatting$time_scales$choices$days$verb
#> [1] "Days"
#> 
#> 
#> $MC$formatting$time_scales$choices$hours
#> $MC$formatting$time_scales$choices$hours$conv
#> [1] "60*60"
#> 
#> $MC$formatting$time_scales$choices$hours$verb
#> [1] "Hours"
#> 
#> 
#> $MC$formatting$time_scales$choices$min
#> $MC$formatting$time_scales$choices$min$conv
#> [1] "60"
#> 
#> $MC$formatting$time_scales$choices$min$verb
#> [1] "Minutes"
#> 
#> 
#> $MC$formatting$time_scales$choices$sec
#> $MC$formatting$time_scales$choices$sec$conv
#> [1] "1"
#> 
#> $MC$formatting$time_scales$choices$sec$verb
#> [1] "Seconds"
#> 
#> 
#> 
#> 
#> 
#> $MC$labels
#> $MC$labels$current_element
#> NULL
#> 
#> $MC$labels$element_name
#> NULL
#> 
#> $MC$labels$upload_model_file
#> NULL
#> 
#> $MC$labels$base_from
#> NULL
#> 
#> $MC$labels$catalog_selection
#> NULL
#> 
#> $MC$labels$model_type_selection
#> NULL
#> 
#> $MC$labels$time_scale
#> NULL
#> 
#> $MC$labels$catalog_empty
#> [1] "No models were found, the catalog is empty"
#> 
#> $MC$labels$save_btn
#> [1] "Save"
#> 
#> $MC$labels$clip_btn
#> [1] "Code"
#> 
#> $MC$labels$copy_btn
#> [1] "Copy"
#> 
#> $MC$labels$del_btn
#> [1] "Delete"
#> 
#> $MC$labels$new_btn
#> [1] "New"
#> 
#> $MC$labels$building_model
#> [1] "Building model, be patient."
#> 
#> $MC$labels$element_name_diff
#> [1] "The model name has changed."
#> 
#> $MC$labels$model_code_diff
#> [1] "Manual changes have been made to the model."
#> 
#> $MC$labels$save_change_detected
#> [1] "You need to click on the save button for these changes to take effect."
#> 
#> $MC$labels$head_base_model
#> [1] "Starting model"
#> 
#> $MC$labels$head_model_code
#> [1] "Model code"
#> 
#> $MC$labels$head_time_scale
#> [1] "Model time-scale"
#> 
#> 
#> $MC$errors
#> $MC$errors$no_model_found
#> [1] "No model was found."
#> 
#> $MC$errors$base_model_build_failed
#> [1] "The base model build failed."
#> 
#> $MC$errors$manual_model_update_failed
#> [1] "Manual model update failed."
#> 
#> $MC$errors$user_file_upload_failed
#> [1] "User-defined model failed."
#> 
#> $MC$errors$nlmixr2lib_not_found
#> [1] "The nlmixr2lib package was not found. This library will not be available."
#> 
#> $MC$errors$selected_id_bad_list
#> [1] "Unable to find a list for the selected component."
#> 
#> $MC$errors$selected_id_bad_row
#> [1] "Selected component should have 1 row."
#> 
#> 
#> $MC$tooltips
#> $MC$tooltips$include
#> [1] TRUE
#> 
#> $MC$tooltips$elements
#> [1] "Show model components"
#> 
#> $MC$tooltips$show_code
#> [1] "Show model code"
#> 
#> $MC$tooltips$url_rxode
#> [1] "https://nlmixr2.github.io/rxode2/articles/rxode2-syntax.html"
#> 
#> $MC$tooltips$url_model_types
#> [1] "https://ruminate.ubiquity.tools/articles/model_builder.html#model-catalog"
#> 
#> $MC$tooltips$components
#> $MC$tooltips$components$input_tip
#> [1] "My tool tip"
#> 
#> 
#> 
#> 
#> $MB
#> $MB$button_counters
#> $MB$button_counters$button_clk_save
#> [1] 0
#> 
#> $MB$button_counters$button_clk_clip
#> [1] 0
#> 
#> $MB$button_counters$button_clk_del
#> [1] 0
#> 
#> $MB$button_counters$button_clk_copy
#> [1] 0
#> 
#> $MB$button_counters$button_clk_new
#> [1] 0
#> 
#> 
#> $MB$ui_hold
#> $MB$ui_hold$button_clk_save
#> [1] FALSE
#> 
#> $MB$ui_hold$button_clk_clip
#> [1] FALSE
#> 
#> $MB$ui_hold$button_clk_del
#> [1] FALSE
#> 
#> $MB$ui_hold$button_clk_copy
#> [1] FALSE
#> 
#> $MB$ui_hold$button_clk_new
#> [1] FALSE
#> 
#> $MB$ui_hold$model_type_selection
#> [1] FALSE
#> 
#> $MB$ui_hold$ui_select_element
#> [1] FALSE
#> 
#> $MB$ui_hold$ui_mb_model
#> [1] FALSE
#> 
#> $MB$ui_hold$time_scale
#> [1] FALSE
#> 
#> $MB$ui_hold$uploaded_model
#> [1] FALSE
#> 
#> $MB$ui_hold$base_from
#> [1] FALSE
#> 
#> $MB$ui_hold$element_selection
#> [1] FALSE
#> 
#> $MB$ui_hold$catalog_selection
#> [1] FALSE
#> 
#> $MB$ui_hold$element_name
#> [1] FALSE
#> 
#> 
#> $MB$ui_ids
#>  [1] "button_clk_save"      "button_clk_clip"      "button_clk_del"      
#>  [4] "button_clk_copy"      "button_clk_new"       "model_type_selection"
#>  [7] "ui_select_element"    "ui_mb_model"          "time_scale"          
#> [10] "model_type_selection" "uploaded_model"       "base_from"           
#> [13] "element_selection"    "catalog_selection"    "element_name"        
#> 
#> $MB$isgood
#> [1] TRUE
#> 
#> $MB$ui_ele
#> [1] "catalog_selection" "element_name"      "time_scale"       
#> 
#> $MB$element_cntr
#> [1] 1
#> 
#> $MB$model_type_selection
#> [1] "rxode2"
#> 
#> $MB$base_from
#> [1] "catalog"
#> 
#> $MB$suggested
#> $MB$suggested$found
#> [1] FALSE
#> 
#> $MB$suggested$pkgs
#> $MB$suggested$pkgs$rxode2
#> $MB$suggested$pkgs$rxode2$found
#> [1] FALSE
#> 
#> $MB$suggested$pkgs$rxode2$msg
#> [1] "rxode2 package was not found."
#> 
#> 
#> $MB$suggested$pkgs$nonmem2rx
#> $MB$suggested$pkgs$nonmem2rx$found
#> [1] FALSE
#> 
#> $MB$suggested$pkgs$nonmem2rx$msg
#> [1] "nonmem2rx package was not found."
#> 
#> 
#> $MB$suggested$pkgs$nlmixr2lib
#> $MB$suggested$pkgs$nlmixr2lib$found
#> [1] FALSE
#> 
#> $MB$suggested$pkgs$nlmixr2lib$msg
#> [1] "nlmixr2lib package was not found."
#> 
#> 
#> 
#> 
#> $MB$model_catalog
#> $MB$model_catalog$sources
#> $MB$model_catalog$sources[[1]]
#> $MB$model_catalog$sources[[1]]$source
#> $MB$model_catalog$sources[[1]]$source$group
#> [1] "User-defined Models"
#> 
#> $MB$model_catalog$sources[[1]]$source$name
#> [1] "rxode2 User model"
#> 
#> $MB$model_catalog$sources[[1]]$source$description
#> [1] "User-defined rxode2 model"
#> 
#> $MB$model_catalog$sources[[1]]$source$type
#> [1] "rxode2"
#> 
#> $MB$model_catalog$sources[[1]]$source$obj
#> [1] "my_fcn"
#> 
#> $MB$model_catalog$sources[[1]]$source$file
#> [1] "file.path(getwd(),\"user_model.R\")"
#> 
#> 
#> 
#> $MB$model_catalog$sources[[2]]
#> $MB$model_catalog$sources[[2]]$source
#> $MB$model_catalog$sources[[2]]$source$group
#> [1] "User-defined Models"
#> 
#> $MB$model_catalog$sources[[2]]$source$name
#> [1] "NONMEM User model"
#> 
#> $MB$model_catalog$sources[[2]]$source$description
#> [1] "User-defined NONMEM model"
#> 
#> $MB$model_catalog$sources[[2]]$source$type
#> [1] "NONMEM"
#> 
#> $MB$model_catalog$sources[[2]]$source$file
#> [1] "file.path(getwd(),\"user_model.ctl\")"
#> 
#> 
#> 
#> $MB$model_catalog$sources[[3]]
#> $MB$model_catalog$sources[[3]]$source
#> $MB$model_catalog$sources[[3]]$source$group
#> [1] "Model Library"
#> 
#> $MB$model_catalog$sources[[3]]$source$type
#> [1] "nlmixr2lib"
#> 
#> $MB$model_catalog$sources[[3]]$source$name
#> [1] "nlmixr2 Model Library"
#> 
#> 
#> 
#> 
#> $MB$model_catalog$summary
#> NULL
#> 
#> $MB$model_catalog$select_group
#> list()
#> 
#> $MB$model_catalog$select_plain
#> list()
#> 
#> $MB$model_catalog$select_subtext
#> NULL
#> 
#> $MB$model_catalog$msgs
#> NULL
#> 
#> $MB$model_catalog$hasmdl
#> [1] TRUE
#> 
#> $MB$model_catalog$isgood
#> [1] TRUE
#> 
#> 
#> $MB$elements
#> $MB$elements$element_1
#> $MB$elements$element_1$isgood
#> [1] TRUE
#> 
#> $MB$elements$element_1$ui
#> $MB$elements$element_1$ui$ui_mb_model
#> [1] ""
#> 
#> $MB$elements$element_1$ui$time_scale
#> [1] "weeks"
#> 
#> $MB$elements$element_1$ui$element_name
#> [1] "Model 1"
#> 
#> $MB$elements$element_1$ui$catalog_selection
#> NULL
#> 
#> 
#> $MB$elements$element_1$id
#> [1] "element_1"
#> 
#> $MB$elements$element_1$idx
#> [1] 1
#> 
#> $MB$elements$element_1$fcn_obj_name
#> [1] "MB_obj_1_fcn"
#> 
#> $MB$elements$element_1$rx_obj_name
#> [1] "MB_obj_1_rx"
#> 
#> $MB$elements$element_1$msgs
#> NULL
#> 
#> $MB$elements$element_1$code_previous
#> [1] ""
#> 
#> $MB$elements$element_1$update_model_code
#> [1] FALSE
#> 
#> $MB$elements$element_1$components_table
#> data frame with 0 columns and 0 rows
#> 
#> $MB$elements$element_1$selected_component_id
#> NULL
#> 
#> $MB$elements$element_1$components_list
#> list()
#> 
#> $MB$elements$element_1$checksum
#> [1] "b2e8fd34d20c49c229c7bf3b796157d2"
#> 
#> $MB$elements$element_1$base_model_name
#> [1] ""
#> 
#> $MB$elements$element_1$base_model
#> [1] ""
#> 
#> 
#> 
#> $MB$current_element
#> [1] "element_1"
#> 
#> $MB$checksum
#> [1] "8ac755447dc73c185137d18702a2cdb6"
#> 
#> 
#> $MOD_TYPE
#> [1] "MB"
#> 
#> $id
#> [1] "MB"
#> 
#> $FM_yaml_file
#> [1] "/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/formods/templates/formods.yaml"
#> 
#> $MOD_yaml_file
#> [1] "/private/var/folders/l9/6pm7xq5j1hn6l00zrs22j6mc0000gn/T/Rtmp3mXuRf/temp_libpathd3297b8d3bb9/ruminate/templates/MB.yaml"
#> 
#> $shiny_token
#> [1] "non_shiny"
#>