Append Report Elements

Description

Usage

NCA_append_report(state, rpt, rpttype, gen_code_only = FALSE)

Arguments

state

NCA state from NCA_fetch_state()

rpt

Report with the current content of the report which will be appended to in this function. For details on the structure see the documentation for FM_generate_report.

rpttype

Type of report to generate (supported "xlsx", "pptx", "docx").

gen_code_only

Boolean value indicating that only code should be generated (FALSE).

Value

list containing the following elements

isgood: Return status of the function.

hasrptele: Boolean indicator if the module has any reportable elements.

code: Data wrangling R command.

msgs: Messages to be passed back to the user.

rpt: Report with any additions passed back to the user.

See also

Examples

# We need a state object to use below
sess_res = NCA_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
#> → NCA: State initialized
#> → NCA: State initialized
#> → NCA: loading analysis scenario: sd_iv
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> → NCA: module checksum updated:4b6a10335f14b25490b6b17c88ea872e
#> → NCA: loading analysis scenario: sd_iv
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 2 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 2 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Values from `PPORRES` are not uniquely identified; output will contain list-cols.
#> * Use `values_fn = list` to suppress this warning.
#> * Use `values_fn = {summary_fun}` to summarise duplicates.
#> * Use the following dplyr code to identify duplicates.
#>   {data} %>%
#>     dplyr::group_by(ID, pnames) %>%
#>     dplyr::summarise(n = dplyr::n(), .groups = "drop") %>%
#>     dplyr::filter(n > 1L)
#> → NCA: module checksum updated:60a317526b965f8ef3159a2c49d5e025
state = sess_res$state

rpt = list(summary = list(), sheets=list())

rpt_res = NCA_append_report(state,
 rpt     = rpt,
 rpttype = "xlsx")

# Shows if report elements are present
rpt_res$hasrptele
#> [1] TRUE

# Code chunk to generate report element
cat(paste(rpt_res$code, collapse="\n"))
#> rpt[["sheets"]][["NCA_1_RES"]]=NCA_1_table_ind_params[["one_table"]]
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_1_RES",
#>     Description="PK NCA Analysis Example (indiviudal results)"
#>   )
#> )
#> rpt[["sheets"]][["NCA_1_RAW"]]=as.data.frame(NCA_1_res)
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_1_RAW",
#>     Description="PK NCA Analysis Example (raw NCA results)"
#>   )
#> )
#> rpt[["sheets"]][["NCA_2_RES"]]=NCA_2_table_ind_params[["one_table"]]
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_2_RES",
#>     Description="PK/PD NCA Analysis Example (indiviudal results)"
#>   )
#> )
#> rpt[["sheets"]][["NCA_2_RAW"]]=as.data.frame(NCA_2_res)
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_2_RAW",
#>     Description="PK/PD NCA Analysis Example (raw NCA results)"
#>   )
#> )

# Tabular summary of data views
rpt_res$rpt$summary
#>   Sheet_Name                                     Description
#> 1  NCA_1_RES    PK NCA Analysis Example (indiviudal results)
#> 2  NCA_1_RAW       PK NCA Analysis Example (raw NCA results)
#> 3  NCA_2_RES PK/PD NCA Analysis Example (indiviudal results)
#> 4  NCA_2_RAW    PK/PD NCA Analysis Example (raw NCA results)
# We need a state object to use below
sess_res = NCA_test_mksession(session=list())
#> → UD: State initialized
#> → UD: module checksum updated:48559d397570afc9fba8ad9b252fa75a
#> → DW: State initialized
#> → DW: module checksum updated:5a6adb8be080914162e848e04fb61f85
#> → DW: module checksum updated:a5d1d3b8a436befcfa119b5d72d3697b
#> → DW: module checksum updated:48d9e35c3f568c7a6f72914538d4ecd4
#> → DW: module checksum updated:680bde3fe602684945b93dda69abc020
#> → DW: module checksum updated:1f571c6eabb372eff6c1ef806344b2e4
#> → DW: module checksum updated:a1632ef431da47f98334a9281270a3fd
#> → DW: module checksum updated:518e56af743d086ea4b1b6070bf70472
#> → DW: module checksum updated:460fd6e3a4a46f56f17bde8ff4ef5c01
#> → DW: module checksum updated:12335ca5ced95462c2ed1df919a72d15
#> → DW: module checksum updated:0794933dac491313b32ca4bbfd996641
#> → DW: module checksum updated:75abe1a1cfc787fc68049b40c9e5749e
#> → DW: module checksum updated:070868f0125fd16b3e648dc05be84cd0
#> → DW: module checksum updated:1da012ee43c59bf8bedc7376511d179e
#> → DW: module checksum updated:6dfd06414929a762cc4f07d6d428289b
#> → DW: module checksum updated:83f7106b33c98e7650432749c16d7a36
#> → DW: module checksum updated:32652abd1996af932cbdf304cc509bd0
#> → DW: module checksum updated:dfa171056875bc474e03e0c911ace2b6
#> → DW: module checksum updated:5d4dfb7affd116c593a856034804aa04
#> → DW: module checksum updated:f47016aa2630055e8a1ae44ab79cabd2
#> → DW: module checksum updated:46cc2c9503359a39be98c3c0b889851a
#> → DW: module checksum updated:58ed3ad9e797e77eba05a2576b78e50f
#> → DW: module checksum updated:aaf6ebcd2be24f05cd9b4776ef042986
#> → DW: module checksum updated:afbbf99a8ed5f01191217b8370603a9d
#> → DW: module checksum updated:086f760681a2b47a9a1dfd572f876fcd
#> → DW: module checksum updated:8c045cb5e0fe9bd74e84d87d3c6725b5
#> → DW: module checksum updated:779a7d3c81e58ea4c8a756c2e422fee0
#> → DW: module checksum updated:104438a337e6617c4504477506e90863
#> → DW: module checksum updated:1bd57823c76715ccceab0be02ed7a79e
#> → DW: module checksum updated:8b901fc11d897dd27c774accf0eb7097
#> → DW: module checksum updated:7d55b640df1207dcdc56a10c9b8d8b87
#> → DW: module checksum updated:8d91f192ea994a133c8e2fbb7ce7c11a
#> → DW: module checksum updated:d5c81022154ffb7347d7a30d66a38f60
#> → NCA: State initialized
#> → NCA: State initialized
#> → NCA: loading analysis scenario: sd_iv
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> → NCA: module checksum updated:f2e415a63bcbb9e158e950ed09fdd4ee
#> → NCA: loading analysis scenario: sd_iv
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 1 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 2 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 2 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Too few points for half-life calculation (min.hl.points=3 with only 0 points)
#> Warning: Values from `PPORRES` are not uniquely identified; output will contain list-cols.
#> * Use `values_fn = list` to suppress this warning.
#> * Use `values_fn = {summary_fun}` to summarise duplicates.
#> * Use the following dplyr code to identify duplicates.
#>   {data} %>%
#>     dplyr::group_by(ID, pnames) %>%
#>     dplyr::summarise(n = dplyr::n(), .groups = "drop") %>%
#>     dplyr::filter(n > 1L)
#> → NCA: module checksum updated:70064a2ffe6ed99805b24c82b444be57
state = sess_res$state

# here we need an empty report object for tabular data
rpt = list(summary = list(), sheets=list())

# Now we append the report indicating we want 
# Excel output:
rpt_res = NCA_append_report(state,
  rpt     = rpt,
  rpttype = "xlsx")

# Shows if report elements are present
rpt_res$hasrptele
#> [1] TRUE

# Code chunk to generate report element
cat(paste(rpt_res$code, collapse="\n"))
#> rpt[["sheets"]][["NCA_1_RES"]]=NCA_1_table_ind_params[["one_table"]]
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_1_RES",
#>     Description="PK NCA Analysis Example (indiviudal results)"
#>   )
#> )
#> rpt[["sheets"]][["NCA_1_RAW"]]=as.data.frame(NCA_1_res)
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_1_RAW",
#>     Description="PK NCA Analysis Example (raw NCA results)"
#>   )
#> )
#> rpt[["sheets"]][["NCA_2_RES"]]=NCA_2_table_ind_params[["one_table"]]
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_2_RES",
#>     Description="PK/PD NCA Analysis Example (indiviudal results)"
#>   )
#> )
#> rpt[["sheets"]][["NCA_2_RAW"]]=as.data.frame(NCA_2_res)
#> rpt[["summary"]] = rbind(rpt[["summary"]],
#>   data.frame(
#>     Sheet_Name="NCA_2_RAW",
#>     Description="PK/PD NCA Analysis Example (raw NCA results)"
#>   )
#> )

# Tabular results from the first NCA analysis
rpt_res$rpt$sheets$NCA_1_RES
#>    ID auclast:[0,21] cmax:[0,21] aucinf.obs:[0,inf]
#> 1  ID      AUC~last~      C~max~       AUC~inf,obs~
#> 2          day⋅ng/mL       ng/mL          day⋅ng/mL
#> 3             [0,21]      [0,21]            [0,inf]
#> 4   1           6140         816               9840
#> 5   2           8810        1230              17600
#> 6   3           6410         786               9800
#> 7   4           6390         824              11300
#> 8   5           8090        1150              16400
#> 9   6           6740         737              12900
#> 10  7           6780         679              13500
#> 11  8           7010         586              16600
#> 12  9           9370        1440              16800
#> 13 10           9100         951              16400
#> 14 11           6780         640              14700
#> 15 12           5360         675               8900
#> 16 13           8620        1060              16000
#> 17 14           5930         717              11300
#> 18 15           8760        1060              16800
#> 19 16           5760         643              10300
#> 20 17           9300         985              19800
#> 21 18           5800         574               9980
#> 22 19           8360        1060              16700
#> 23 20           6590         706              12600