Fetches PKNCA Metadata

Compiles Metadata from PKNCA

Usage

NCA_fetch_PKNCA_meta()

Value

Dataframe containing PKCNA metadata for NCA parameters.

Examples

PKNCA_meta = NCA_fetch_PKNCA_meta()
utils::head(PKNCA_meta)
#> $parameters
#>                parameter       unit_type
#> 1                  start            time
#> 2                    end            time
#> 3                auclast             auc
#> 4                 aucall             auc
#> 5               aumclast            aumc
#> 6                aumcall            aumc
#> 7            aucint.last             auc
#> 8       aucint.last.dose             auc
#> 9             aucint.all             auc
#> 10       aucint.all.dose             auc
#> 11                    c0            conc
#> 12                  cmax            conc
#> 13                  cmin            conc
#> 14                  tmax            time
#> 15                 tlast            time
#> 16                tfirst            time
#> 17             clast.obs            conc
#> 18               cl.last       clearance
#> 19                cl.all       clearance
#> 20                     f        fraction
#> 21              mrt.last            time
#> 22           mrt.iv.last            time
#> 23              vss.last          volume
#> 24           vss.iv.last          volume
#> 25                   cav            conc
#> 26          cav.int.last            conc
#> 27           cav.int.all            conc
#> 28               ctrough            conc
#> 29                cstart            conc
#> 30                   ptr        fraction
#> 31                  tlag            time
#> 32              deg.fluc               %
#> 33                 swing               %
#> 34                  ceoi            conc
#> 35  aucabove.predose.all             auc
#> 36   aucabove.trough.all             auc
#> 37            count_conc           count
#> 38   count_conc_measured           count
#> 39               totdose            dose
#> 40                    ae          amount
#> 41              clr.last renal_clearance
#> 42               clr.obs renal_clearance
#> 43              clr.pred renal_clearance
#> 44                    fe     amount_dose
#> 45        sparse_auclast             auc
#> 46         sparse_auc_se             auc
#> 47         sparse_auc_df           count
#> 48            time_above            time
#> 49             aucivlast             auc
#> 50              aucivall             auc
#> 51         aucivint.last             auc
#> 52          aucivint.all             auc
#> 53        aucivpbextlast               %
#> 54         aucivpbextall               %
#> 55    aucivpbextint.last               %
#> 56     aucivpbextint.all               %
#> 57             half.life            time
#> 58             r.squared        unitless
#> 59         adj.r.squared        unitless
#> 60              lambda.z    inverse_time
#> 61   lambda.z.time.first            time
#> 62    lambda.z.time.last            time
#> 63     lambda.z.n.points           count
#> 64            clast.pred            conc
#> 65            span.ratio        fraction
#> 66        thalf.eff.last            time
#> 67     thalf.eff.iv.last            time
#> 68              kel.last    inverse_time
#> 69           kel.iv.last    inverse_time
#> 70            aucinf.obs             auc
#> 71           aucinf.pred             auc
#> 72           aumcinf.obs            aumc
#> 73          aumcinf.pred            aumc
#> 74        aucint.inf.obs             auc
#> 75   aucint.inf.obs.dose             auc
#> 76       aucint.inf.pred             auc
#> 77  aucint.inf.pred.dose             auc
#> 78          aucivinf.obs             auc
#> 79         aucivinf.pred             auc
#> 80     aucivpbextinf.obs               %
#> 81    aucivpbextinf.pred               %
#> 82           aucpext.obs               %
#> 83          aucpext.pred               %
#> 84                cl.obs       clearance
#> 85               cl.pred       clearance
#> 86               mrt.obs            time
#> 87              mrt.pred            time
#> 88            mrt.iv.obs            time
#> 89           mrt.iv.pred            time
#> 90            mrt.md.obs            time
#> 91           mrt.md.pred            time
#> 92                vz.obs          volume
#> 93               vz.pred          volume
#> 94               vss.obs          volume
#> 95              vss.pred          volume
#> 96            vss.iv.obs          volume
#> 97           vss.iv.pred          volume
#> 98            vss.md.obs          volume
#> 99           vss.md.pred          volume
#> 100      cav.int.inf.obs            conc
#> 101     cav.int.inf.pred            conc
#> 102        thalf.eff.obs            time
#> 103       thalf.eff.pred            time
#> 104     thalf.eff.iv.obs            time
#> 105    thalf.eff.iv.pred            time
#> 106              kel.obs    inverse_time
#> 107             kel.pred    inverse_time
#> 108           kel.iv.obs    inverse_time
#> 109          kel.iv.pred    inverse_time
#> 110           auclast.dn    auc_dosenorm
#> 111            aucall.dn    auc_dosenorm
#> 112        aucinf.obs.dn    auc_dosenorm
#> 113       aucinf.pred.dn    auc_dosenorm
#> 114          aumclast.dn   aumc_dosenorm
#> 115           aumcall.dn   aumc_dosenorm
#> 116       aumcinf.obs.dn   aumc_dosenorm
#> 117      aumcinf.pred.dn   aumc_dosenorm
#> 118              cmax.dn   conc_dosenorm
#> 119              cmin.dn   conc_dosenorm
#> 120         clast.obs.dn   conc_dosenorm
#> 121        clast.pred.dn   conc_dosenorm
#> 122               cav.dn   conc_dosenorm
#> 123           ctrough.dn   conc_dosenorm
#>                                                 pretty_name data_type
#> 1                                            Interval Start  interval
#> 2                                              Interval End  interval
#> 3                                                   AUClast  interval
#> 4                                                    AUCall  interval
#> 5                                                 AUMC,last  interval
#> 6                                                  AUMC,all  interval
#> 7                   AUCint (based on AUClast extrapolation)  interval
#> 8       AUCint (based on AUClast extrapolation, dose-aware)  interval
#> 9                    AUCint (based on AUCall extrapolation)  interval
#> 10       AUCint (based on AUCall extrapolation, dose-aware)  interval
#> 11                                                       C0  interval
#> 12                                                     Cmax  interval
#> 13                                                     Cmin  interval
#> 14                                                     Tmax  interval
#> 15                                                    Tlast  interval
#> 16                                                   Tfirst  interval
#> 17                                                    Clast  interval
#> 18                                    CL (based on AUClast)  interval
#> 19                                     CL (based on AUCall)  interval
#> 20                                          Bioavailability  interval
#> 21                                   MRT (based on AUClast)  interval
#> 22                    MRT (for IV dosing, based on AUClast)  interval
#> 23                                   Vss (based on AUClast)  interval
#> 24                    Vss (for IV dosing, based on AUClast)  interval
#> 25                                                      Cav  interval
#> 26                                                      Cav  interval
#> 27                                                      Cav  interval
#> 28                                                  Ctrough  interval
#> 29                                                   Cstart  interval
#> 30                                     Peak-to-trough ratio  interval
#> 31                                                     Tlag  interval
#> 32                                    Degree of fluctuation  interval
#> 33                                                    Swing  interval
#> 34                                                     Ceoi  interval
#> 35                                                AUC,above  interval
#> 36                                                AUC,above  interval
#> 37                                      Concentration count  interval
#> 38                             Measured concentration count  interval
#> 39                                               Total dose  interval
#> 40                                          Amount excreted  interval
#> 41                           Renal clearance (from AUClast)  interval
#> 42                        Renal clearance (from AUCinf,obs)  interval
#> 43                       Renal clearance (from AUCinf,pred)  interval
#> 44                                        Fraction excreted  interval
#> 45                                           Sparse AUClast  interval
#> 46                            Sparse AUClast standard error  interval
#> 47                        Sparse AUClast degrees of freedom  interval
#> 48                                 Time above Concentration  interval
#> 49                                      AUClast (IV dosing)  interval
#> 50                                       AUCall (IV dosing)  interval
#> 51                                  AUCint,last (IV dosing)  interval
#> 52                                   AUCint,all (IV dosing)  interval
#> 53                               AUCbext (based on AUClast)  interval
#> 54                                AUCbext (based on AUCall)  interval
#> 55                           AUCbext (based on AUCint,last)  interval
#> 56                            AUCbext (based on AUCint,all)  interval
#> 57                                                Half-life  interval
#> 58                                                    $r^2$  interval
#> 59                                              $r^2_{adj}$  interval
#> 60                                             $\\lambda_z$  interval
#> 61                              First time for $\\lambda_z$  interval
#> 62                               Last time for $\\lambda_z$  interval
#> 63                       Number of points used for lambda_z  interval
#> 64                                               Clast,pred  interval
#> 65                                               Span ratio  interval
#> 66                  Effective half-life (based on MRT,last)  interval
#> 67    Effective half-life (for IV dosing, based on MRTlast)  interval
#> 68                                   Kel (based on AUClast)  interval
#> 69                    Kel (for IV dosing, based on AUClast)  interval
#> 70                                               AUCinf,obs  interval
#> 71                                              AUCinf,pred  interval
#> 72                                             AUMC,inf,obs  interval
#> 73                                            AUMC,inf,pred  interval
#> 74               AUCint (based on AUCinf,obs extrapolation)  interval
#> 75   AUCint (based on AUCinf,obs extrapolation, dose-aware)  interval
#> 76              AUCint (based on AUCinf,pred extrapolation)  interval
#> 77  AUCint (based on AUCinf,pred extrapolation, dose-aware)  interval
#> 78                                   AUCinf,obs (IV dosing)  interval
#> 79                                  AUCinf,pred (IV dosing)  interval
#> 80                            AUCbext (based on AUCinf,obs)  interval
#> 81                           AUCbext (based on AUCinf,pred)  interval
#> 82                            AUCpext (based on AUCinf,obs)  interval
#> 83                           AUCpext (based on AUCinf,pred)  interval
#> 84                                 CL (based on AUCinf,obs)  interval
#> 85                                CL (based on AUCinf,pred)  interval
#> 86                                MRT (based on AUCinf,obs)  interval
#> 87                               MRT (based on AUCinf,pred)  interval
#> 88                 MRT (for IV dosing, based on AUCinf,obs)  interval
#> 89                MRT (for IV dosing, based on AUCinf,pred)  interval
#> 90           MRT (for multiple dosing, based on AUCinf,obs)  interval
#> 91          MRT (for multiple dosing, based on AUCinf,pred)  interval
#> 92                                 Vz (based on AUCinf,obs)  interval
#> 93                                Vz (based on AUCinf,pred)  interval
#> 94                                Vss (based on AUCinf,obs)  interval
#> 95                               Vss (based on AUCinf,pred)  interval
#> 96                 Vss (for IV dosing, based on AUCinf,obs)  interval
#> 97                Vss (for IV dosing, based on AUCinf,pred)  interval
#> 98              Vss (for multiple-dose, based on Clast,obs)  interval
#> 99             Vss (for multiple-dose, based on Clast,pred)  interval
#> 100                                                     Cav  interval
#> 101                                                     Cav  interval
#> 102                  Effective half-life (based on MRT,obs)  interval
#> 103                 Effective half-life (based on MRT,pred)  interval
#> 104   Effective half-life (for IV dosing, based on MRT,obs)  interval
#> 105  Effective half-life (for IV dosing, based on MRT,pred)  interval
#> 106                               Kel (based on AUCinf,obs)  interval
#> 107                              Kel (based on AUCinf,pred)  interval
#> 108                Kel (for IV dosing, based on AUCinf,obs)  interval
#> 109               Kel (for IV dosing, based on AUCinf,pred)  interval
#> 110                               AUClast (dose-normalized)  interval
#> 111                                AUCall (dose-normalized)  interval
#> 112                            AUCinf,obs (dose-normalized)  interval
#> 113                           AUCinf,pred (dose-normalized)  interval
#> 114                             AUMC,last (dose-normalized)  interval
#> 115                              AUMC,all (dose-normalized)  interval
#> 116                          AUMC,inf,obs (dose-normalized)  interval
#> 117                         AUMC,inf,pred (dose-normalized)  interval
#> 118                                  Cmax (dose-normalized)  interval
#> 119                                  Cmin (dose-normalized)  interval
#> 120                                 Clast (dose-normalized)  interval
#> 121                            Clast,pred (dose-normalized)  interval
#> 122                                   Cav (dose-normalized)  interval
#> 123                               Ctrough (dose-normalized)  interval
#>                                                                                                                                                                                                                                                                                                                                                              desc
#> 1                                                                                                                                                                                                                                                                                                                                   Starting time of the interval
#> 2                                                                                                                                                                                                                                                                                                              Ending time of the interval (potentially infinity)
#> 3                                                                                                                                                                                                                      The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 4                                                                                                     The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 5                                                                                                                                                                                                               The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification
#> 6                                                                                The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification plus the moment of the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 7                                                                                                                                                                                                                                  The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 8                                                                                                                                                                    The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 9                                                                                                                                                                           The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 10                                                                                                            The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 11                                                                                                                                                                                                                                                                                                                        Initial concentration after an IV bolus
#> 12                                                                                                                                                                                                                                                                                                                                 Maximum observed concentration
#> 13                                                                                                                                                                                                                                                                                                                                 Minimum observed concentration
#> 14                                                                                                                                                                                                                                                                                                                     Time of the maximum observed concentration
#> 15                                                                                                                                                                                                                                                                                      Time of the last concentration observed above the limit of quantification
#> 16                                                                                                                                                                                                                                                                                              Time of the first concentration above the limit of quantification
#> 17                                                                                                                                                                                                                                                                                              The last concentration observed above the limit of quantification
#> 18                                                                                                                                                                                                                                                                                                       Clearance or observed oral clearance calculated to Clast
#> 19                                                                                                                                                                                                                                                                                                    Clearance or observed oral clearance calculated with AUCall
#> 20                                                                                                                                                                                                                                                                                                                    Bioavailability or relative bioavailability
#> 21                                                                                                                                                                                                                                                                                       The mean residence time to the last observed concentration above the LOQ
#> 22                                                                                                                                                                                                                                                        The mean residence time to the last observed concentration above the LOQ correcting for dosing duration
#> 23                                                                                                                                                                                                                                                                                              The steady-state volume of distribution calculating through Tlast
#> 24                                                                                                                                                                                                                                                                    The steady-state volume of distribution with intravenous infusion calculating through Tlast
#> 25                                                                                                                                                                                                                                                                                         The average concentration during an interval (calculated with AUClast)
#> 26                                                                                                                                                                                                                                                                                     The average concentration during an interval (calculated with AUCint.last)
#> 27                                                                                                                                                                                                                                                                                      The average concentration during an interval (calculated with AUCint.all)
#> 28                                                                                                                                                                                                                                                                                                                     The trough (end of interval) concentration
#> 29                                                                                                                                                                                                                                                                                                                                      The predose concentration
#> 30                                                                                                                                                                                                                                                                                                                                Peak-to-Trough ratio (fraction)
#> 31                                                                                                                                                                                                                                                                                                                                                       Lag time
#> 32                                                                                                                                                                                                                                                                                                                                          Degree of fluctuation
#> 33                                                                                                                                                                                                                                                                                                                                         Swing relative to Cmin
#> 34                                                                                                                                                                                                                                                                                                                           Concentration at the end of infusion
#> 35  The area under the concentration time the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification, with a concentration subtracted from all concentrations and values below zero after subtraction set to zero
#> 36  The area under the concentration time the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification, with a concentration subtracted from all concentrations and values below zero after subtraction set to zero
#> 37                                                                                                                                                                                                                                                                                                           Number of non-missing concentrations for an interval
#> 38                                                                                                                                                                                                                                                                                              Number of measured and non BLQ/ALQ concentrations for an interval
#> 39                                                                                                                                                                                                                                                                                                                     Total dose administered during an interval
#> 40                                                                                                                                                                                                                                                                                                            The amount excreted (typically into urine or feces)
#> 41                                                                                                                                                                                                                                                                                                                   The renal clearance calculated using AUClast
#> 42                                                                                                                                                                                                                                                                                                                The renal clearance calculated using AUCinf,obs
#> 43                                                                                                                                                                                                                                                                                                               The renal clearance calculated using AUCinf,pred
#> 44                                                                                                                                                                                                                                                                                                                              The fraction of the dose excreted
#> 45                                                                                                                                                                                             For sparse PK sampling, the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 46                                                                                                                                                                       For sparse PK sampling, the standard error of the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 47                                                                                                                                                    For sparse PK sampling, the standard error degrees of freedom of the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 48                                                                                                                                                                                                                                                                                                                               Time above a given concentration
#> 49                                                                                                                                                                                                                                                                    The AUClast calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 50                                                                                                                                                                                                                                                                     The AUCall calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 51                                                                                                                                                                                                                                                                The AUCint,last calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 52                                                                                                                                                                                                                                                                 The AUCint,all calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 53                                                                                                                                                                                                                                                                                         The back-extrapolation percent for intravenous dosing based on AUClast
#> 54                                                                                                                                                                                                                                                                                          The back-extrapolation percent for intravenous dosing based on AUCall
#> 55                                                                                                                                                                                                                                                                                     The back-extrapolation percent for intravenous dosing based on AUCint,last
#> 56                                                                                                                                                                                                                                                                                      The back-extrapolation percent for intravenous dosing based on AUCint,all
#> 57                                                                                                                                                                                                                                                                                                                                       The (terminal) half-life
#> 58                                                                                                                                                                                                                                                                                                                     The r^2 value of the half-life calculation
#> 59                                                                                                                                                                                                                                                                                                            The adjusted r^2 value of the half-life calculation
#> 60                                                                                                                                                                                                                                                                                                                 The elimination rate of the terminal half-life
#> 61                                                                                                                                                                                                                                                                                                     The first time point used for the calculation of half-life
#> 62                                                                                                                                                                                                                                                                                                      The last time point used for the calculation of half-life
#> 63                                                                                                                                                                                                                                                                                                     The number of points used for the calculation of half-life
#> 64                                                                                                                                                                                                                                                                                                       The concentration at Tlast as predicted by the half-life
#> 65                                                                                                                                                                                                                                                                                      The ratio of the half-life to the duration used for half-life calculation
#> 66                                                                                                                                                                                                                                                                                                       The effective half-life (as determined from the MRTlast)
#> 67                                                                                                                                                                                                                                                                                           The effective half-life (as determined from the intravenous MRTlast)
#> 68                                                                                                                                                                                                                                                                                                    Elimination rate (as calculated from the MRT using AUClast)
#> 69                                                                                                                                                                                                                                                                                                  Elimination rate (as calculated from the intravenous MRTlast)
#> 70                                                                                                                                                                                                              The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 71                                                                                                                                                                                                             The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast
#> 72                                                                                                                                                                                                       The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 73                                                                                                                                                                                                      The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast
#> 74                                                                                                                                                                                                                                 The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 75                                                                                                                                                                   The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 76                                                                                                                                                                          The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 77                                                                                                            The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 78                                                                                                                                                                                                                                                                 The AUCinf,obs calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 79                                                                                                                                                                                                                                                                           The  calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 80                                                                                                                                                                                                                                                                                      The back-extrapolation percent for intravenous dosing based on AUCinf,obs
#> 81                                                                                                                                                                                                                                                                                     The back-extrapolation percent for intravenous dosing based on AUCinf,pred
#> 82                                                                                                                                                                                                                                                                      Percent of the AUCinf that is extrapolated after Tlast calculated from the observed Clast
#> 83                                                                                                                                                                                                                                                                     Percent of the AUCinf that is extrapolated after Tlast calculated from the predicted Clast
#> 84                                                                                                                                                                                                                                                                                            Clearance or observed oral clearance calculated with observed Clast
#> 85                                                                                                                                                                                                                                                                                           Clearance or observed oral clearance calculated with predicted Clast
#> 86                                                                                                                                                                                                                                                                                                       The mean residence time to infinity using observed Clast
#> 87                                                                                                                                                                                                                                                                                                      The mean residence time to infinity using predicted Clast
#> 88                                                                                                                                                                                                                                                                        The mean residence time to infinity using observed Clast correcting for dosing duration
#> 89                                                                                                                                                                                                                                                                       The mean residence time to infinity using predicted Clast correcting for dosing duration
#> 90                                                                                                                                                                                                                                                                       The mean residence time with multiple dosing and nonlinear kinetics using observed Clast
#> 91                                                                                                                                                                                                                                                                      The mean residence time with multiple dosing and nonlinear kinetics using predicted Clast
#> 92                                                                                                                                                                                                                                                                                                       The terminal volume of distribution using observed Clast
#> 93                                                                                                                                                                                                                                                                                                      The terminal volume of distribution using predicted Clast
#> 94                                                                                                                                                                                                                                                                                                   The steady-state volume of distribution using observed Clast
#> 95                                                                                                                                                                                                                                                                                                  The steady-state volume of distribution using predicted Clast
#> 96                                                                                                                                                                                                                                                                         The steady-state volume of distribution with intravenous infusion using observed Clast
#> 97                                                                                                                                                                                                                                                                        The steady-state volume of distribution with intravenous infusion using predicted Clast
#> 98                                                                                                                                                                                                                                                                  The steady-state volume of distribution for nonlinear multiple-dose data using observed Clast
#> 99                                                                                                                                                                                                                                                                 The steady-state volume of distribution for nonlinear multiple-dose data using predicted Clast
#> 100                                                                                                                                                                                                                                                                                 The average concentration during an interval (calculated with AUCint.inf.obs)
#> 101                                                                                                                                                                                                                                                                                The average concentration during an interval (calculated with AUCint.inf.pred)
#> 102                                                                                                                                                                                                                                                                                                       The effective half-life (as determined from the MRTobs)
#> 103                                                                                                                                                                                                                                                                                                      The effective half-life (as determined from the MRTpred)
#> 104                                                                                                                                                                                                                                                                                           The effective half-life (as determined from the intravenous MRTobs)
#> 105                                                                                                                                                                                                                                                                                          The effective half-life (as determined from the intravenous MRTpred)
#> 106                                                                                                                                                                                                                                                                                             Elimination rate (as calculated from the MRT with observed Clast)
#> 107                                                                                                                                                                                                                                                                                            Elimination rate (as calculated from the MRT with predicted Clast)
#> 108                                                                                                                                                                                                                                                                                                  Elimination rate (as calculated from the intravenous MRTobs)
#> 109                                                                                                                                                                                                                                                                                                 Elimination rate (as calculated from the intravenous MRTpred)
#> 110                                                                                                                                                                                                                                                                                                                                       Dose normalized auclast
#> 111                                                                                                                                                                                                                                                                                                                                        Dose normalized aucall
#> 112                                                                                                                                                                                                                                                                                                                                    Dose normalized aucinf.obs
#> 113                                                                                                                                                                                                                                                                                                                                   Dose normalized aucinf.pred
#> 114                                                                                                                                                                                                                                                                                                                                      Dose normalized aumclast
#> 115                                                                                                                                                                                                                                                                                                                                       Dose normalized aumcall
#> 116                                                                                                                                                                                                                                                                                                                                   Dose normalized aumcinf.obs
#> 117                                                                                                                                                                                                                                                                                                                                  Dose normalized aumcinf.pred
#> 118                                                                                                                                                                                                                                                                                                                                          Dose normalized cmax
#> 119                                                                                                                                                                                                                                                                                                                                          Dose normalized cmin
#> 120                                                                                                                                                                                                                                                                                                                                     Dose normalized clast.obs
#> 121                                                                                                                                                                                                                                                                                                                                    Dose normalized clast.pred
#> 122                                                                                                                                                                                                                                                                                                                                           Dose normalized cav
#> 123                                                                                                                                                                                                                                                                                                                                       Dose normalized ctrough
#>