Fetches NCA Parameter Meta Information

This provides meta information about NCA parameters. This includes parameter names, text descriptions, formatting (md and LaTeX).

Usage

NCA_fetch_np_meta(
  MOD_yaml_file = system.file(package = "ruminate", "templates", "NCA.yaml")
)

Arguments

MOD_yaml_file

Module configuration file with MC as main section.

Value

list with the following elements:

choices: List parameter choices grouped by values specified in the module configuration file.

summary: Data frame with meta data about the NCA parameters with the following columns:

parameter: Name of parameter in PKNCA.

text: Name of parameter in plain text.

md: Parameter name formatted in Markdown.

latex: Parameter name formatted using LaTeX.

description: Verbose description in plain text for the parameter.

Examples

NCA_fetch_np_meta()
#> $choices
#> $choices$`Common Parameters`
#> $choices$`Common Parameters`$AUCall
#> [1] "aucall"
#> 
#> $choices$`Common Parameters`$C0
#> [1] "c0"
#> 
#> $choices$`Common Parameters`$`CL (based on AUCall)`
#> [1] "cl.all"
#> 
#> $choices$`Common Parameters`$`CL (based on AUClast)`
#> [1] "cl.last"
#> 
#> $choices$`Common Parameters`$Cmax
#> [1] "cmax"
#> 
#> $choices$`Common Parameters`$`Half-life`
#> [1] "half.life"
#> 
#> $choices$`Common Parameters`$`r²`
#> [1] "r.squared"
#> 
#> $choices$`Common Parameters`$`Sparse AUClast`
#> [1] "sparse_auclast"
#> 
#> 
#> $choices$Other
#> $choices$Other$`r² (Adj)`
#> [1] "adj.r.squared"
#> 
#> $choices$Other$`Amount excreted`
#> [1] "ae"
#> 
#> $choices$Other$`AUCall (dose-normalized)`
#> [1] "aucall.dn"
#> 
#> $choices$Other$`AUCinf,obs`
#> [1] "aucinf.obs"
#> 
#> $choices$Other$`AUCinf,obs (dose-normalized)`
#> [1] "aucinf.obs.dn"
#> 
#> $choices$Other$`AUCinf,pred (dose-normalized)`
#> [1] "aucinf.pred.dn"
#> 
#> $choices$Other$`AUCint (based on AUCall extrapolation)`
#> [1] "aucint.all"
#> 
#> $choices$Other$`AUCint (based on AUCall extrapolation, dose-aware)`
#> [1] "aucint.all.dose"
#> 
#> $choices$Other$`AUCint (based on AUCinf,obs extrapolation)`
#> [1] "aucint.inf.obs"
#> 
#> $choices$Other$`AUCint (based on AUCinf,obs extrapolation, dose-aware)`
#> [1] "aucint.inf.obs.dose"
#> 
#> $choices$Other$`AUCint (based on AUCinf,pred extrapolation)`
#> [1] "aucint.inf.pred"
#> 
#> $choices$Other$`AUCint (based on AUCinf,pred extrapolation, dose-aware)`
#> [1] "aucint.inf.pred.dose"
#> 
#> $choices$Other$`AUCint (based on AUClast extrapolation)`
#> [1] "aucint.last"
#> 
#> $choices$Other$`AUCint (based on AUClast extrapolation, dose-aware)`
#> [1] "aucint.last.dose"
#> 
#> $choices$Other$`AUCall (IV dosing)`
#> [1] "aucivall"
#> 
#> $choices$Other$`AUCinf,obs (IV dosing)`
#> [1] "aucivinf.obs"
#> 
#> $choices$Other$`AUCinf,pred (IV dosing)`
#> [1] "aucivinf.pred"
#> 
#> $choices$Other$`AUCint,all (IV dosing)`
#> [1] "aucivint.all"
#> 
#> $choices$Other$`AUCint,last (IV dosing)`
#> [1] "aucivint.last"
#> 
#> $choices$Other$`AUClast (IV dosing)`
#> [1] "aucivlast"
#> 
#> $choices$Other$`AUCbext (based on AUCall)`
#> [1] "aucivpbextall"
#> 
#> $choices$Other$`AUCbext (based on AUCinf,obs)`
#> [1] "aucivpbextinf.obs"
#> 
#> $choices$Other$`AUCbext (based on AUCinf,pred)`
#> [1] "aucivpbextinf.pred"
#> 
#> $choices$Other$`AUCbext (based on AUCint,all)`
#> [1] "aucivpbextint.all"
#> 
#> $choices$Other$`AUCbext (based on AUCint,last)`
#> [1] "aucivpbextint.last"
#> 
#> $choices$Other$`AUCbext (based on AUClast)`
#> [1] "aucivpbextlast"
#> 
#> $choices$Other$AUClast
#> [1] "auclast"
#> 
#> $choices$Other$`AUClast (dose-normalized)`
#> [1] "auclast.dn"
#> 
#> $choices$Other$`AUCpext (based on AUCinf,obs)`
#> [1] "aucpext.obs"
#> 
#> $choices$Other$`AUCpext (based on AUCinf,pred)`
#> [1] "aucpext.pred"
#> 
#> $choices$Other$`AUMC,all`
#> [1] "aumcall"
#> 
#> $choices$Other$`AUMC,all (dose-normalized)`
#> [1] "aumcall.dn"
#> 
#> $choices$Other$`AUMC,inf,obs`
#> [1] "aumcinf.obs"
#> 
#> $choices$Other$`AUMC,inf,obs (dose-normalized)`
#> [1] "aumcinf.obs.dn"
#> 
#> $choices$Other$`AUMC,inf,pred (dose-normalized)`
#> [1] "aumcinf.pred.dn"
#> 
#> $choices$Other$`AUMC,last`
#> [1] "aumclast"
#> 
#> $choices$Other$`AUMC,last (dose-normalized)`
#> [1] "aumclast.dn"
#> 
#> $choices$Other$Cav
#> [1] "cav"
#> 
#> $choices$Other$`Cav (dose-normalized)`
#> [1] "cav.dn"
#> 
#> $choices$Other$Ceoi
#> [1] "ceoi"
#> 
#> $choices$Other$`CL (based on AUCinf,obs)`
#> [1] "cl.obs"
#> 
#> $choices$Other$`CL (based on AUCinf,pred)`
#> [1] "cl.pred"
#> 
#> $choices$Other$Clast
#> [1] "clast.obs"
#> 
#> $choices$Other$`Clast,pred`
#> [1] "clast.pred"
#> 
#> $choices$Other$`Clast,pred (dose-normalized)`
#> [1] "clast.pred.dn"
#> 
#> $choices$Other$`Renal clearance (from AUClast)`
#> [1] "clr.last"
#> 
#> $choices$Other$`Renal clearance (from AUCinf,obs)`
#> [1] "clr.obs"
#> 
#> $choices$Other$`Renal clearance (from AUCinf,pred)`
#> [1] "clr.pred"
#> 
#> $choices$Other$`Cmax (dose-normalized)`
#> [1] "cmax.dn"
#> 
#> $choices$Other$Cmin
#> [1] "cmin"
#> 
#> $choices$Other$`Cmin (dose-normalized)`
#> [1] "cmin.dn"
#> 
#> $choices$Other$Ctrough
#> [1] "ctrough"
#> 
#> $choices$Other$`Ctrough (dose-normalized)`
#> [1] "ctrough.dn"
#> 
#> $choices$Other$`Degree of fluctuation`
#> [1] "deg.fluc"
#> 
#> $choices$Other$`Interval End`
#> [1] "end"
#> 
#> $choices$Other$Bioavailability
#> [1] "f"
#> 
#> $choices$Other$`Fraction excreted`
#> [1] "fe"
#> 
#> $choices$Other$`Kel (for IV dosing, based on AUClast)`
#> [1] "kel.iv.last"
#> 
#> $choices$Other$`Kel (for IV dosing, based on AUCinf,obs)`
#> [1] "kel.iv.obs"
#> 
#> $choices$Other$`Kel (for IV dosing, based on AUCinf,pred)`
#> [1] "kel.iv.pred"
#> 
#> $choices$Other$`Kel (based on AUClast)`
#> [1] "kel.last"
#> 
#> $choices$Other$`Kel (based on AUCinf,obs)`
#> [1] "kel.obs"
#> 
#> $choices$Other$`Kel (based on AUCinf,pred)`
#> [1] "kel.pred"
#> 
#> $choices$Other$λz
#> [1] "lambda.z"
#> 
#> $choices$Other$`λz,n`
#> [1] "lambda.z.n.points"
#> 
#> $choices$Other$`λz (t 1ˢᵗ)`
#> [1] "lambda.z.time.first"
#> 
#> $choices$Other$`MRT (for IV dosing, based on AUClast)`
#> [1] "mrt.iv.last"
#> 
#> $choices$Other$`MRT (for IV dosing, based on AUCinf,obs)`
#> [1] "mrt.iv.obs"
#> 
#> $choices$Other$`MRT (for IV dosing, based on AUCinf,pred)`
#> [1] "mrt.iv.pred"
#> 
#> $choices$Other$`MRT (based on AUClast)`
#> [1] "mrt.last"
#> 
#> $choices$Other$`MRT (for multiple dosing, based on AUCinf,obs)`
#> [1] "mrt.md.obs"
#> 
#> $choices$Other$`MRT (for multiple dosing, based on AUCinf,pred)`
#> [1] "mrt.md.pred"
#> 
#> $choices$Other$`MRT (based on AUCinf,obs)`
#> [1] "mrt.obs"
#> 
#> $choices$Other$`MRT (based on AUCinf,pred)`
#> [1] "mrt.pred"
#> 
#> $choices$Other$`Peak-to-trough ratio`
#> [1] "ptr"
#> 
#> $choices$Other$`Span ratio`
#> [1] "span.ratio"
#> 
#> $choices$Other$`Interval Start`
#> [1] "start"
#> 
#> $choices$Other$Swing
#> [1] "swing"
#> 
#> $choices$Other$`t 1ˢᵗ`
#> [1] "tfirst"
#> 
#> $choices$Other$`Effective half-life (for IV dosing, based on MRTlast)`
#> [1] "thalf.eff.iv.last"
#> 
#> $choices$Other$`Effective half-life (for IV dosing, based on MRT,obs)`
#> [1] "thalf.eff.iv.obs"
#> 
#> $choices$Other$`Effective half-life (for IV dosing, based on MRT,pred)`
#> [1] "thalf.eff.iv.pred"
#> 
#> $choices$Other$`Effective half-life (based on MRT,last)`
#> [1] "thalf.eff.last"
#> 
#> $choices$Other$`Effective half-life (based on MRT,obs)`
#> [1] "thalf.eff.obs"
#> 
#> $choices$Other$`Effective half-life (based on MRT,pred)`
#> [1] "thalf.eff.pred"
#> 
#> $choices$Other$`Time above Concentration`
#> [1] "time_above"
#> 
#> $choices$Other$Tlag
#> [1] "tlag"
#> 
#> $choices$Other$Tlast
#> [1] "tlast"
#> 
#> $choices$Other$`Vd (based on Clast,obs)`
#> [1] "vd.obs"
#> 
#> $choices$Other$`Vd (based on Clast,pred)`
#> [1] "vd.pred"
#> 
#> $choices$Other$`Vss (for IV dosing, based on AUClast)`
#> [1] "vss.iv.last"
#> 
#> $choices$Other$`Vss (for IV dosing, based on AUCinf,obs)`
#> [1] "vss.iv.obs"
#> 
#> $choices$Other$`Vss (for IV dosing, based on AUCinf,pred)`
#> [1] "vss.iv.pred"
#> 
#> $choices$Other$`Vss (based on AUClast)`
#> [1] "vss.last"
#> 
#> $choices$Other$`Vss (for multiple-dose, based on Clast,obs)`
#> [1] "vss.md.obs"
#> 
#> $choices$Other$`Vss (for multiple-dose, based on Clast,pred)`
#> [1] "vss.md.pred"
#> 
#> $choices$Other$`Vss (based on AUCinf,obs)`
#> [1] "vss.obs"
#> 
#> $choices$Other$`Vss (based on AUCinf,pred)`
#> [1] "vss.pred"
#> 
#> $choices$Other$`Vz (based on AUCinf,obs)`
#> [1] "vz.obs"
#> 
#> $choices$Other$`Vz (based on AUCinf,pred)`
#> [1] "vz.pred"
#> 
#> 
#> 
#> $summary
#>                parameter
#> 1                 aucall
#> 2          adj.r.squared
#> 3                     ae
#> 4              aucall.dn
#> 5             aucinf.obs
#> 6          aucinf.obs.dn
#> 7         aucinf.pred.dn
#> 8             aucint.all
#> 9        aucint.all.dose
#> 10        aucint.inf.obs
#> 11   aucint.inf.obs.dose
#> 12       aucint.inf.pred
#> 13  aucint.inf.pred.dose
#> 14           aucint.last
#> 15      aucint.last.dose
#> 16              aucivall
#> 17          aucivinf.obs
#> 18         aucivinf.pred
#> 19          aucivint.all
#> 20         aucivint.last
#> 21             aucivlast
#> 22         aucivpbextall
#> 23     aucivpbextinf.obs
#> 24    aucivpbextinf.pred
#> 25     aucivpbextint.all
#> 26    aucivpbextint.last
#> 27        aucivpbextlast
#> 28               auclast
#> 29            auclast.dn
#> 30           aucpext.obs
#> 31          aucpext.pred
#> 32               aumcall
#> 33            aumcall.dn
#> 34           aumcinf.obs
#> 35        aumcinf.obs.dn
#> 36       aumcinf.pred.dn
#> 37              aumclast
#> 38           aumclast.dn
#> 39                    c0
#> 40                   cav
#> 41                cav.dn
#> 42                  ceoi
#> 43                cl.all
#> 44               cl.last
#> 45                cl.obs
#> 46               cl.pred
#> 47             clast.obs
#> 48            clast.pred
#> 49         clast.pred.dn
#> 50              clr.last
#> 51               clr.obs
#> 52              clr.pred
#> 53                  cmax
#> 54               cmax.dn
#> 55                  cmin
#> 56               cmin.dn
#> 57               ctrough
#> 58            ctrough.dn
#> 59              deg.fluc
#> 60                   end
#> 61                     f
#> 62                    fe
#> 63             half.life
#> 64           kel.iv.last
#> 65            kel.iv.obs
#> 66           kel.iv.pred
#> 67              kel.last
#> 68               kel.obs
#> 69              kel.pred
#> 70              lambda.z
#> 71     lambda.z.n.points
#> 72   lambda.z.time.first
#> 73           mrt.iv.last
#> 74            mrt.iv.obs
#> 75           mrt.iv.pred
#> 76              mrt.last
#> 77            mrt.md.obs
#> 78           mrt.md.pred
#> 79               mrt.obs
#> 80              mrt.pred
#> 81                   ptr
#> 82             r.squared
#> 83            span.ratio
#> 84        sparse_auclast
#> 85                 start
#> 86                 swing
#> 87                tfirst
#> 88     thalf.eff.iv.last
#> 89      thalf.eff.iv.obs
#> 90     thalf.eff.iv.pred
#> 91        thalf.eff.last
#> 92         thalf.eff.obs
#> 93        thalf.eff.pred
#> 94            time_above
#> 95                  tlag
#> 96                 tlast
#> 97                vd.obs
#> 98               vd.pred
#> 99           vss.iv.last
#> 100           vss.iv.obs
#> 101          vss.iv.pred
#> 102             vss.last
#> 103           vss.md.obs
#> 104          vss.md.pred
#> 105              vss.obs
#> 106             vss.pred
#> 107               vz.obs
#> 108              vz.pred
#>                                                        text
#> 1                                                    AUCall
#> 2                                                  r² (Adj)
#> 3                                           Amount excreted
#> 4                                  AUCall (dose-normalized)
#> 5                                                AUCinf,obs
#> 6                              AUCinf,obs (dose-normalized)
#> 7                             AUCinf,pred (dose-normalized)
#> 8                    AUCint (based on AUCall extrapolation)
#> 9        AUCint (based on AUCall extrapolation, dose-aware)
#> 10               AUCint (based on AUCinf,obs extrapolation)
#> 11   AUCint (based on AUCinf,obs extrapolation, dose-aware)
#> 12              AUCint (based on AUCinf,pred extrapolation)
#> 13  AUCint (based on AUCinf,pred extrapolation, dose-aware)
#> 14                  AUCint (based on AUClast extrapolation)
#> 15      AUCint (based on AUClast extrapolation, dose-aware)
#> 16                                       AUCall (IV dosing)
#> 17                                   AUCinf,obs (IV dosing)
#> 18                                  AUCinf,pred (IV dosing)
#> 19                                   AUCint,all (IV dosing)
#> 20                                  AUCint,last (IV dosing)
#> 21                                      AUClast (IV dosing)
#> 22                                AUCbext (based on AUCall)
#> 23                            AUCbext (based on AUCinf,obs)
#> 24                           AUCbext (based on AUCinf,pred)
#> 25                            AUCbext (based on AUCint,all)
#> 26                           AUCbext (based on AUCint,last)
#> 27                               AUCbext (based on AUClast)
#> 28                                                  AUClast
#> 29                                AUClast (dose-normalized)
#> 30                            AUCpext (based on AUCinf,obs)
#> 31                           AUCpext (based on AUCinf,pred)
#> 32                                                 AUMC,all
#> 33                               AUMC,all (dose-normalized)
#> 34                                             AUMC,inf,obs
#> 35                           AUMC,inf,obs (dose-normalized)
#> 36                          AUMC,inf,pred (dose-normalized)
#> 37                                                AUMC,last
#> 38                              AUMC,last (dose-normalized)
#> 39                                                       C0
#> 40                                                      Cav
#> 41                                    Cav (dose-normalized)
#> 42                                                     Ceoi
#> 43                                     CL (based on AUCall)
#> 44                                    CL (based on AUClast)
#> 45                                 CL (based on AUCinf,obs)
#> 46                                CL (based on AUCinf,pred)
#> 47                                                    Clast
#> 48                                               Clast,pred
#> 49                             Clast,pred (dose-normalized)
#> 50                           Renal clearance (from AUClast)
#> 51                        Renal clearance (from AUCinf,obs)
#> 52                       Renal clearance (from AUCinf,pred)
#> 53                                                     Cmax
#> 54                                   Cmax (dose-normalized)
#> 55                                                     Cmin
#> 56                                   Cmin (dose-normalized)
#> 57                                                  Ctrough
#> 58                                Ctrough (dose-normalized)
#> 59                                    Degree of fluctuation
#> 60                                             Interval End
#> 61                                          Bioavailability
#> 62                                        Fraction excreted
#> 63                                                Half-life
#> 64                    Kel (for IV dosing, based on AUClast)
#> 65                 Kel (for IV dosing, based on AUCinf,obs)
#> 66                Kel (for IV dosing, based on AUCinf,pred)
#> 67                                   Kel (based on AUClast)
#> 68                                Kel (based on AUCinf,obs)
#> 69                               Kel (based on AUCinf,pred)
#> 70                                                       λz
#> 71                                                     λz,n
#> 72                                               λz (t 1ˢᵗ)
#> 73                    MRT (for IV dosing, based on AUClast)
#> 74                 MRT (for IV dosing, based on AUCinf,obs)
#> 75                MRT (for IV dosing, based on AUCinf,pred)
#> 76                                   MRT (based on AUClast)
#> 77           MRT (for multiple dosing, based on AUCinf,obs)
#> 78          MRT (for multiple dosing, based on AUCinf,pred)
#> 79                                MRT (based on AUCinf,obs)
#> 80                               MRT (based on AUCinf,pred)
#> 81                                     Peak-to-trough ratio
#> 82                                                       r²
#> 83                                               Span ratio
#> 84                                           Sparse AUClast
#> 85                                           Interval Start
#> 86                                                    Swing
#> 87                                                    t 1ˢᵗ
#> 88    Effective half-life (for IV dosing, based on MRTlast)
#> 89    Effective half-life (for IV dosing, based on MRT,obs)
#> 90   Effective half-life (for IV dosing, based on MRT,pred)
#> 91                  Effective half-life (based on MRT,last)
#> 92                   Effective half-life (based on MRT,obs)
#> 93                  Effective half-life (based on MRT,pred)
#> 94                                 Time above Concentration
#> 95                                                     Tlag
#> 96                                                    Tlast
#> 97                                  Vd (based on Clast,obs)
#> 98                                 Vd (based on Clast,pred)
#> 99                    Vss (for IV dosing, based on AUClast)
#> 100                Vss (for IV dosing, based on AUCinf,obs)
#> 101               Vss (for IV dosing, based on AUCinf,pred)
#> 102                                  Vss (based on AUClast)
#> 103             Vss (for multiple-dose, based on Clast,obs)
#> 104            Vss (for multiple-dose, based on Clast,pred)
#> 105                               Vss (based on AUCinf,obs)
#> 106                              Vss (based on AUCinf,pred)
#> 107                                Vz (based on AUCinf,obs)
#> 108                               Vz (based on AUCinf,pred)
#>                                 md
#> 1                         AUC~all~
#> 2                        Adj. R^2^
#> 3                               AE
#> 4                      AUC~all,dn~
#> 5                     AUC~inf,obs~
#> 6                  AUC~inf,obs,dn~
#> 7                 AUC~inf,pred,dn~
#> 8                     AUC~int,all~
#> 9                AUC~int,all,dose~
#> 10                AUC~int,inf,obs~
#> 11           AUC~int,inf,obs,dose~
#> 12               AUC~int,inf,pred~
#> 13          AUC~int,inf,pred,dose~
#> 14                   AUC~int,last~
#> 15              AUC~int,last,dose~
#> 16                     AUC~iv,all~
#> 17                 AUC~iv,inf,obs~
#> 18                AUC~iv,inf,pred~
#> 19                 AUC~iv,int,all~
#> 20                AUC~iv,int,last~
#> 21                    AUC~iv,last~
#> 22                 AUC~iv,pbe,all~
#> 23             AUC~iv,pbe,inf,obs~
#> 24            AUC~iv,pbe,inf,pred~
#> 25             AUC~iv,pbe,inf,all~
#> 26            AUC~iv,pbe,inf,last~
#> 27                AUC~iv,pbe,last~
#> 28                       AUC~last~
#> 29                    AUC~last,dn~
#> 30                   AUC~pext,obs~
#> 31                  AUC~pext,pred~
#> 32                       AUMC~all~
#> 33                    AUMC~all,dn~
#> 34                   AUMC~inf,obs~
#> 35                AUMC~inf,obs,dn~
#> 36               AUMC~inf,pred,dn~
#> 37                      AUMC~last~
#> 38                   AUMC~last,dn~
#> 39                            C~0~
#> 40                          C~ave~
#> 41                       C~ave,dn~
#> 42                          C~EOI~
#> 43                              CL
#> 44                        CL~last~
#> 45                         CL~obs~
#> 46                        CL~pred~
#> 47                    CL~last,obs~
#> 48                   CL~last,pred~
#> 49                CL~last,pred,dn~
#> 50                      CL~R,last~
#> 51                       CL~R,obs~
#> 52                      CL~R,pred~
#> 53                          C~max~
#> 54                       C~max,dn~
#> 55                          C~min~
#> 56                       C~min,dn~
#> 57                           C~tr~
#> 58                        C~tr,dn~
#> 59                             dgf
#> 60                          t~end~
#> 61                            f~b~
#> 62                            f~e~
#> 63                         t~half~
#> 64                   k~el,iv,last~
#> 65                    k~el,iv,obs~
#> 66                   k~el,iv,pred~
#> 67                      k~el,last~
#> 68                       k~el,obs~
#> 69                      k~el,pred~
#> 70            <ff:symbol>l</ff>~z~
#> 71          <ff:symbol>l</ff>~z,n~
#> 72  <ff:symbol>l</ff>~z~ (t 1^st^)
#> 73                    MRT~iv,last~
#> 74                     MRT~iv,obs~
#> 75                    MRT~iv,pred~
#> 76                       MRT~last~
#> 77                     MRT~md,obs~
#> 78                    MRT~md,pred~
#> 79                        MRT~obs~
#> 80                       MRT~pred~
#> 81                             PTR
#> 82                            R^2^
#> 83                              SR
#> 84                    AUC~last,sp~
#> 85                        t~start~
#> 86                           SWING
#> 87                        t~first~
#> 88             t~half,eff,iv,last~
#> 89              t~half,eff,iv,obs~
#> 90             t~half,eff,iv,pred~
#> 91                t~half,eff,last~
#> 92                 t~half,eff,obs~
#> 93                t~half,eff,pred~
#> 94                           t~ab~
#> 95                          t~lag~
#> 96                         t~last~
#> 97                        V~d,obs~
#> 98                       V~d,pred~
#> 99                   V~ss,iv,last~
#> 100                   V~ss,iv,obs~
#> 101                  V~ss,iv,pred~
#> 102                  V~ss,iv,last~
#> 103                   V~ss,md,obs~
#> 104                  V~ss,md,pred~
#> 105                      V~ss,obs~
#> 106                     V~ss,pred~
#> 107                       V~z,obs~
#> 108                      V~z,pred~
#>                                                       latex
#> 1                                                    AUCall
#> 2                                               $r^2_{adj}$
#> 3                                           Amount excreted
#> 4                                  AUCall (dose-normalized)
#> 5                                                AUCinf,obs
#> 6                              AUCinf,obs (dose-normalized)
#> 7                             AUCinf,pred (dose-normalized)
#> 8                    AUCint (based on AUCall extrapolation)
#> 9        AUCint (based on AUCall extrapolation, dose-aware)
#> 10               AUCint (based on AUCinf,obs extrapolation)
#> 11   AUCint (based on AUCinf,obs extrapolation, dose-aware)
#> 12              AUCint (based on AUCinf,pred extrapolation)
#> 13  AUCint (based on AUCinf,pred extrapolation, dose-aware)
#> 14                  AUCint (based on AUClast extrapolation)
#> 15      AUCint (based on AUClast extrapolation, dose-aware)
#> 16                                       AUCall (IV dosing)
#> 17                                   AUCinf,obs (IV dosing)
#> 18                                  AUCinf,pred (IV dosing)
#> 19                                   AUCint,all (IV dosing)
#> 20                                  AUCint,last (IV dosing)
#> 21                                      AUClast (IV dosing)
#> 22                                AUCbext (based on AUCall)
#> 23                            AUCbext (based on AUCinf,obs)
#> 24                           AUCbext (based on AUCinf,pred)
#> 25                            AUCbext (based on AUCint,all)
#> 26                           AUCbext (based on AUCint,last)
#> 27                               AUCbext (based on AUClast)
#> 28                                                  AUClast
#> 29                                AUClast (dose-normalized)
#> 30                            AUCpext (based on AUCinf,obs)
#> 31                           AUCpext (based on AUCinf,pred)
#> 32                                                 AUMC,all
#> 33                               AUMC,all (dose-normalized)
#> 34                                             AUMC,inf,obs
#> 35                           AUMC,inf,obs (dose-normalized)
#> 36                          AUMC,inf,pred (dose-normalized)
#> 37                                                AUMC,last
#> 38                              AUMC,last (dose-normalized)
#> 39                                                       C0
#> 40                                                      Cav
#> 41                                    Cav (dose-normalized)
#> 42                                                     Ceoi
#> 43                                     CL (based on AUCall)
#> 44                                    CL (based on AUClast)
#> 45                                 CL (based on AUCinf,obs)
#> 46                                CL (based on AUCinf,pred)
#> 47                                                    Clast
#> 48                                               Clast,pred
#> 49                             Clast,pred (dose-normalized)
#> 50                           Renal clearance (from AUClast)
#> 51                        Renal clearance (from AUCinf,obs)
#> 52                       Renal clearance (from AUCinf,pred)
#> 53                                                     Cmax
#> 54                                   Cmax (dose-normalized)
#> 55                                                     Cmin
#> 56                                   Cmin (dose-normalized)
#> 57                                                  Ctrough
#> 58                                Ctrough (dose-normalized)
#> 59                                    Degree of fluctuation
#> 60                                             Interval End
#> 61                                          Bioavailability
#> 62                                        Fraction excreted
#> 63                                                Half-life
#> 64                    Kel (for IV dosing, based on AUClast)
#> 65                 Kel (for IV dosing, based on AUCinf,obs)
#> 66                Kel (for IV dosing, based on AUCinf,pred)
#> 67                                   Kel (based on AUClast)
#> 68                                Kel (based on AUCinf,obs)
#> 69                               Kel (based on AUCinf,pred)
#> 70                                             $\\lambda_z$
#> 71                       Number of points used for lambda_z
#> 72                              First time for $\\lambda_z$
#> 73                    MRT (for IV dosing, based on AUClast)
#> 74                 MRT (for IV dosing, based on AUCinf,obs)
#> 75                MRT (for IV dosing, based on AUCinf,pred)
#> 76                                   MRT (based on AUClast)
#> 77           MRT (for multiple dosing, based on AUCinf,obs)
#> 78          MRT (for multiple dosing, based on AUCinf,pred)
#> 79                                MRT (based on AUCinf,obs)
#> 80                               MRT (based on AUCinf,pred)
#> 81                                     Peak-to-trough ratio
#> 82                                                    $r^2$
#> 83                                               Span ratio
#> 84                                           Sparse AUClast
#> 85                                           Interval Start
#> 86                                                    Swing
#> 87                                                   Tfirst
#> 88    Effective half-life (for IV dosing, based on MRTlast)
#> 89    Effective half-life (for IV dosing, based on MRT,obs)
#> 90   Effective half-life (for IV dosing, based on MRT,pred)
#> 91                  Effective half-life (based on MRT,last)
#> 92                   Effective half-life (based on MRT,obs)
#> 93                  Effective half-life (based on MRT,pred)
#> 94                                 Time above Concentration
#> 95                                                     Tlag
#> 96                                                    Tlast
#> 97                                  Vd (based on Clast,obs)
#> 98                                 Vd (based on Clast,pred)
#> 99                    Vss (for IV dosing, based on AUClast)
#> 100                Vss (for IV dosing, based on AUCinf,obs)
#> 101               Vss (for IV dosing, based on AUCinf,pred)
#> 102                                  Vss (based on AUClast)
#> 103             Vss (for multiple-dose, based on Clast,obs)
#> 104            Vss (for multiple-dose, based on Clast,pred)
#> 105                               Vss (based on AUCinf,obs)
#> 106                              Vss (based on AUCinf,pred)
#> 107                                Vz (based on AUCinf,obs)
#> 108                               Vz (based on AUCinf,pred)
#>                                                                                                                                                                                                                                                                          description
#> 1                        The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 2                                                                                                                                                                                                                                The adjusted r^2 value of the half-life calculation
#> 3                                                                                                                                                                                                                                The amount excreted (typically into urine or feces)
#> 4                                                                                                                                                                                                                                                             Dose normalized aucall
#> 5                                                                                                                                  The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 6                                                                                                                                                                                                                                                         Dose normalized aucinf.obs
#> 7                                                                                                                                                                                                                                                        Dose normalized aucinf.pred
#> 8                                                                                              The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 9                                The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 10                                                                                                                                                    The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 11                                                                                      The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 12                                                                                             The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 13                               The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 14                                                                                                                                                    The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 15                                                                                      The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 16                                                                                                                                                                                        The AUCall calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 17                                                                                                                                                                                    The AUCinf,obs calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 18                                                                                                                                                                                              The  calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 19                                                                                                                                                                                    The AUCint,all calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 20                                                                                                                                                                                   The AUCint,last calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 21                                                                                                                                                                                       The AUClast calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 22                                                                                                                                                                                                             The back-extrapolation percent for intravenous dosing based on AUCall
#> 23                                                                                                                                                                                                         The back-extrapolation percent for intravenous dosing based on AUCinf,obs
#> 24                                                                                                                                                                                                        The back-extrapolation percent for intravenous dosing based on AUCinf,pred
#> 25                                                                                                                                                                                                         The back-extrapolation percent for intravenous dosing based on AUCint,all
#> 26                                                                                                                                                                                                        The back-extrapolation percent for intravenous dosing based on AUCint,last
#> 27                                                                                                                                                                                                            The back-extrapolation percent for intravenous dosing based on AUClast
#> 28                                                                                                                                        The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 29                                                                                                                                                                                                                                                           Dose normalized auclast
#> 30                                                                                                                                                                                         Percent of the AUCinf that is extrapolated after Tlast calculated from the observed Clast
#> 31                                                                                                                                                                                        Percent of the AUCinf that is extrapolated after Tlast calculated from the predicted Clast
#> 32  The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification plus the moment of the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 33                                                                                                                                                                                                                                                           Dose normalized aumcall
#> 34                                                                                                                          The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 35                                                                                                                                                                                                                                                       Dose normalized aumcinf.obs
#> 36                                                                                                                                                                                                                                                      Dose normalized aumcinf.pred
#> 37                                                                                                                                 The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification
#> 38                                                                                                                                                                                                                                                          Dose normalized aumclast
#> 39                                                                                                                                                                                                                                           Initial concentration after an IV bolus
#> 40                                                                                                                                                                                                                                      The average concentration during an interval
#> 41                                                                                                                                                                                                                                                               Dose normalized cav
#> 42                                                                                                                                                                                                                                              Concentration at the end of infusion
#> 43                                                                                                                                                                                                                       Clearance or observed oral clearance calculated with AUCall
#> 44                                                                                                                                                                                                                          Clearance or observed oral clearance calculated to Clast
#> 45                                                                                                                                                                                                               Clearance or observed oral clearance calculated with observed Clast
#> 46                                                                                                                                                                                                              Clearance or observed oral clearance calculated with predicted Clast
#> 47                                                                                                                                                                                                                 The last concentration observed above the limit of quantification
#> 48                                                                                                                                                                                                                          The concentration at Tlast as predicted by the half-life
#> 49                                                                                                                                                                                                                                                        Dose normalized clast.pred
#> 50                                                                                                                                                                                                                                      The renal clearance calculated using AUClast
#> 51                                                                                                                                                                                                                                   The renal clearance calculated using AUCinf,obs
#> 52                                                                                                                                                                                                                                  The renal clearance calculated using AUCinf,pred
#> 53                                                                                                                                                                                                                                                    Maximum observed concentration
#> 54                                                                                                                                                                                                                                                              Dose normalized cmax
#> 55                                                                                                                                                                                                                                                    Minimum observed concentration
#> 56                                                                                                                                                                                                                                                              Dose normalized cmin
#> 57                                                                                                                                                                                                                                        The trough (end of interval) concentration
#> 58                                                                                                                                                                                                                                                           Dose normalized ctrough
#> 59                                                                                                                                                                                                                                                             Degree of fluctuation
#> 60                                                                                                                                                                                                                                Ending time of the interval (potentially infinity)
#> 61                                                                                                                                                                                                                                       Bioavailability or relative bioavailability
#> 62                                                                                                                                                                                                                                                 The fraction of the dose excreted
#> 63                                                                                                                                                                                                                                                          The (terminal) half-life
#> 64                                                                                                                                                                                                                     Elimination rate (as calculated from the intravenous MRTlast)
#> 65                                                                                                                                                                                                                      Elimination rate (as calculated from the intravenous MRTobs)
#> 66                                                                                                                                                                                                                     Elimination rate (as calculated from the intravenous MRTpred)
#> 67                                                                                                                                                                                                                       Elimination rate (as calculated from the MRT using AUClast)
#> 68                                                                                                                                                                                                                 Elimination rate (as calculated from the MRT with observed Clast)
#> 69                                                                                                                                                                                                                Elimination rate (as calculated from the MRT with predicted Clast)
#> 70                                                                                                                                                                                                                                    The elimination rate of the terminal half-life
#> 71                                                                                                                                                                                                                        The number of points used for the calculation of half-life
#> 72                                                                                                                                                                                                                        The first time point used for the calculation of half-life
#> 73                                                                                                                                                                           The mean residence time to the last observed concentration above the LOQ correcting for dosing duration
#> 74                                                                                                                                                                                           The mean residence time to infinity using observed Clast correcting for dosing duration
#> 75                                                                                                                                                                                          The mean residence time to infinity using predicted Clast correcting for dosing duration
#> 76                                                                                                                                                                                                          The mean residence time to the last observed concentration above the LOQ
#> 77                                                                                                                                                                                          The mean residence time with multiple dosing and nonlinear kinetics using observed Clast
#> 78                                                                                                                                                                                         The mean residence time with multiple dosing and nonlinear kinetics using predicted Clast
#> 79                                                                                                                                                                                                                          The mean residence time to infinity using observed Clast
#> 80                                                                                                                                                                                                                         The mean residence time to infinity using predicted Clast
#> 81                                                                                                                                                                                                                                                   Peak-to-Trough ratio (fraction)
#> 82                                                                                                                                                                                                                                        The r^2 value of the half-life calculation
#> 83                                                                                                                                                                                                         The ratio of the half-life to the duration used for half-life calculation
#> 84                                                                                                                For sparse PK sampling, the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 85                                                                                                                                                                                                                                                     Starting time of the interval
#> 86                                                                                                                                                                                                                                                            Swing relative to Cmin
#> 87                                                                                                                                                                                                                 Time of the first concentration above the limit of quantification
#> 88                                                                                                                                                                                                              The effective half-life (as determined from the intravenous MRTlast)
#> 89                                                                                                                                                                                                               The effective half-life (as determined from the intravenous MRTobs)
#> 90                                                                                                                                                                                                              The effective half-life (as determined from the intravenous MRTpred)
#> 91                                                                                                                                                                                                                          The effective half-life (as determined from the MRTlast)
#> 92                                                                                                                                                                                                                           The effective half-life (as determined from the MRTobs)
#> 93                                                                                                                                                                                                                          The effective half-life (as determined from the MRTpred)
#> 94                                                                                                                                                                                                                                                  Time above a given concentration
#> 95                                                                                                                                                                                                                                                                          Lag time
#> 96                                                                                                                                                                                                         Time of the last concentration observed above the limit of quantification
#> 97                                                                                                                                                                                                           Apparent observed volume of distribution calculated with observed Clast
#> 98                                                                                                                                                                                                          Apparent observed volume of distribution calculated with predicted Clast
#> 99                                                                                                                                                                                       The steady-state volume of distribution with intravenous infusion calculating through Tlast
#> 100                                                                                                                                                                                           The steady-state volume of distribution with intravenous infusion using observed Clast
#> 101                                                                                                                                                                                          The steady-state volume of distribution with intravenous infusion using predicted Clast
#> 102                                                                                                                                                                                                                The steady-state volume of distribution calculating through Tlast
#> 103                                                                                                                                                                                    The steady-state volume of distribution for nonlinear multiple-dose data using observed Clast
#> 104                                                                                                                                                                                   The steady-state volume of distribution for nonlinear multiple-dose data using predicted Clast
#> 105                                                                                                                                                                                                                     The steady-state volume of distribution using observed Clast
#> 106                                                                                                                                                                                                                    The steady-state volume of distribution using predicted Clast
#> 107                                                                                                                                                                                                                         The terminal volume of distribution using observed Clast
#> 108                                                                                                                                                                                                                        The terminal volume of distribution using predicted Clast
#>