Fetch ruminate State

Merges default app options with the changes made in the UI

Usage

NCA_fetch_state(
  id,
  input,
  session,
  FM_yaml_file,
  MOD_yaml_file,
  id_ASM,
  id_UD,
  id_DW,
  react_state
)

Arguments

id

Shiny module ID

input

Shiny input variable

session

Shiny session variable

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

id_ASM

ID string for the app state management module used to save and load app states

id_UD

ID string for the upload data module used to save and load app states

id_DW

ID string for the data wrangling module used to save and load app states

Value

list containing the current state of the app including default values from the yaml file as well as any changes made by the user. The list has the following structure:

yaml: Full contents of the supplied yaml file.

MC: Module components of the yaml file.

NCA: # JMH populate the rest of this list

ana_cntr: Analysis counter.

anas: List of analyses: Each analysis has the following structure:

ana_dsview: Dataset view/ID (name from DSV) selected as a data source for this analysis.

ana_scenario: JMH

checksum: JMH

code: Code to generate analysis from start to finish or error messages if code generation/analysis failed.

code_components: List containing the different components from code

col_conc: Column from ana_dsview containing the concentration data.

col_cycle: Column from ana_dsview containing the dose dose cycle/number.

col_dose: Column from ana_dsview containing the dose amount.

col_dur: Column from ana_dsview containing the infusion duration or N/A if unused.

col_group: Columns from ana_dsview containing other grouping variables.

col_id: Column from ana_dsview containing the subject IDs.

col_ntime: Column from ana_dsview containing the nominal time values

col_route: Column from ana_dsview containing the dosing route.

col_time: Column from ana_dsview containing the time values.

id: Character id (ana_idx)

idx: Numeric id (1)

include_units: Boolean variable indicating in units should included in the analysis. be

interval_start: Beginning of the analysis interval from the UI.

interval_stop: End of the analysis interval from the UI.

intervals: List of the intervals to include.

isgood: Current status of the analysis.

key: Analysis key acts as a title/caption (user editable)

msgs: Messages generated when checking configuration and analysis options.

nca_config: List of NCA configuration options for this analysis.

nca_object_name: Prefix for NCA objects associated with this analyis.

nca_parameters: NCA parameters selected for calculation in the UI.

notes: Analysis notes (user editable)

objs: List of names and values for objects created with generated code.

sampling: Sampling method either "sparse" or "serial"

units_amt: Amount units.

units_conc: Concentration units.

units_dose: Dosing units.

units_time: Time units.

button_counters: JMH.

current_ana: Currently selected analysis (list name element from anas).

DSV: Available data source views (see FM_fetch_ds)

checksum: This is an MD5 sum of the (JMH update) element and can be

nca_config: List of PKNCA configuration options for this analysis.

ui: JMH.

ui_ana_map: Map between UI element names and analysis in the object you get from NCA_fetch_current_ana()

ui_hold: JMH.

ui_ids: JMH.

MOD_TYPE: Character data containing the type of module "NCA"

id: Character data containing the module id module in the session variable.

FM_yaml_file: App configuration file with FM as main section.

MOD_yaml_file: Module configuration file with MC as main section.