ruminate: Shiny app and module to facilitate pharamacometrics analysis

This is done by creating a Shiny interface to different tools for data transformation (dplyr and tidyr), plotting (ggplot2), and noncompartmental analysis (PKNCA). These results can be reported in Excel, Word or PowerPoint. The state of the app can be saved and loaded at a later date. When saved, a script is generated to reproduce the different actions in the Shiny interface.

Runs the pharmacometrics ruminate app.

Usage

ruminate(
  host = "127.0.0.1",
  port = 3838,
  server_opts = list(shiny.maxRequestSize = 30 * 1024^2),
  mksession = FALSE
)

Arguments

host

Hostname of the server ("127.0.0.1")

port

Port number for the app (3838)

server_opts

List of options (names) and their vlues (value) e.g. list(shiny.maxRequestSize = 30 * 1024^2).

mksession

Boolean value, when TRUE will load test session data for app testing

Value

Nothing is returned, this function just runs the built-in ruminate app.

See also

Author

Maintainer: John Harrold john.m.harrold@gmail.com (ORCID)

Examples

if (interactive()) {
ruminate()
}