Fetch Clinical Trial Simulator Module Datasets

Fetches the datasets produced by the module. For each cohort this will be the simulation timecourse and the event table

Usage

CTS_fetch_ds(state)

Arguments

state

CTS state from CTS_fetch_state()

Value

Character object vector with the lines of code

list containing the following elements

isgood: Return status of the function.

hasds: Boolean indicator if the module has any datasets

msgs: Messages to be passed back to the user.

ds: List with datasets. Each list element has the name of the R-object for that dataset. Each element has the following structure:

label: Text label for the dataset

MOD_TYPE: Short name for the type of module.

id: module ID

DS: Dataframe containing the actual dataset.

DSMETA: Metadata describing DS

code: Complete code to build dataset.

checksum: Module checksum.

DSchecksum: Dataset checksum.

Examples

# We need a module state:
sess_res = CTS_test_mksession(session=list(), full_session=FALSE)
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#> → MB: module checksum updated:a8d5d330f91feb7be769e08518458dd8
#> → MB: State initialized
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/My Drive/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: cdfcbf111f1e9cc959894800df9e3aa7
#> → MB: module checksum updated:bf44543da2f27ba698ac16981e68c0cb
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → CTS:   dest:   file.path("config","report.docx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → CTS:   dest:   file.path("config","report.pptx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → CTS:   dest:   file.path("config","report.yaml")
#> → CTS: module checksum updated: f312310118fbc0d82246501049f0eac7
#> → CTS: State initialized
#> → CTS: cohort checksum updated: f07847307239b964d91692e8b32808db
#> → CTS: module checksum updated: c462b0919a3c63afc12588f62e835a11
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: 8b7ef832249bdb946586fa9b02b596db
#> → CTS: module checksum updated: c3763952a51c7f44564b6448058cc876
#>  
#>  
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> → CTS: cohort checksum updated: 3cce8f7adcc2fa8ee03fc3caf9470d7d
#> → CTS: module checksum updated: dd42c9a26614d3edfb85e3a2b0c062c6
state = sess_res$state

ds = CTS_fetch_ds(state)

ds
#> $hasds
#> [1] TRUE
#> 
#> $isgood
#> [1] FALSE
#> 
#> $msgs
#> NULL
#> 
#> $ds
#> $ds$CTS_obj_1_sim_tc
#> $ds$CTS_obj_1_sim_tc$label
#> character(0)
#> 
#> $ds$CTS_obj_1_sim_tc$MOD_TYPE
#> [1] "CTS"
#> 
#> $ds$CTS_obj_1_sim_tc$id
#> NULL
#> 
#> $ds$CTS_obj_1_sim_tc$DS
#>    id   time       ka       cl       vc        kel           Cc     ipredSim
#> 13  1   0.00 1.568312 3.007394 31.50039 0.09547165 3.174564e-02 3.174564e-02
#> 2   1   0.01 1.568312 3.007394 31.50039 0.09547165 3.171534e-02 3.171534e-02
#> 31  1  25.00 1.568312 3.007394 31.50039 0.09547165 2.918198e-03 2.918198e-03
#> 4   1  50.00 1.568312 3.007394 31.50039 0.09547165 2.682523e-04 2.682523e-04
#> 5   1  75.00 1.568312 3.007394 31.50039 0.09547165 2.465879e-05 2.465879e-05
#> 6   1 100.00 1.568312 3.007394 31.50039 0.09547165 2.266722e-06 2.266722e-06
#> 7   2   0.00 1.568312 2.258416 31.50039 0.07169486 3.174564e-02 3.174564e-02
#> 8   2   0.01 1.568312 2.258416 31.50039 0.07169486 3.172288e-02 3.172288e-02
#> 9   2  25.00 1.568312 2.258416 31.50039 0.07169486 5.287709e-03 5.287709e-03
#> 10  2  50.00 1.568312 2.258416 31.50039 0.07169486 8.807459e-04 8.807459e-04
#> 11  2  75.00 1.568312 2.258416 31.50039 0.07169486 1.467008e-04 1.467008e-04
#> 12  2 100.00 1.568312 2.258416 31.50039 0.07169486 2.443510e-05 2.443510e-05
#>             sim depot      central Single_Dose
#> 13 3.217246e-03     0 1.000000e+00        true
#> 2  4.717418e-02     0 9.990457e-01        true
#> 31 1.981518e-03     0 9.192439e-02        true
#> 4  3.345132e-04     0 8.450054e-03        true
#> 5  3.219750e-05     0 7.767614e-04        true
#> 6  2.587942e-06     0 7.140263e-05        true
#> 7  2.675555e-02     0 1.000000e+00        true
#> 8  8.676035e-03     0 9.992833e-01        true
#> 9  7.611851e-03     0 1.665649e-01        true
#> 10 1.358158e-03     0 2.774384e-02        true
#> 11 1.777412e-04     0 4.621132e-03        true
#> 12 3.178687e-05     0 7.697152e-04        true
#> 
#> $ds$CTS_obj_1_sim_tc$DSMETA
#> [1] "Timecourse"
#> 
#> $ds$CTS_obj_1_sim_tc$code
#> [1] "\n# Defining covariates\nCTS_obj_1_cov = list()\n\n# Generating the subjects\nCTS_obj_1_subs = mk_subjects(object = MB_obj_1_rx,\n  nsub   = 2,\n  covs   = CTS_obj_1_cov)\n\n# Creating rules\nCTS_obj_1_rules = list()\n\nCTS_obj_1_rules[[\"Single_Dose\"]] = \nlist(condition = \"time == 0\", action = list(type = \"dose\", state = \"central\", \n    values = \"c(1)\", times = \"c(0)\", durations = \"c(0)\"))\n\n# rxSolve options\nCTS_obj_1_rxopts = list(\n  atol = 1e-8,\n  rtol = 1e-6,\n  sigdig = NULL)\n\n# Output times\nCTS_obj_1_output_times = formods::linspace(0,100,5)\n\n# Running simulation\nCTS_obj_1_simres =                                 \n simulate_rules(object        = MB_obj_1_rx,\n                subjects      = CTS_obj_1_subs[[\"subjects\"]],\n                eval_times    = c(0),\n                output_times  = CTS_obj_1_output_times,\n                rules         = CTS_obj_1_rules,\n                rx_options    = CTS_obj_1_rxopts)\n# Collecting the simulation and event history values\nCTS_obj_1_sim_tc = CTS_obj_1_simres[[\"simall\"]]\nCTS_obj_1_sim_ev = CTS_obj_1_simres[[\"ev_history\"]]\n"
#> 
#> $ds$CTS_obj_1_sim_tc$checksum
#> [1] "dd42c9a26614d3edfb85e3a2b0c062c6"
#> 
#> $ds$CTS_obj_1_sim_tc$DSchecksum
#> [1] "edc19386a0709503304dfd094af35e38"
#> 
#> 
#> $ds$CTS_obj_1_sim_ev
#> $ds$CTS_obj_1_sim_ev$label
#> character(0)
#> 
#> $ds$CTS_obj_1_sim_ev$MOD_TYPE
#> [1] "CTS"
#> 
#> $ds$CTS_obj_1_sim_ev$id
#> NULL
#> 
#> $ds$CTS_obj_1_sim_ev$DS
#> ── EventTable with 18 records ──
#> 6 dosing records (see $get.dosing(); add with add.dosing or et)
#> 12 observation times (see $get.sampling(); add with add.sampling or et)
#> ── First part of : ──
#> # A tibble: 18 × 5
#>       id   time cmt       amt evid                        
#>    <int>  <dbl> <chr>   <dbl> <evid>                      
#>  1     1   0    Cc         NA 0:Observation               
#>  2     1   0    central     1 1:Dose (Add)                
#>  3     1   0    depot       0 4:Reset&Dose                
#>  4     1   0    central     0 4:Reset&Dose                
#>  5     1   0.01 Cc         NA 0:Observation               
#>  6     1  25    Cc         NA 0:Observation               
#>  7     1  50    Cc         NA 0:Observation               
#>  8     1  75    Cc         NA 0:Observation               
#>  9     1 100    Cc         NA 0:Observation               
#> 10     2   0    Cc         NA 0:Observation               
#> 11     2   0    central     1 1:Dose (Add)                
#> 12     2   0    depot       0 4:Reset&Dose                
#> 13     2   0    central     0 4:Reset&Dose                
#> 14     2   0.01 Cc         NA 0:Observation               
#> 15     2  25    Cc         NA 0:Observation               
#> 16     2  50    Cc         NA 0:Observation               
#> 17     2  75    Cc         NA 0:Observation               
#> 18     2 100    Cc         NA 0:Observation               
#> 
#> $ds$CTS_obj_1_sim_ev$DSMETA
#> [1] "Events"
#> 
#> $ds$CTS_obj_1_sim_ev$code
#> [1] "\n# Defining covariates\nCTS_obj_1_cov = list()\n\n# Generating the subjects\nCTS_obj_1_subs = mk_subjects(object = MB_obj_1_rx,\n  nsub   = 2,\n  covs   = CTS_obj_1_cov)\n\n# Creating rules\nCTS_obj_1_rules = list()\n\nCTS_obj_1_rules[[\"Single_Dose\"]] = \nlist(condition = \"time == 0\", action = list(type = \"dose\", state = \"central\", \n    values = \"c(1)\", times = \"c(0)\", durations = \"c(0)\"))\n\n# rxSolve options\nCTS_obj_1_rxopts = list(\n  atol = 1e-8,\n  rtol = 1e-6,\n  sigdig = NULL)\n\n# Output times\nCTS_obj_1_output_times = formods::linspace(0,100,5)\n\n# Running simulation\nCTS_obj_1_simres =                                 \n simulate_rules(object        = MB_obj_1_rx,\n                subjects      = CTS_obj_1_subs[[\"subjects\"]],\n                eval_times    = c(0),\n                output_times  = CTS_obj_1_output_times,\n                rules         = CTS_obj_1_rules,\n                rx_options    = CTS_obj_1_rxopts)\n# Collecting the simulation and event history values\nCTS_obj_1_sim_tc = CTS_obj_1_simres[[\"simall\"]]\nCTS_obj_1_sim_ev = CTS_obj_1_simres[[\"ev_history\"]]\n"
#> 
#> $ds$CTS_obj_1_sim_ev$checksum
#> [1] "dd42c9a26614d3edfb85e3a2b0c062c6"
#> 
#> $ds$CTS_obj_1_sim_ev$DSchecksum
#> [1] "2152329158c081210a67850665e6cf62"
#> 
#> 
#>