Populate Session Data for Module Testing

Populates the supplied session variable for testing.

Usage

CTS_test_mksession(session = list())

Arguments

session

Shiny session variable (in app) or a list (outside of app)

Value

The CTS portion of the all_sess_res returned from FM_app_preload

See also

Examples

# \donttest{
sess_res = CTS_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: module isgood: TRUE
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> → MB: module checksum updated:24933f86b657b9503f22440e8c4d3cac
#> → MB: State initialized
#> → MB: loading model idx: 1
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 8b3f3b8af91627c0483413c0e0442669
#> → MB: module checksum updated:286a139c83f58035b4e9dc5acf49d038
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 5f1b23b2d9a79cf5846ff0d31aa546e5
#> → MB: module checksum updated:0bdd4b6718625094549797c04d55d98e
#> → MB:   -> setting name: One compartment model
#> → MB:   -> setting time scale: hours
#> → MB:   -> setting base from: user
#> → MB:   -> setting catalog selection: 
#> → MB:   -> setting base model id: manual
#> → MB:   -> setting base model name: manual
#> → MB: model checksum updated: dabcf0060488da635fd86bd7199eeaac
#> → MB: module checksum updated:d92ba98b47330355c06634c9d161c53e
#> → MB: added element idx: 1
#> → MB: loading model idx: 2
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: 581647073aa247db2cfb62a7f8824341
#> → MB: module checksum updated:76bf26c1c68ea7d0ba734efe07932ca9
#> → MB:   -> setting name: PK Biomarker
#> → MB:   -> setting time scale: days
#> → MB:   -> setting base from: user
#> → MB:   -> setting catalog selection: 
#> → MB:   -> setting base model id: manual
#> → MB:   -> setting base model name: manual
#> → MB: model checksum updated: 285c2f14ea0a75939504345a76fa647f
#> → MB: module checksum updated:defa767de59af224ca683c7c6cf2983a
#> → MB: added element idx: 2
#> → MB: module isgood: TRUE
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → CTS:   dest:   file.path("config","report.docx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → CTS:   dest:   file.path("config","report.pptx")
#> → CTS: including file
#> → CTS:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → CTS:   dest:   file.path("config","report.yaml")
#> → CTS: source model change detected
#> → CTS:  > covariates reset
#> → CTS: cohort checksum updated: 7662291cfda678368edef72abe61a255
#> → CTS: module checksum updated: 351269c95d7155b66deffb91bd3f52fc
#> → CTS: State initialized
#> → CTS: source model change detected
#> → CTS:  > covariates reset
#> → CTS: cohort checksum updated: f2736e3fa1b53e235b00e78e2ee13b37
#> → CTS: module checksum updated: cd06cf7a9ca5f5f5594b18daf3411dfe
#> → CTS: loading element idx: 1
#> → CTS: setting cohort name: Single dose
#> → CTS: setting model source: MB_obj_1_rx
#> → CTS: setting trial options:
#> → CTS:   - nsub: 3
#> → CTS:   - visit_times: 0
#> → CTS:   - trial_end: 100
#> → CTS:   - fpage: 1
#> → CTS:   - dvcols: Cc
#> → CTS:   - time_scale: 
#> → CTS:   - tc_dim: d2x4
#> → CTS:   - evplot: 1
#> → CTS:   - cts_config_atol: 1e-8
#> → CTS:   - cts_config_rtol: 1e-6
#> → CTS:   - cts_config_sigdig: NULL
#> → CTS:   - cts_config_seed: 5446
#> → CTS:   - cts_config_nsteps: 5
#> → CTS: cohort checksum updated: 9448ed03b3e8641502fa8a38b39bae6d
#> → CTS: module checksum updated: 4826a1c05cc1af102709f9a723efc628
#> → CTS: adding rule: first_dose
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: 85f8a17c030e10b80d7f00bd9202b176
#> → CTS: module checksum updated: f070406eea1986f3858ab3c985e07025
#>  
#>  
#> using C compiler: ‘Apple clang version 15.0.0 (clang-1500.3.9.4)’
#> using SDK: ‘’
#>  
#>  
#> using C compiler: ‘Apple clang version 15.0.0 (clang-1500.3.9.4)’
#> using SDK: ‘’
#> Warning: multi-subject simulation without without 'omega'
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> → CTS: simulation complete
#> → CTS: plot complete
#> → CTS: cohort checksum updated: b9acb3e1f0db7b0dfec5d281bebd561e
#> → CTS: module checksum updated: 84ff06dd86631023d95f5f4302dab83f
#> → CTS: added element idx: 1
#> → CTS: loading element idx: 2
#> → CTS: setting cohort name: Biomarker Titration
#> → CTS: setting model source: MB_obj_2_rx
#> → CTS: source model change detected
#> → CTS:  > covariates reset
#> → CTS: setting trial options:
#> → CTS:   - nsub: 12
#> → CTS:   - visit_times: (0:6)*28*2
#> → CTS:   - trial_end: 400
#> → CTS:   - fpage: 1
#> → CTS:   - dvcols: C_ng_ml
#> → CTS:   - dvcols: BM
#> → CTS:   - time_scale: weeks
#> → CTS:   - tc_dim: d2x4
#> → CTS:   - evplot: 1
#> → CTS:   - cts_config_atol: 1e-8
#> → CTS:   - cts_config_rtol: 1e-6
#> → CTS:   - cts_config_sigdig: NULL
#> → CTS:   - cts_config_seed: 5446
#> → CTS:   - cts_config_nsteps: 50
#> → CTS: adding covariates:
#> → CTS:   - WT: cont_lognormal (70, .1)
#> → CTS:   - SEX_ID: discrete (0, 1)
#> → CTS:   - SUBTYPE_ID: fixed (1)
#> → CTS: cohort checksum updated: 21169e8176e6637d79512ed9fec65104
#> → CTS: module checksum updated: 291eb17ee4540ac0416ad8e3283c368f
#> → CTS: adding rule: first_dose
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: d4400762e1d0327b9fc92a861a14ad39
#> → CTS: module checksum updated: efcf92e826d8587d92a34b1be9088bbe
#> → CTS: adding rule: keep_dose
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: 495d7d176f9f3b2e950ce1120d18701e
#> → CTS: module checksum updated: a320180fc636367f667ad7ddb0c369f5
#> → CTS: adding rule: decrease_dose
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: 9039993fa25bcb50300ed1c964f5a661
#> → CTS: module checksum updated: 1a9e0ba567ebecb5071fc07e7a76a821
#> → CTS: adding rule: increase_dose
#> → CTS: add rule success
#> → CTS: rule added
#> → CTS: cohort checksum updated: f961d2185a7996315f302cf36d19a32e
#> → CTS: module checksum updated: 0a34f2fae4d7fb4b9cf6a0392aa0c0f2
#>  
#>  
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Evaluation times ■■■■■■■■■                         25% | ETA:  4s
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Warning: IDs without zero-time start at the first observed time: 12 11 10 9 8 7 6 5 4 3 2 1
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Warning: IDs without zero-time start at the first observed time: 12 11 10 9 8 7 6 5 4 3 2 1
#> Evaluation times ■■■■■■■■■■■■■■■■                  50% | ETA:  4s
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Warning: IDs without zero-time start at the first observed time: 12 11 10 9 8 7 6 5 4 3 2 1
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Warning: IDs without zero-time start at the first observed time: 12 11 10 9 8 7 6 5 4 3 2 1
#> Evaluation times ■■■■■■■■■■■■■■■■■■■■■■■           75% | ETA:  2s
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Warning: IDs without zero-time start at the first observed time: 12 11 10 9 8 7 6 5 4 3 2 1
#>  
#>  
#> Warning: multi-subject simulation without without 'omega'
#> Warning: IDs without zero-time start at the first observed time: 12 11 10 9 8 7 6 5 4 3 2 1
#> Evaluation times ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> → CTS: simulation complete
#> → CTS: plot complete
#> → CTS: cohort checksum updated: 764a686fd25f2a10461c154718fa5076
#> → CTS: module checksum updated: 948b8cda224ddec181d685484722ad23
#> → CTS: added element idx: 2
#> → CTS: module isgood: TRUE
# }