Model Builder State Server

Server function for the Model Builder Shiny Module

Usage

MB_Server(
  id,
  id_ASM = "ASM",
  FM_yaml_file = system.file(package = "formods", "templates", "formods.yaml"),
  MOD_yaml_file = system.file(package = "ruminate", "templates", "MB.yaml"),
  deployed = FALSE,
  react_state = NULL
)

Arguments

id

An ID string that corresponds with the ID used to call the modules UI elements

id_ASM

ID string for the app state managment module used to save and load app states

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

deployed

Boolean variable indicating whether the app is deployed or not.

react_state

Variable passed to server to allow reaction outside of module (NULL)

Value

MB Server object