Populate Session Data for Module Testing

Populates the supplied session variable for testing.

Usage

MB_test_mksession(session, id = "MB", id_ASM = "ASM", full_session = TRUE)

Arguments

session

Shiny session variable (in app) or a list (outside of app)

id

An ID string that corresponds with the ID used to call the modules UI elements

id_ASM

An ID string that corresponds with the ID used to call the ASM module

full_session

Boolean to indicate if the full test session should be created (default TRUE).

Value

list with the following elements

isgood: Boolean indicating the exit status of the function.

session: The value Shiny session variable (in app) or a list (outside of app) after initialization.

input: The value of the shiny input at the end of the session initialization.

state: App state.

rsc: The react_state components.

Examples

sess_res = MB_test_mksession(session=list(), full_session=FALSE)
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → MB:   dest:   file.path("config","report.docx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → MB:   dest:   file.path("config","report.pptx")
#> → MB: including file
#> → MB:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → MB:   dest:   file.path("config","report.yaml")
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#> → MB: module checksum updated:a8d5d330f91feb7be769e08518458dd8
#> → MB: State initialized
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.R not found (skipping)
#> ! MB: User-defined model: /Users/jmh/projects/ruminate/github/ruminate/docs/reference/user_model.ctl not found (skipping)
#> Warning: incomplete final line found on '/Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/nlmixr2lib/modeldb/indirect_0cpt_transitEx.R'
#>  
#>  
#> ℹ parameter labels from comments are typically ignored in non-interactive mode
#> ℹ Need to run with the source intact to parse comments
#> → MB: model checksum updated: b83dbf6ef4c621c6799f7fb6b6504060
#> → MB: module checksum updated:21e568266ee52e343cd70987729d8e51