Fetches PKNCA Metadata

Compiles Metadata from PKNCA

Usage

NCA_fetch_PKNCA_meta()

Value

Dataframe containing PKCNA metadata for NCA parameters.

Examples

PKNCA_meta = NCA_fetch_PKNCA_meta()
utils::head(PKNCA_meta)
#> $parameters
#>                parameter       unit_type
#> 1                  start            time
#> 2                    end            time
#> 3                auclast             auc
#> 4                 aucall             auc
#> 5               aumclast            aumc
#> 6                aumcall            aumc
#> 7            aucint.last             auc
#> 8       aucint.last.dose             auc
#> 9             aucint.all             auc
#> 10       aucint.all.dose             auc
#> 11                    c0            conc
#> 12                  cmax            conc
#> 13                  cmin            conc
#> 14                  tmax            time
#> 15                 tlast            time
#> 16                tfirst            time
#> 17             clast.obs            conc
#> 18               cl.last       clearance
#> 19                cl.all       clearance
#> 20                     f        fraction
#> 21              mrt.last            time
#> 22           mrt.iv.last            time
#> 23              vss.last          volume
#> 24           vss.iv.last          volume
#> 25                   cav            conc
#> 26          cav.int.last            conc
#> 27           cav.int.all            conc
#> 28               ctrough            conc
#> 29                cstart            conc
#> 30                   ptr        fraction
#> 31                  tlag            time
#> 32              deg.fluc               %
#> 33                 swing               %
#> 34                  ceoi            conc
#> 35  aucabove.predose.all             auc
#> 36   aucabove.trough.all             auc
#> 37            count_conc           count
#> 38               totdose            dose
#> 39                    ae          amount
#> 40              clr.last renal_clearance
#> 41               clr.obs renal_clearance
#> 42              clr.pred renal_clearance
#> 43                    fe        fraction
#> 44        sparse_auclast             auc
#> 45         sparse_auc_se             auc
#> 46         sparse_auc_df           count
#> 47            time_above            time
#> 48             aucivlast             auc
#> 49              aucivall             auc
#> 50         aucivint.last             auc
#> 51          aucivint.all             auc
#> 52        aucivpbextlast               %
#> 53         aucivpbextall               %
#> 54    aucivpbextint.last               %
#> 55     aucivpbextint.all               %
#> 56             half.life            time
#> 57             r.squared        unitless
#> 58         adj.r.squared        unitless
#> 59              lambda.z    inverse_time
#> 60   lambda.z.time.first            time
#> 61     lambda.z.n.points           count
#> 62            clast.pred            conc
#> 63            span.ratio        fraction
#> 64        thalf.eff.last            time
#> 65     thalf.eff.iv.last            time
#> 66              kel.last    inverse_time
#> 67           kel.iv.last    inverse_time
#> 68            aucinf.obs             auc
#> 69           aucinf.pred             auc
#> 70           aumcinf.obs            aumc
#> 71          aumcinf.pred            aumc
#> 72        aucint.inf.obs             auc
#> 73   aucint.inf.obs.dose             auc
#> 74       aucint.inf.pred             auc
#> 75  aucint.inf.pred.dose             auc
#> 76          aucivinf.obs             auc
#> 77         aucivinf.pred             auc
#> 78     aucivpbextinf.obs               %
#> 79    aucivpbextinf.pred               %
#> 80           aucpext.obs               %
#> 81          aucpext.pred               %
#> 82                cl.obs       clearance
#> 83               cl.pred       clearance
#> 84               mrt.obs            time
#> 85              mrt.pred            time
#> 86            mrt.iv.obs            time
#> 87           mrt.iv.pred            time
#> 88            mrt.md.obs            time
#> 89           mrt.md.pred            time
#> 90                vz.obs          volume
#> 91               vz.pred          volume
#> 92               vss.obs          volume
#> 93              vss.pred          volume
#> 94            vss.iv.obs          volume
#> 95           vss.iv.pred          volume
#> 96            vss.md.obs          volume
#> 97           vss.md.pred          volume
#> 98       cav.int.inf.obs            conc
#> 99      cav.int.inf.pred            conc
#> 100        thalf.eff.obs            time
#> 101       thalf.eff.pred            time
#> 102     thalf.eff.iv.obs            time
#> 103    thalf.eff.iv.pred            time
#> 104              kel.obs    inverse_time
#> 105             kel.pred    inverse_time
#> 106           kel.iv.obs    inverse_time
#> 107          kel.iv.pred    inverse_time
#> 108           auclast.dn    auc_dosenorm
#> 109            aucall.dn    auc_dosenorm
#> 110        aucinf.obs.dn    auc_dosenorm
#> 111       aucinf.pred.dn    auc_dosenorm
#> 112          aumclast.dn   aumc_dosenorm
#> 113           aumcall.dn   aumc_dosenorm
#> 114       aumcinf.obs.dn   aumc_dosenorm
#> 115      aumcinf.pred.dn   aumc_dosenorm
#> 116              cmax.dn   conc_dosenorm
#> 117              cmin.dn   conc_dosenorm
#> 118         clast.obs.dn   conc_dosenorm
#> 119        clast.pred.dn   conc_dosenorm
#> 120               cav.dn   conc_dosenorm
#> 121           ctrough.dn   conc_dosenorm
#>                                                 pretty_name data_type
#> 1                                            Interval Start  interval
#> 2                                              Interval End  interval
#> 3                                                   AUClast  interval
#> 4                                                    AUCall  interval
#> 5                                                 AUMC,last  interval
#> 6                                                  AUMC,all  interval
#> 7                   AUCint (based on AUClast extrapolation)  interval
#> 8       AUCint (based on AUClast extrapolation, dose-aware)  interval
#> 9                    AUCint (based on AUCall extrapolation)  interval
#> 10       AUCint (based on AUCall extrapolation, dose-aware)  interval
#> 11                                                       C0  interval
#> 12                                                     Cmax  interval
#> 13                                                     Cmin  interval
#> 14                                                     Tmax  interval
#> 15                                                    Tlast  interval
#> 16                                                   Tfirst  interval
#> 17                                                    Clast  interval
#> 18                                    CL (based on AUClast)  interval
#> 19                                     CL (based on AUCall)  interval
#> 20                                          Bioavailability  interval
#> 21                                   MRT (based on AUClast)  interval
#> 22                    MRT (for IV dosing, based on AUClast)  interval
#> 23                                   Vss (based on AUClast)  interval
#> 24                    Vss (for IV dosing, based on AUClast)  interval
#> 25                                                      Cav  interval
#> 26                                                      Cav  interval
#> 27                                                      Cav  interval
#> 28                                                  Ctrough  interval
#> 29                                                   Cstart  interval
#> 30                                     Peak-to-trough ratio  interval
#> 31                                                     Tlag  interval
#> 32                                    Degree of fluctuation  interval
#> 33                                                    Swing  interval
#> 34                                                     Ceoi  interval
#> 35                                                AUC,above  interval
#> 36                                                AUC,above  interval
#> 37                                      Concentration count  interval
#> 38                                               Total dose  interval
#> 39                                          Amount excreted  interval
#> 40                           Renal clearance (from AUClast)  interval
#> 41                        Renal clearance (from AUCinf,obs)  interval
#> 42                       Renal clearance (from AUCinf,pred)  interval
#> 43                                        Fraction excreted  interval
#> 44                                           Sparse AUClast  interval
#> 45                            Sparse AUClast standard error  interval
#> 46                        Sparse AUClast degrees of freedom  interval
#> 47                                 Time above Concentration  interval
#> 48                                      AUClast (IV dosing)  interval
#> 49                                       AUCall (IV dosing)  interval
#> 50                                  AUCint,last (IV dosing)  interval
#> 51                                   AUCint,all (IV dosing)  interval
#> 52                               AUCbext (based on AUClast)  interval
#> 53                                AUCbext (based on AUCall)  interval
#> 54                           AUCbext (based on AUCint,last)  interval
#> 55                            AUCbext (based on AUCint,all)  interval
#> 56                                                Half-life  interval
#> 57                                                    $r^2$  interval
#> 58                                              $r^2_{adj}$  interval
#> 59                                             $\\lambda_z$  interval
#> 60                              First time for $\\lambda_z$  interval
#> 61                       Number of points used for lambda_z  interval
#> 62                                               Clast,pred  interval
#> 63                                               Span ratio  interval
#> 64                  Effective half-life (based on MRT,last)  interval
#> 65    Effective half-life (for IV dosing, based on MRTlast)  interval
#> 66                                   Kel (based on AUClast)  interval
#> 67                    Kel (for IV dosing, based on AUClast)  interval
#> 68                                               AUCinf,obs  interval
#> 69                                              AUCinf,pred  interval
#> 70                                             AUMC,inf,obs  interval
#> 71                                            AUMC,inf,pred  interval
#> 72               AUCint (based on AUCinf,obs extrapolation)  interval
#> 73   AUCint (based on AUCinf,obs extrapolation, dose-aware)  interval
#> 74              AUCint (based on AUCinf,pred extrapolation)  interval
#> 75  AUCint (based on AUCinf,pred extrapolation, dose-aware)  interval
#> 76                                   AUCinf,obs (IV dosing)  interval
#> 77                                  AUCinf,pred (IV dosing)  interval
#> 78                            AUCbext (based on AUCinf,obs)  interval
#> 79                           AUCbext (based on AUCinf,pred)  interval
#> 80                            AUCpext (based on AUCinf,obs)  interval
#> 81                           AUCpext (based on AUCinf,pred)  interval
#> 82                                 CL (based on AUCinf,obs)  interval
#> 83                                CL (based on AUCinf,pred)  interval
#> 84                                MRT (based on AUCinf,obs)  interval
#> 85                               MRT (based on AUCinf,pred)  interval
#> 86                 MRT (for IV dosing, based on AUCinf,obs)  interval
#> 87                MRT (for IV dosing, based on AUCinf,pred)  interval
#> 88           MRT (for multiple dosing, based on AUCinf,obs)  interval
#> 89          MRT (for multiple dosing, based on AUCinf,pred)  interval
#> 90                                 Vz (based on AUCinf,obs)  interval
#> 91                                Vz (based on AUCinf,pred)  interval
#> 92                                Vss (based on AUCinf,obs)  interval
#> 93                               Vss (based on AUCinf,pred)  interval
#> 94                 Vss (for IV dosing, based on AUCinf,obs)  interval
#> 95                Vss (for IV dosing, based on AUCinf,pred)  interval
#> 96              Vss (for multiple-dose, based on Clast,obs)  interval
#> 97             Vss (for multiple-dose, based on Clast,pred)  interval
#> 98                                                      Cav  interval
#> 99                                                      Cav  interval
#> 100                  Effective half-life (based on MRT,obs)  interval
#> 101                 Effective half-life (based on MRT,pred)  interval
#> 102   Effective half-life (for IV dosing, based on MRT,obs)  interval
#> 103  Effective half-life (for IV dosing, based on MRT,pred)  interval
#> 104                               Kel (based on AUCinf,obs)  interval
#> 105                              Kel (based on AUCinf,pred)  interval
#> 106                Kel (for IV dosing, based on AUCinf,obs)  interval
#> 107               Kel (for IV dosing, based on AUCinf,pred)  interval
#> 108                               AUClast (dose-normalized)  interval
#> 109                                AUCall (dose-normalized)  interval
#> 110                            AUCinf,obs (dose-normalized)  interval
#> 111                           AUCinf,pred (dose-normalized)  interval
#> 112                             AUMC,last (dose-normalized)  interval
#> 113                              AUMC,all (dose-normalized)  interval
#> 114                          AUMC,inf,obs (dose-normalized)  interval
#> 115                         AUMC,inf,pred (dose-normalized)  interval
#> 116                                  Cmax (dose-normalized)  interval
#> 117                                  Cmin (dose-normalized)  interval
#> 118                                 Clast (dose-normalized)  interval
#> 119                            Clast,pred (dose-normalized)  interval
#> 120                                   Cav (dose-normalized)  interval
#> 121                               Ctrough (dose-normalized)  interval
#>                                                                                                                                                                                                                                                                                                                                                              desc
#> 1                                                                                                                                                                                                                                                                                                                                   Starting time of the interval
#> 2                                                                                                                                                                                                                                                                                                              Ending time of the interval (potentially infinity)
#> 3                                                                                                                                                                                                                      The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 4                                                                                                     The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 5                                                                                                                                                                                                               The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification
#> 6                                                                                The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification plus the moment of the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 7                                                                                                                                                                                                                                  The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 8                                                                                                                                                                    The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 9                                                                                                                                                                           The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 10                                                                                                            The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 11                                                                                                                                                                                                                                                                                                                        Initial concentration after an IV bolus
#> 12                                                                                                                                                                                                                                                                                                                                 Maximum observed concentration
#> 13                                                                                                                                                                                                                                                                                                                                 Minimum observed concentration
#> 14                                                                                                                                                                                                                                                                                                                     Time of the maximum observed concentration
#> 15                                                                                                                                                                                                                                                                                      Time of the last concentration observed above the limit of quantification
#> 16                                                                                                                                                                                                                                                                                              Time of the first concentration above the limit of quantification
#> 17                                                                                                                                                                                                                                                                                              The last concentration observed above the limit of quantification
#> 18                                                                                                                                                                                                                                                                                                       Clearance or observed oral clearance calculated to Clast
#> 19                                                                                                                                                                                                                                                                                                    Clearance or observed oral clearance calculated with AUCall
#> 20                                                                                                                                                                                                                                                                                                                    Bioavailability or relative bioavailability
#> 21                                                                                                                                                                                                                                                                                       The mean residence time to the last observed concentration above the LOQ
#> 22                                                                                                                                                                                                                                                        The mean residence time to the last observed concentration above the LOQ correcting for dosing duration
#> 23                                                                                                                                                                                                                                                                                              The steady-state volume of distribution calculating through Tlast
#> 24                                                                                                                                                                                                                                                                    The steady-state volume of distribution with intravenous infusion calculating through Tlast
#> 25                                                                                                                                                                                                                                                                                         The average concentration during an interval (calculated with AUClast)
#> 26                                                                                                                                                                                                                                                                                     The average concentration during an interval (calculated with AUCint.last)
#> 27                                                                                                                                                                                                                                                                                      The average concentration during an interval (calculated with AUCint.all)
#> 28                                                                                                                                                                                                                                                                                                                     The trough (end of interval) concentration
#> 29                                                                                                                                                                                                                                                                                                                                      The predose concentration
#> 30                                                                                                                                                                                                                                                                                                                                Peak-to-Trough ratio (fraction)
#> 31                                                                                                                                                                                                                                                                                                                                                       Lag time
#> 32                                                                                                                                                                                                                                                                                                                                          Degree of fluctuation
#> 33                                                                                                                                                                                                                                                                                                                                         Swing relative to Cmin
#> 34                                                                                                                                                                                                                                                                                                                           Concentration at the end of infusion
#> 35  The area under the concentration time the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification, with a concentration subtracted from all concentrations and values below zero after subtraction set to zero
#> 36  The area under the concentration time the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification, with a concentration subtracted from all concentrations and values below zero after subtraction set to zero
#> 37                                                                                                                                                                                                                                                                                                             Number of non-missing concentrations for a subject
#> 38                                                                                                                                                                                                                                                                                                                     Total dose administered during an interval
#> 39                                                                                                                                                                                                                                                                                                            The amount excreted (typically into urine or feces)
#> 40                                                                                                                                                                                                                                                                                                                   The renal clearance calculated using AUClast
#> 41                                                                                                                                                                                                                                                                                                                The renal clearance calculated using AUCinf,obs
#> 42                                                                                                                                                                                                                                                                                                               The renal clearance calculated using AUCinf,pred
#> 43                                                                                                                                                                                                                                                                                                                              The fraction of the dose excreted
#> 44                                                                                                                                                                                             For sparse PK sampling, the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 45                                                                                                                                                                       For sparse PK sampling, the standard error of the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 46                                                                                                                                                    For sparse PK sampling, the standard error degrees of freedom of the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 47                                                                                                                                                                                                                                                                                                                               Time above a given concentration
#> 48                                                                                                                                                                                                                                                                    The AUClast calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 49                                                                                                                                                                                                                                                                     The AUCall calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 50                                                                                                                                                                                                                                                                The AUCint,last calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 51                                                                                                                                                                                                                                                                 The AUCint,all calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 52                                                                                                                                                                                                                                                                                         The back-extrapolation percent for intravenous dosing based on AUClast
#> 53                                                                                                                                                                                                                                                                                          The back-extrapolation percent for intravenous dosing based on AUCall
#> 54                                                                                                                                                                                                                                                                                     The back-extrapolation percent for intravenous dosing based on AUCint,last
#> 55                                                                                                                                                                                                                                                                                      The back-extrapolation percent for intravenous dosing based on AUCint,all
#> 56                                                                                                                                                                                                                                                                                                                                       The (terminal) half-life
#> 57                                                                                                                                                                                                                                                                                                                     The r^2 value of the half-life calculation
#> 58                                                                                                                                                                                                                                                                                                            The adjusted r^2 value of the half-life calculation
#> 59                                                                                                                                                                                                                                                                                                                 The elimination rate of the terminal half-life
#> 60                                                                                                                                                                                                                                                                                                     The first time point used for the calculation of half-life
#> 61                                                                                                                                                                                                                                                                                                     The number of points used for the calculation of half-life
#> 62                                                                                                                                                                                                                                                                                                       The concentration at Tlast as predicted by the half-life
#> 63                                                                                                                                                                                                                                                                                      The ratio of the half-life to the duration used for half-life calculation
#> 64                                                                                                                                                                                                                                                                                                       The effective half-life (as determined from the MRTlast)
#> 65                                                                                                                                                                                                                                                                                           The effective half-life (as determined from the intravenous MRTlast)
#> 66                                                                                                                                                                                                                                                                                                    Elimination rate (as calculated from the MRT using AUClast)
#> 67                                                                                                                                                                                                                                                                                                  Elimination rate (as calculated from the intravenous MRTlast)
#> 68                                                                                                                                                                                                              The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 69                                                                                                                                                                                                             The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast
#> 70                                                                                                                                                                                                       The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 71                                                                                                                                                                                                      The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast
#> 72                                                                                                                                                                                                                                 The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 73                                                                                                                                                                   The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 74                                                                                                                                                                          The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 75                                                                                                            The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 76                                                                                                                                                                                                                                                                 The AUCinf,obs calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 77                                                                                                                                                                                                                                                                           The  calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 78                                                                                                                                                                                                                                                                                      The back-extrapolation percent for intravenous dosing based on AUCinf,obs
#> 79                                                                                                                                                                                                                                                                                     The back-extrapolation percent for intravenous dosing based on AUCinf,pred
#> 80                                                                                                                                                                                                                                                                      Percent of the AUCinf that is extrapolated after Tlast calculated from the observed Clast
#> 81                                                                                                                                                                                                                                                                     Percent of the AUCinf that is extrapolated after Tlast calculated from the predicted Clast
#> 82                                                                                                                                                                                                                                                                                            Clearance or observed oral clearance calculated with observed Clast
#> 83                                                                                                                                                                                                                                                                                           Clearance or observed oral clearance calculated with predicted Clast
#> 84                                                                                                                                                                                                                                                                                                       The mean residence time to infinity using observed Clast
#> 85                                                                                                                                                                                                                                                                                                      The mean residence time to infinity using predicted Clast
#> 86                                                                                                                                                                                                                                                                        The mean residence time to infinity using observed Clast correcting for dosing duration
#> 87                                                                                                                                                                                                                                                                       The mean residence time to infinity using predicted Clast correcting for dosing duration
#> 88                                                                                                                                                                                                                                                                       The mean residence time with multiple dosing and nonlinear kinetics using observed Clast
#> 89                                                                                                                                                                                                                                                                      The mean residence time with multiple dosing and nonlinear kinetics using predicted Clast
#> 90                                                                                                                                                                                                                                                                                                       The terminal volume of distribution using observed Clast
#> 91                                                                                                                                                                                                                                                                                                      The terminal volume of distribution using predicted Clast
#> 92                                                                                                                                                                                                                                                                                                   The steady-state volume of distribution using observed Clast
#> 93                                                                                                                                                                                                                                                                                                  The steady-state volume of distribution using predicted Clast
#> 94                                                                                                                                                                                                                                                                         The steady-state volume of distribution with intravenous infusion using observed Clast
#> 95                                                                                                                                                                                                                                                                        The steady-state volume of distribution with intravenous infusion using predicted Clast
#> 96                                                                                                                                                                                                                                                                  The steady-state volume of distribution for nonlinear multiple-dose data using observed Clast
#> 97                                                                                                                                                                                                                                                                 The steady-state volume of distribution for nonlinear multiple-dose data using predicted Clast
#> 98                                                                                                                                                                                                                                                                                  The average concentration during an interval (calculated with AUCint.inf.obs)
#> 99                                                                                                                                                                                                                                                                                 The average concentration during an interval (calculated with AUCint.inf.pred)
#> 100                                                                                                                                                                                                                                                                                                       The effective half-life (as determined from the MRTobs)
#> 101                                                                                                                                                                                                                                                                                                      The effective half-life (as determined from the MRTpred)
#> 102                                                                                                                                                                                                                                                                                           The effective half-life (as determined from the intravenous MRTobs)
#> 103                                                                                                                                                                                                                                                                                          The effective half-life (as determined from the intravenous MRTpred)
#> 104                                                                                                                                                                                                                                                                                             Elimination rate (as calculated from the MRT with observed Clast)
#> 105                                                                                                                                                                                                                                                                                            Elimination rate (as calculated from the MRT with predicted Clast)
#> 106                                                                                                                                                                                                                                                                                                  Elimination rate (as calculated from the intravenous MRTobs)
#> 107                                                                                                                                                                                                                                                                                                 Elimination rate (as calculated from the intravenous MRTpred)
#> 108                                                                                                                                                                                                                                                                                                                                       Dose normalized auclast
#> 109                                                                                                                                                                                                                                                                                                                                        Dose normalized aucall
#> 110                                                                                                                                                                                                                                                                                                                                    Dose normalized aucinf.obs
#> 111                                                                                                                                                                                                                                                                                                                                   Dose normalized aucinf.pred
#> 112                                                                                                                                                                                                                                                                                                                                      Dose normalized aumclast
#> 113                                                                                                                                                                                                                                                                                                                                       Dose normalized aumcall
#> 114                                                                                                                                                                                                                                                                                                                                   Dose normalized aumcinf.obs
#> 115                                                                                                                                                                                                                                                                                                                                  Dose normalized aumcinf.pred
#> 116                                                                                                                                                                                                                                                                                                                                          Dose normalized cmax
#> 117                                                                                                                                                                                                                                                                                                                                          Dose normalized cmin
#> 118                                                                                                                                                                                                                                                                                                                                     Dose normalized clast.obs
#> 119                                                                                                                                                                                                                                                                                                                                    Dose normalized clast.pred
#> 120                                                                                                                                                                                                                                                                                                                                           Dose normalized cav
#> 121                                                                                                                                                                                                                                                                                                                                       Dose normalized ctrough
#>