Fetches PKNCA Metadata

Compiles Metadata from PKNCA

Usage

NCA_fetch_PKNCA_meta()

Value

Dataframe containing PKCNA metadata for NCA parameters.

Examples

PKNCA_meta = NCA_fetch_PKNCA_meta()
utils::head(PKNCA_meta)
#> $parameters
#>                parameter       unit_type
#> 1                  start            time
#> 2                    end            time
#> 3                auclast             auc
#> 4                 aucall             auc
#> 5               aumclast            aumc
#> 6                aumcall            aumc
#> 7            aucint.last             auc
#> 8       aucint.last.dose             auc
#> 9             aucint.all             auc
#> 10       aucint.all.dose             auc
#> 11                    c0            conc
#> 12                  cmax            conc
#> 13                  cmin            conc
#> 14                  tmax            time
#> 15                 tlast            time
#> 16                tfirst            time
#> 17             clast.obs            conc
#> 18               cl.last       clearance
#> 19                cl.all       clearance
#> 20                     f        fraction
#> 21              mrt.last            time
#> 22           mrt.iv.last            time
#> 23              vss.last          volume
#> 24           vss.iv.last          volume
#> 25                   cav            conc
#> 26               ctrough            conc
#> 27                cstart            conc
#> 28                   ptr        fraction
#> 29                  tlag            time
#> 30              deg.fluc               %
#> 31                 swing               %
#> 32                  ceoi            conc
#> 33  aucabove.predose.all             auc
#> 34   aucabove.trough.all             auc
#> 35                    ae          amount
#> 36              clr.last renal_clearance
#> 37               clr.obs renal_clearance
#> 38              clr.pred renal_clearance
#> 39                    fe        fraction
#> 40        sparse_auclast             auc
#> 41            time_above            time
#> 42             aucivlast             auc
#> 43              aucivall             auc
#> 44         aucivint.last             auc
#> 45          aucivint.all             auc
#> 46        aucivpbextlast               %
#> 47         aucivpbextall               %
#> 48    aucivpbextint.last               %
#> 49     aucivpbextint.all               %
#> 50             half.life            time
#> 51             r.squared        unitless
#> 52         adj.r.squared        unitless
#> 53              lambda.z    inverse_time
#> 54   lambda.z.time.first            time
#> 55     lambda.z.n.points           count
#> 56            clast.pred            conc
#> 57            span.ratio        fraction
#> 58        thalf.eff.last            time
#> 59     thalf.eff.iv.last            time
#> 60              kel.last    inverse_time
#> 61           kel.iv.last    inverse_time
#> 62            aucinf.obs             auc
#> 63           aucinf.pred             auc
#> 64           aumcinf.obs            aumc
#> 65          aumcinf.pred            aumc
#> 66        aucint.inf.obs             auc
#> 67   aucint.inf.obs.dose             auc
#> 68       aucint.inf.pred             auc
#> 69  aucint.inf.pred.dose             auc
#> 70          aucivinf.obs             auc
#> 71         aucivinf.pred             auc
#> 72     aucivpbextinf.obs               %
#> 73    aucivpbextinf.pred               %
#> 74           aucpext.obs               %
#> 75          aucpext.pred               %
#> 76                cl.obs       clearance
#> 77               cl.pred       clearance
#> 78               mrt.obs            time
#> 79              mrt.pred            time
#> 80            mrt.iv.obs            time
#> 81           mrt.iv.pred            time
#> 82            mrt.md.obs            time
#> 83           mrt.md.pred            time
#> 84                vz.obs          volume
#> 85               vz.pred          volume
#> 86               vss.obs          volume
#> 87              vss.pred          volume
#> 88            vss.iv.obs          volume
#> 89           vss.iv.pred          volume
#> 90            vss.md.obs          volume
#> 91           vss.md.pred          volume
#> 92                vd.obs          volume
#> 93               vd.pred          volume
#> 94         thalf.eff.obs            time
#> 95        thalf.eff.pred            time
#> 96      thalf.eff.iv.obs            time
#> 97     thalf.eff.iv.pred            time
#> 98               kel.obs    inverse_time
#> 99              kel.pred    inverse_time
#> 100           kel.iv.obs    inverse_time
#> 101          kel.iv.pred    inverse_time
#> 102           auclast.dn    auc_dosenorm
#> 103            aucall.dn    auc_dosenorm
#> 104        aucinf.obs.dn    auc_dosenorm
#> 105       aucinf.pred.dn    auc_dosenorm
#> 106          aumclast.dn   aumc_dosenorm
#> 107           aumcall.dn   aumc_dosenorm
#> 108       aumcinf.obs.dn   aumc_dosenorm
#> 109      aumcinf.pred.dn   aumc_dosenorm
#> 110              cmax.dn   conc_dosenorm
#> 111              cmin.dn   conc_dosenorm
#> 112         clast.obs.dn   conc_dosenorm
#> 113        clast.pred.dn   conc_dosenorm
#> 114               cav.dn   conc_dosenorm
#> 115           ctrough.dn   conc_dosenorm
#>                                                 pretty_name data_type
#> 1                                            Interval Start  interval
#> 2                                              Interval End  interval
#> 3                                                   AUClast  interval
#> 4                                                    AUCall  interval
#> 5                                                 AUMC,last  interval
#> 6                                                  AUMC,all  interval
#> 7                   AUCint (based on AUClast extrapolation)  interval
#> 8       AUCint (based on AUClast extrapolation, dose-aware)  interval
#> 9                    AUCint (based on AUCall extrapolation)  interval
#> 10       AUCint (based on AUCall extrapolation, dose-aware)  interval
#> 11                                                       C0  interval
#> 12                                                     Cmax  interval
#> 13                                                     Cmin  interval
#> 14                                                     Tmax  interval
#> 15                                                    Tlast  interval
#> 16                                                   Tfirst  interval
#> 17                                                    Clast  interval
#> 18                                    CL (based on AUClast)  interval
#> 19                                     CL (based on AUCall)  interval
#> 20                                          Bioavailability  interval
#> 21                                   MRT (based on AUClast)  interval
#> 22                    MRT (for IV dosing, based on AUClast)  interval
#> 23                                   Vss (based on AUClast)  interval
#> 24                    Vss (for IV dosing, based on AUClast)  interval
#> 25                                                      Cav  interval
#> 26                                                  Ctrough  interval
#> 27                                                   Cstart  interval
#> 28                                     Peak-to-trough ratio  interval
#> 29                                                     Tlag  interval
#> 30                                    Degree of fluctuation  interval
#> 31                                                    Swing  interval
#> 32                                                     Ceoi  interval
#> 33                                                AUC,above  interval
#> 34                                                AUC,above  interval
#> 35                                          Amount excreted  interval
#> 36                           Renal clearance (from AUClast)  interval
#> 37                        Renal clearance (from AUCinf,obs)  interval
#> 38                       Renal clearance (from AUCinf,pred)  interval
#> 39                                        Fraction excreted  interval
#> 40                                           Sparse AUClast  interval
#> 41                                 Time above Concentration  interval
#> 42                                      AUClast (IV dosing)  interval
#> 43                                       AUCall (IV dosing)  interval
#> 44                                  AUCint,last (IV dosing)  interval
#> 45                                   AUCint,all (IV dosing)  interval
#> 46                               AUCbext (based on AUClast)  interval
#> 47                                AUCbext (based on AUCall)  interval
#> 48                           AUCbext (based on AUCint,last)  interval
#> 49                            AUCbext (based on AUCint,all)  interval
#> 50                                                Half-life  interval
#> 51                                                    $r^2$  interval
#> 52                                              $r^2_{adj}$  interval
#> 53                                             $\\lambda_z$  interval
#> 54                              First time for $\\lambda_z$  interval
#> 55                       Number of points used for lambda_z  interval
#> 56                                               Clast,pred  interval
#> 57                                               Span ratio  interval
#> 58                  Effective half-life (based on MRT,last)  interval
#> 59    Effective half-life (for IV dosing, based on MRTlast)  interval
#> 60                                   Kel (based on AUClast)  interval
#> 61                    Kel (for IV dosing, based on AUClast)  interval
#> 62                                               AUCinf,obs  interval
#> 63                                              AUCinf,pred  interval
#> 64                                             AUMC,inf,obs  interval
#> 65                                            AUMC,inf,pred  interval
#> 66               AUCint (based on AUCinf,obs extrapolation)  interval
#> 67   AUCint (based on AUCinf,obs extrapolation, dose-aware)  interval
#> 68              AUCint (based on AUCinf,pred extrapolation)  interval
#> 69  AUCint (based on AUCinf,pred extrapolation, dose-aware)  interval
#> 70                                   AUCinf,obs (IV dosing)  interval
#> 71                                  AUCinf,pred (IV dosing)  interval
#> 72                            AUCbext (based on AUCinf,obs)  interval
#> 73                           AUCbext (based on AUCinf,pred)  interval
#> 74                            AUCpext (based on AUCinf,obs)  interval
#> 75                           AUCpext (based on AUCinf,pred)  interval
#> 76                                 CL (based on AUCinf,obs)  interval
#> 77                                CL (based on AUCinf,pred)  interval
#> 78                                MRT (based on AUCinf,obs)  interval
#> 79                               MRT (based on AUCinf,pred)  interval
#> 80                 MRT (for IV dosing, based on AUCinf,obs)  interval
#> 81                MRT (for IV dosing, based on AUCinf,pred)  interval
#> 82           MRT (for multiple dosing, based on AUCinf,obs)  interval
#> 83          MRT (for multiple dosing, based on AUCinf,pred)  interval
#> 84                                 Vz (based on AUCinf,obs)  interval
#> 85                                Vz (based on AUCinf,pred)  interval
#> 86                                Vss (based on AUCinf,obs)  interval
#> 87                               Vss (based on AUCinf,pred)  interval
#> 88                 Vss (for IV dosing, based on AUCinf,obs)  interval
#> 89                Vss (for IV dosing, based on AUCinf,pred)  interval
#> 90              Vss (for multiple-dose, based on Clast,obs)  interval
#> 91             Vss (for multiple-dose, based on Clast,pred)  interval
#> 92                                  Vd (based on Clast,obs)  interval
#> 93                                 Vd (based on Clast,pred)  interval
#> 94                   Effective half-life (based on MRT,obs)  interval
#> 95                  Effective half-life (based on MRT,pred)  interval
#> 96    Effective half-life (for IV dosing, based on MRT,obs)  interval
#> 97   Effective half-life (for IV dosing, based on MRT,pred)  interval
#> 98                                Kel (based on AUCinf,obs)  interval
#> 99                               Kel (based on AUCinf,pred)  interval
#> 100                Kel (for IV dosing, based on AUCinf,obs)  interval
#> 101               Kel (for IV dosing, based on AUCinf,pred)  interval
#> 102                               AUClast (dose-normalized)  interval
#> 103                                AUCall (dose-normalized)  interval
#> 104                            AUCinf,obs (dose-normalized)  interval
#> 105                           AUCinf,pred (dose-normalized)  interval
#> 106                             AUMC,last (dose-normalized)  interval
#> 107                              AUMC,all (dose-normalized)  interval
#> 108                          AUMC,inf,obs (dose-normalized)  interval
#> 109                         AUMC,inf,pred (dose-normalized)  interval
#> 110                                  Cmax (dose-normalized)  interval
#> 111                                  Cmin (dose-normalized)  interval
#> 112                                 Clast (dose-normalized)  interval
#> 113                            Clast,pred (dose-normalized)  interval
#> 114                                   Cav (dose-normalized)  interval
#> 115                               Ctrough (dose-normalized)  interval
#>                                                                                                                                                                                                                                                                                                                                                              desc
#> 1                                                                                                                                                                                                                                                                                                                                   Starting time of the interval
#> 2                                                                                                                                                                                                                                                                                                              Ending time of the interval (potentially infinity)
#> 3                                                                                                                                                                                                                      The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 4                                                                                                     The area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 5                                                                                                                                                                                                               The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification
#> 6                                                                                The area under the concentration time moment curve from the beginning of the interval to the last concentration above the limit of quantification plus the moment of the triangle from that last concentration to 0 at the first concentration below the limit of quantification
#> 7                                                                                                                                                                                                                                  The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 8                                                                                                                                                                    The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 9                                                                                                                                                                           The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 10                                                                                                            The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 11                                                                                                                                                                                                                                                                                                                        Initial concentration after an IV bolus
#> 12                                                                                                                                                                                                                                                                                                                                 Maximum observed concentration
#> 13                                                                                                                                                                                                                                                                                                                                 Minimum observed concentration
#> 14                                                                                                                                                                                                                                                                                                                     Time of the maximum observed concentration
#> 15                                                                                                                                                                                                                                                                                      Time of the last concentration observed above the limit of quantification
#> 16                                                                                                                                                                                                                                                                                              Time of the first concentration above the limit of quantification
#> 17                                                                                                                                                                                                                                                                                              The last concentration observed above the limit of quantification
#> 18                                                                                                                                                                                                                                                                                                       Clearance or observed oral clearance calculated to Clast
#> 19                                                                                                                                                                                                                                                                                                    Clearance or observed oral clearance calculated with AUCall
#> 20                                                                                                                                                                                                                                                                                                                    Bioavailability or relative bioavailability
#> 21                                                                                                                                                                                                                                                                                       The mean residence time to the last observed concentration above the LOQ
#> 22                                                                                                                                                                                                                                                        The mean residence time to the last observed concentration above the LOQ correcting for dosing duration
#> 23                                                                                                                                                                                                                                                                                              The steady-state volume of distribution calculating through Tlast
#> 24                                                                                                                                                                                                                                                                    The steady-state volume of distribution with intravenous infusion calculating through Tlast
#> 25                                                                                                                                                                                                                                                                                                                   The average concentration during an interval
#> 26                                                                                                                                                                                                                                                                                                                     The trough (end of interval) concentration
#> 27                                                                                                                                                                                                                                                                                                                                      The predose concentration
#> 28                                                                                                                                                                                                                                                                                                                                Peak-to-Trough ratio (fraction)
#> 29                                                                                                                                                                                                                                                                                                                                                       Lag time
#> 30                                                                                                                                                                                                                                                                                                                                          Degree of fluctuation
#> 31                                                                                                                                                                                                                                                                                                                                         Swing relative to Cmin
#> 32                                                                                                                                                                                                                                                                                                                           Concentration at the end of infusion
#> 33  The area under the concentration time the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification, with a concentration subtracted from all concentrations and values below zero after subtraction set to zero
#> 34  The area under the concentration time the beginning of the interval to the last concentration above the limit of quantification plus the triangle from that last concentration to 0 at the first concentration below the limit of quantification, with a concentration subtracted from all concentrations and values below zero after subtraction set to zero
#> 35                                                                                                                                                                                                                                                                                                            The amount excreted (typically into urine or feces)
#> 36                                                                                                                                                                                                                                                                                                                   The renal clearance calculated using AUClast
#> 37                                                                                                                                                                                                                                                                                                                The renal clearance calculated using AUCinf,obs
#> 38                                                                                                                                                                                                                                                                                                               The renal clearance calculated using AUCinf,pred
#> 39                                                                                                                                                                                                                                                                                                                              The fraction of the dose excreted
#> 40                                                                                                                                                                                             For sparse PK sampling, the area under the concentration time curve from the beginning of the interval to the last concentration above the limit of quantification
#> 41                                                                                                                                                                                                                                                                                                                               Time above a given concentration
#> 42                                                                                                                                                                                                                                                                    The AUClast calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 43                                                                                                                                                                                                                                                                     The AUCall calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 44                                                                                                                                                                                                                                                                The AUCint,last calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 45                                                                                                                                                                                                                                                                 The AUCint,all calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 46                                                                                                                                                                                                                                                                                         The back-extrapolation percent for intravenous dosing based on AUClast
#> 47                                                                                                                                                                                                                                                                                          The back-extrapolation percent for intravenous dosing based on AUCall
#> 48                                                                                                                                                                                                                                                                                     The back-extrapolation percent for intravenous dosing based on AUCint,last
#> 49                                                                                                                                                                                                                                                                                      The back-extrapolation percent for intravenous dosing based on AUCint,all
#> 50                                                                                                                                                                                                                                                                                                                                       The (terminal) half-life
#> 51                                                                                                                                                                                                                                                                                                                     The r^2 value of the half-life calculation
#> 52                                                                                                                                                                                                                                                                                                            The adjusted r^2 value of the half-life calculation
#> 53                                                                                                                                                                                                                                                                                                                 The elimination rate of the terminal half-life
#> 54                                                                                                                                                                                                                                                                                                     The first time point used for the calculation of half-life
#> 55                                                                                                                                                                                                                                                                                                     The number of points used for the calculation of half-life
#> 56                                                                                                                                                                                                                                                                                                       The concentration at Tlast as predicted by the half-life
#> 57                                                                                                                                                                                                                                                                                      The ratio of the half-life to the duration used for half-life calculation
#> 58                                                                                                                                                                                                                                                                                                       The effective half-life (as determined from the MRTlast)
#> 59                                                                                                                                                                                                                                                                                           The effective half-life (as determined from the intravenous MRTlast)
#> 60                                                                                                                                                                                                                                                                                                    Elimination rate (as calculated from the MRT using AUClast)
#> 61                                                                                                                                                                                                                                                                                                  Elimination rate (as calculated from the intravenous MRTlast)
#> 62                                                                                                                                                                                                              The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 63                                                                                                                                                                                                             The area under the concentration time curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast
#> 64                                                                                                                                                                                                       The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the observed Clast
#> 65                                                                                                                                                                                                      The area under the concentration time moment curve from the beginning of the interval to infinity with extrapolation to infinity from the predicted Clast
#> 66                                                                                                                                                                                                                                 The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast)
#> 67                                                                                                                                                                   The area under the concentration time curve in the interval extrapolating from Tlast to infinity with zeros (matching AUClast) with dose-aware interpolation/extrapolation of concentrations
#> 68                                                                                                                                                                          The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall)
#> 69                                                                                                            The area under the concentration time curve in the interval extrapolating from Tlast to infinity with the triangle from Tlast to the next point and zero thereafter (matching AUCall) with dose-aware interpolation/extrapolation of concentrations
#> 70                                                                                                                                                                                                                                                                 The AUCinf,obs calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 71                                                                                                                                                                                                                                                                           The  calculated with back-extrapolation for intravenous dosing using extrapolated C0
#> 72                                                                                                                                                                                                                                                                                      The back-extrapolation percent for intravenous dosing based on AUCinf,obs
#> 73                                                                                                                                                                                                                                                                                     The back-extrapolation percent for intravenous dosing based on AUCinf,pred
#> 74                                                                                                                                                                                                                                                                      Percent of the AUCinf that is extrapolated after Tlast calculated from the observed Clast
#> 75                                                                                                                                                                                                                                                                     Percent of the AUCinf that is extrapolated after Tlast calculated from the predicted Clast
#> 76                                                                                                                                                                                                                                                                                            Clearance or observed oral clearance calculated with observed Clast
#> 77                                                                                                                                                                                                                                                                                           Clearance or observed oral clearance calculated with predicted Clast
#> 78                                                                                                                                                                                                                                                                                                       The mean residence time to infinity using observed Clast
#> 79                                                                                                                                                                                                                                                                                                      The mean residence time to infinity using predicted Clast
#> 80                                                                                                                                                                                                                                                                        The mean residence time to infinity using observed Clast correcting for dosing duration
#> 81                                                                                                                                                                                                                                                                       The mean residence time to infinity using predicted Clast correcting for dosing duration
#> 82                                                                                                                                                                                                                                                                       The mean residence time with multiple dosing and nonlinear kinetics using observed Clast
#> 83                                                                                                                                                                                                                                                                      The mean residence time with multiple dosing and nonlinear kinetics using predicted Clast
#> 84                                                                                                                                                                                                                                                                                                       The terminal volume of distribution using observed Clast
#> 85                                                                                                                                                                                                                                                                                                      The terminal volume of distribution using predicted Clast
#> 86                                                                                                                                                                                                                                                                                                   The steady-state volume of distribution using observed Clast
#> 87                                                                                                                                                                                                                                                                                                  The steady-state volume of distribution using predicted Clast
#> 88                                                                                                                                                                                                                                                                         The steady-state volume of distribution with intravenous infusion using observed Clast
#> 89                                                                                                                                                                                                                                                                        The steady-state volume of distribution with intravenous infusion using predicted Clast
#> 90                                                                                                                                                                                                                                                                  The steady-state volume of distribution for nonlinear multiple-dose data using observed Clast
#> 91                                                                                                                                                                                                                                                                 The steady-state volume of distribution for nonlinear multiple-dose data using predicted Clast
#> 92                                                                                                                                                                                                                                                                                        Apparent observed volume of distribution calculated with observed Clast
#> 93                                                                                                                                                                                                                                                                                       Apparent observed volume of distribution calculated with predicted Clast
#> 94                                                                                                                                                                                                                                                                                                        The effective half-life (as determined from the MRTobs)
#> 95                                                                                                                                                                                                                                                                                                       The effective half-life (as determined from the MRTpred)
#> 96                                                                                                                                                                                                                                                                                            The effective half-life (as determined from the intravenous MRTobs)
#> 97                                                                                                                                                                                                                                                                                           The effective half-life (as determined from the intravenous MRTpred)
#> 98                                                                                                                                                                                                                                                                                              Elimination rate (as calculated from the MRT with observed Clast)
#> 99                                                                                                                                                                                                                                                                                             Elimination rate (as calculated from the MRT with predicted Clast)
#> 100                                                                                                                                                                                                                                                                                                  Elimination rate (as calculated from the intravenous MRTobs)
#> 101                                                                                                                                                                                                                                                                                                 Elimination rate (as calculated from the intravenous MRTpred)
#> 102                                                                                                                                                                                                                                                                                                                                       Dose normalized auclast
#> 103                                                                                                                                                                                                                                                                                                                                        Dose normalized aucall
#> 104                                                                                                                                                                                                                                                                                                                                    Dose normalized aucinf.obs
#> 105                                                                                                                                                                                                                                                                                                                                   Dose normalized aucinf.pred
#> 106                                                                                                                                                                                                                                                                                                                                      Dose normalized aumclast
#> 107                                                                                                                                                                                                                                                                                                                                       Dose normalized aumcall
#> 108                                                                                                                                                                                                                                                                                                                                   Dose normalized aumcinf.obs
#> 109                                                                                                                                                                                                                                                                                                                                  Dose normalized aumcinf.pred
#> 110                                                                                                                                                                                                                                                                                                                                          Dose normalized cmax
#> 111                                                                                                                                                                                                                                                                                                                                          Dose normalized cmin
#> 112                                                                                                                                                                                                                                                                                                                                     Dose normalized clast.obs
#> 113                                                                                                                                                                                                                                                                                                                                    Dose normalized clast.pred
#> 114                                                                                                                                                                                                                                                                                                                                           Dose normalized cav
#> 115                                                                                                                                                                                                                                                                                                                                       Dose normalized ctrough
#>