Fetch ruminate State

Merges default app options with the changes made in the UI

Usage

NCA_fetch_state(
  id,
  input,
  session,
  FM_yaml_file,
  MOD_yaml_file,
  react_state,
  react_internal = NULL
)

Arguments

id

Shiny module ID

input

Shiny input variable

session

Shiny session variable

FM_yaml_file

App configuration file with FM as main section.

MOD_yaml_file

Module configuration file with MC as main section.

react_state

Variable used within server to handle non-standard reactions (e.g. reacting to plotly)

react_internal

Variable passed to server to allow reaction outside of module (NULL)

Value

list containing the current state of the app including default values from the yaml file as well as any changes made by the user. The list has the following structure:

yaml: Full contents of the supplied yaml file.

MC: Module components of the yaml file.

NCA:

ana_cntr: Analysis counter.

anas: List of analyses: Each analysis has the following structure:

ana_dsview: Dataset view/ID (name from DSV) selected as a data source for this analysis.

ana_scenario: Analysis scenario selected in the UI

checksum: checksum of the analysis (used to detect changes in the analysis).

code: Code to generate analysis from start to finish or error messages if code generation/analysis failed.

code_components: List containing the different components from code

col_conc: Column from ana_dsview containing the concentration data.

col_dose: Column from ana_dsview containing the dose amount.

col_dur: Column from ana_dsview containing the infusion duration or N/A if unused.

col_group: Columns from ana_dsview containing other grouping variables.

col_id: Column from ana_dsview containing the subject IDs.

col_ntime: Column from ana_dsview containing the nominal time values

col_route: Column from ana_dsview containing the dosing route.

col_time: Column from ana_dsview containing the time values.

id: Character id (ana_idx).

idx: Numeric id (1).

include_units: Boolean variable indicating in units should included in the analysis.

interval_range: Vector with the first element representing he beginning of the interval and the second element containing the end of the interval.

intervals: List of the intervals to include.

isgood: Current status of the analysis.

key: Analysis key acts as a title/caption (user editable)

msgs: Messages generated when checking configuration and analysis options.

nca_config: List of NCA configuration options for this analysis.

nca_object_name: Prefix for NCA objects associated with this analyis.

nca_parameters: NCA parameters selected for calculation in the UI.

notes: Analysis notes (user editable)

objs: List of names and values for objects created with generated code.

sampling: Sampling method either "sparse" or "serial"

units_amt: Amount units.

units_conc: Concentration units.

units_dose: Dosing units.

units_time: Time units.

current_ana: Currently selected analysis (list name element from anas).

DSV: Available data source views (see FM_fetch_ds)

checksum: This is an MD5 sum of the module (checksum of the analysis checksums).

nca_config: List of PKNCA configuration options for this analysis.

nca_parameters: List with two elements

choices: List consisting of "Common Parameters" and "Other" (used for grouping in the UI). Each of these is a list of text parameter names with a value of the PKNCA parameter name.

summary: Summary table with the following columns:

parameter: PKNCA Paramter name.

text: Name used in text output.

md: Name used markdown output.

latex: Name used in latex output.

description: Verbose textual description of the parameter.

ui: Current value of form elements in the UI.

ui_ana_map: Map between UI element names and analysis in the object you get from NCA_fetch_current_ana

ui_ids: Vector of UI elements for the module.

ui_hold: List of hold elements to disable updates before a full ui referesh is complete.

MOD_TYPE: Character data containing the type of module "NCA"

id: Character data containing the module id module in the session variable.

FM_yaml_file: App configuration file with FM as main section.

MOD_yaml_file: Module configuration file with MC as main section.

Examples

library(ruminate)
# Module IDs                                                                   
id     = "NCA"                                                                 
# We need session and input variables to be define                             
 sess_res = NCA_test_mksession()
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → ASM:   dest:   file.path("config","report.docx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → ASM:   dest:   file.path("config","report.pptx")
#> → ASM: including file
#> → ASM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → ASM:   dest:   file.path("config","report.yaml")
#> → ASM: State initialized
#> → ASM: setting word placeholders: 
#> → ASM:   -> setting docx ph: HEADERLEFT = left header
#> → ASM:   -> setting docx ph: HEADERRIGHT = right header
#> → ASM:   -> setting docx ph: FOOTERLEFT = left footer
#> → ASM: module isgood: TRUE
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → UD:   dest:   file.path("config","report.docx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → UD:   dest:   file.path("config","report.pptx")
#> → UD: including file
#> → UD:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → UD:   dest:   file.path("config","report.yaml")
#> → UD: State initialized
#> → UD: module checksum updated:897d952fecbc804999396a96f9df4b20
#> → UD: module isgood: TRUE
#> → DM: including file
#> → DM:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DM:   dest:   file.path("config","report.docx")
#> → DM: including file
#> → DM:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DM:   dest:   file.path("config","report.pptx")
#> → DM: including file
#> → DM:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DM:   dest:   file.path("config","report.yaml")
#> → DM: module checksum updated: f280116ab6b4488df6de76aff513a83f
#> → DM: State initialized
#> → DM: ds checksum updated: 641fcbc6616d24d05acded4317a3bb12
#> → DM: module checksum updated: 78099100dae67cc0fee50517ea72a920
#> → DM: adding preload sources
#> → DM:  - file added: TEST_DATA.xlsx
#> → DM: added: TEST_DATA.xlsx
#> → DM:  - url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_PC.csv
#> → DM: url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_PC.csv
#> → DM:  - url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_EX.csv
#> → DM: url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_EX.csv
#> → DM:  - url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_DM.csv
#> → DM: url added: https://raw.githubusercontent.com/john-harrold/formods/refs/heads/master/inst/test_data/SDTM_DM.csv
#> → DM: loading element idx: 1
#> → DM: ds checksum updated: a260f0653846f780df917cbe451fdbc4
#> → DM: module checksum updated: dabc19da05e7d47bad63339c834161f2
#> → DM: added element idx: 1
#> → DM: loading element idx: 2
#> → DM: ds checksum updated: 581c9783e11b2ba4cfcc4dae2a1df801
#> → DM: module checksum updated: eb37d311c9d8ed9c5f265285c5beae62
#> → DM: added element idx: 2
#> → DM: loading element idx: 3
#> → DM: ds checksum updated: f5b07e6f6fe909fc07b3e5f49d48fb52
#> → DM: module checksum updated: 463d2600d61f2bbc3a16bfbbb9e76a7b
#> → DM: added element idx: 3
#> → DM: loading element idx: 4
#> → DM: ds checksum updated: e6bf3112154fa24b08644e5d9a10a0a4
#> → DM: module checksum updated: db00c4309cb36441424d1e29b2036c0a
#> → DM: added element idx: 4
#> → DM: module isgood: TRUE
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → DW:   dest:   file.path("config","report.docx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → DW:   dest:   file.path("config","report.pptx")
#> → DW: including file
#> → DW:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → DW:   dest:   file.path("config","report.yaml")
#> → DW: data view checksum updated: 
#> → DW:   old: 
#> → DW:   new: e19ddb59715ba77afdd1907053aa5d14
#> → DW: module checksum updated:a9496b622f92f683cc8c69f65ac1ece4
#> → DW: State initialized
#> → DW: data view checksum updated: 
#> → DW:   old: 
#> → DW:   new: e19ddb59715ba77afdd1907053aa5d14
#> → DW: module checksum updated:0955d1f9a1922b95277b4ad1df38e987
#> → DW: data view checksum updated: 
#> → DW:   old: 
#> → DW:   new: e19ddb59715ba77afdd1907053aa5d14
#> → DW: module checksum updated:3d3523604fcde20ce1792b16611bdba2
#> → DW: data view checksum updated: 
#> → DW:   old: 
#> → DW:   new: e19ddb59715ba77afdd1907053aa5d14
#> → DW: module checksum updated:dfec3d5aa14c2716ccf367706931693f
#> → DW: loading data view idx: 1
#> → DW:   -> setting data source: myDS
#> → DW: setting name: Observations
#> → DW: module checksum updated:4d9fd3d913355669d14e0ac710b0b5dc
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: e19ddb59715ba77afdd1907053aa5d14
#> → DW:   new: 0017bd7cdc15ff708408b2d60672ef88
#> → DW: module checksum updated:2aba88838dee694a3d2813ed9f2837d2
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 0017bd7cdc15ff708408b2d60672ef88
#> → DW:   new: 278da634fa74dc23eae6db31314e0324
#> → DW: module checksum updated:c181a352a2bd3592494d917b87057086
#> → DW:   -> mutate
#> → DW: data view checksum updated: 
#> → DW:   old: 278da634fa74dc23eae6db31314e0324
#> → DW:   new: a164ecd5a527b2b38058230c1daa0080
#> → DW: module checksum updated:048df4e8afa9494121e72808981c842c
#> → DW: loading data view idx: 2
#> → DW:   -> setting data source: myDS
#> → DW: setting name: PK 3mg SD IV
#> → DW: module checksum updated:c54c6fc20ff13fd14fbb61c786f65078
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: e19ddb59715ba77afdd1907053aa5d14
#> → DW:   new: 0017bd7cdc15ff708408b2d60672ef88
#> → DW: module checksum updated:496c1149f14ec1c48d4ddbc5aec72a72
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 0017bd7cdc15ff708408b2d60672ef88
#> → DW:   new: 278da634fa74dc23eae6db31314e0324
#> → DW: module checksum updated:d65a6b3c16bea4262f7fb15a1d961512
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 278da634fa74dc23eae6db31314e0324
#> → DW:   new: 00bd5f01d02d634acf134cf01970a1c3
#> → DW: module checksum updated:f90816b8f2ef9aeaad82e4a00880e536
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 00bd5f01d02d634acf134cf01970a1c3
#> → DW:   new: a01562e32a19fd4b10f88ace089346c6
#> → DW: module checksum updated:452fc9d9131592650a93b4b49de961d1
#> → DW: loading data view idx: 3
#> → DW:   -> setting data source: myDS
#> → DW: setting name: PK 3mg SD IV (NCA)
#> → DW: module checksum updated:3b177ed09a80dfaf15c9f3dde8b9a8fb
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: e19ddb59715ba77afdd1907053aa5d14
#> → DW:   new: 0017bd7cdc15ff708408b2d60672ef88
#> → DW: module checksum updated:ce7e35b3bc2e1f53088ae7edd5b4345e
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 0017bd7cdc15ff708408b2d60672ef88
#> → DW:   new: 198f7d878f23121868e677c0b86a5f3c
#> → DW: module checksum updated:64a5146d5e7070664557a7ff3d80f6f0
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 198f7d878f23121868e677c0b86a5f3c
#> → DW:   new: 4326a7b6b9040676b320b387ba5b8ad9
#> → DW: module checksum updated:2cf968d6e72262f9274ea0389f717e84
#> → DW: loading data view idx: 4
#> → DW:   -> setting data source: myDS
#> → DW: setting name: PKPD 3mg SD IV (NCA)
#> → DW: module checksum updated:67e9d6e6b2e085808be155613f8efe20
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: e19ddb59715ba77afdd1907053aa5d14
#> → DW:   new: 0017bd7cdc15ff708408b2d60672ef88
#> → DW: module checksum updated:a7a211de4eb91c91098a1f20e62bec73
#> → DW:   -> filter
#> → DW: data view checksum updated: 
#> → DW:   old: 0017bd7cdc15ff708408b2d60672ef88
#> → DW:   new: 198f7d878f23121868e677c0b86a5f3c
#> → DW: module checksum updated:306f66790aca305ff1b715f34802610e
#> → DW:   -> filter
#> → DW: module isgood: TRUE
#> → NCA: including file
#> → NCA:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → NCA:   dest:   file.path("config","report.docx")
#> → NCA: including file
#> → NCA:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → NCA:   dest:   file.path("config","report.pptx")
#> → NCA: including file
#> → NCA:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → NCA:   dest:   file.path("config","report.yaml")
#> → NCA: including file
#> → NCA:   source: file.path(system.file(package="onbrand"), "templates", "report.docx")
#> → NCA:   dest:   file.path("config","report.docx")
#> → NCA: including file
#> → NCA:   source: file.path(system.file(package="onbrand"), "templates", "report.pptx")
#> → NCA:   dest:   file.path("config","report.pptx")
#> → NCA: including file
#> → NCA:   source: file.path(system.file(package="onbrand"), "templates", "report.yaml")
#> → NCA:   dest:   file.path("config","report.yaml")
#> → NCA: State initialized
#> → NCA: State initialized
#> → NCA: loading element idx: 1
#> → NCA:   -> setting name: PK Example
#> → NCA:   -> notes found and set
#> → NCA:   -> setting data source: DW_myDS_3
#> → NCA: module checksum updated:87e9bf7cae52913d38f093b2a7e3698c
#> → NCA: NCA_add_int: append
#> → NCA: analysis fresh flag set to FALSE
#> → NCA: analysis checksum updated:119a2d1a02bbcafaba53781ddb0ec701
#> → NCA: module checksum updated:eca633c366f3787eed85f5bea80e8cf8
#> → NCA: added element idx: 1
#> → NCA: module isgood: TRUE
#> → # preload_complete is good: TRUE ------------------------------------------
                                                                               
# Extracting the session and input variables                                   
session      = sess_res$session                                                
input        = sess_res$input                                                  
react_state  = list()                                                          
                                                                               
# We also need configuration files                                             
FM_yaml_file  = system.file(package = "formods",  "templates", "formods.yaml") 
MOD_yaml_file = system.file(package = "ruminate", "templates", "NCA.yaml")     
                                                                               
# Getting the current module state                                             
state = NCA_fetch_state(id             = id,                                   
                       input           = input,                                
                       session         = session,                              
                       FM_yaml_file    = FM_yaml_file,                         
                       MOD_yaml_file   = MOD_yaml_file,                        
                       react_state     = react_state)                          
                                                                               
# Pulls out the active analysis                                                
current_ana = NCA_fetch_current_ana(state)

# This will get the dataset associated with this analysis
ds = NCA_fetch_ana_ds(state, current_ana)

# After making changes you can update those in the state
state = NCA_set_current_ana(state, current_ana)

# You can use this to check the current analysis
current_ana = NCA_process_current_ana(state)

# This will pull out the code for the module
fc_res = NCA_fetch_code(state)

# This will use patterns defined for the site to detect
# columns. In this example we are detecting the id column:
id_col = NCA_find_col(
  patterns = state[["MC"]][["detect_col"]][["id"]],
  dscols   = names(ds))

# This creates a new analysis
state = NCA_new_ana(state)

# Switching back to the previous analysis
state = NCA_mkactive_ana(state, "NCA_1")

# Forcing a checksum update
state = NCA_update_checksum(state)
#> → NCA: module checksum updated:404e1b1b9d72602b8584242f7539adea