Title

Description

Usage

mk_table_nca_params(
  nca_res,
  type = "individual",
  grouping = "interval",
  not_calc = "NC",
  obnd = NULL,
  nps = NULL,
  infinity = list(text = "∞", md = "∞"),
  digits = 3,
  max_row = NULL,
  max_col = NULL
)

Arguments

nca_res

Output of PKNCA.

type

Type of table to generate. Can be either "individual" or "summary"].

grouping

How to group columns in tables. Can be either "interval" or "parameter"].

obnd

onbrand reporting object.

nps

NCA parameter summary table with the following columns.

parameter: PKNCA Paramter name.

text: Name used in text output.

md: Name used markdown output.

latex: Name used in latex output.

description: Verbose textual description of the parameter.

digits

Number of significant figures to report (set to NULL to disable rounding)

max_col

Maximum number of columns to have on a page. Spillover will be wrapped to multiple pages.

Value

list containing the following elements

raw_nca: Raw PKNCA output.

isgood: Boolean indicating the exit status of the function.

msgs: Vector of text messages describing any errors that were found.

tables: Named list of tables. Each list element is of the output format from build_span().